### 2025-05-13 #8190 @earthai-tech
> **Submitting author:** [@earthai-tech](https://github.com/earthai-tech) ([Kouao Laurent KOUADIO](http://orcid.org/0000-0001-7259-7254)) **Repository:** https://github.com/earthai-tech/k-diagram **Branch with paper.md** (empty if default branch): **Version:** V1.0.3 **Editor:** Pending **Reviewers:** Pending **Managing EiC:** Chris Vernon
>
> ## Status
> [![status](https://camo.githubusercontent.com/dae2048f117c156b858538bee7b542fb2ff8bd55c524fa951bc07d914a83e4e9/68747470733a2f2f6a6f73732e7468656f6a2e6f72672f7061706572732f62316238336432303532363331313765306235346236373064343630303137342f7374617475732e737667)](https://joss.theoj.org/papers/b1b83d205263117e0b54b670d4600174)
>
> Status badge code:
>
> ```
> ```
>
> **Author instructions**
>
> Thanks for submitting your paper to JOSS [@earthai-tech](https://github.com/earthai-tech). **Currently, there isn't a JOSS editor assigned** to your paper.
>
> [@earthai-tech](https://github.com/earthai-tech) if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
>
> **Editor instructions**
>
> The JOSS submission bot [@editorialbot](https://github.com/editorialbot) is here to help you find and assign reviewers and start the main review. To find out what [@editorialbot](https://github.com/editorialbot) can do for you type:
>
> ```
> @editorialbot commands
> ```

silaslll
rachitanaik-lyft
AnSwati
ChinniAbburi
