### 2025-05-21 #8190 @logological
@crvernon Sorry, I'm currently at capacity and in any case the paper seems outside my familiarity.
