### 2025-06-01 #8190 @earthai-tech
> 👋 [@earthai-tech](https://github.com/earthai-tech) - I noticed your commit history is young in your repository. Will you tell me a bit more about the development history of your project?

Thank you for your comment and for reviewing my repository.

The development of the k-diagram package officially started in November 2024 locally. Still, it was pushed to the repository last month as part of ongoing research on forecasting uncertainty in urban geohazards, particularly land subsidence. The motivation behind developing the package stems from the challenges observed during my study in urban environments, such as Zhongshan and Nansha, China, where uncertainty in multi-horizon forecasts was challenging to interpret using traditional methods.

Since its inception, the k-diagram package has undergone significant progress, and it is already being used to generate key visualizations for research papers that are currently under submission, such as:

1. **Understanding Uncertainty in Land Subsidence Forecasting** (submitted to IEEE Transactions on Geosciences and Remote Sensing (IEEE TGRS) IEEE TGRS-2025-04174) - This paper discusses the uncertainty structure and consistency in forecasting land subsidence, with contributions from the k-diagram package.

2. **Forecasting Urban Land Subsidence in the Era of Rapid Urbanization and Climate Stress**  (submitted to Nature Communications (NCOMMS-25-34098-T) previously in Nature Sustainability) - This paper applies visualization techniques from k-diagram to forecast urban land subsidence in Nansha and Zhongshan.

In this short period, the k-diagram has proven effective in evaluating uncertainty in complex urban geohazard models, making a notable impact on the clarity of the results in both the IJF and Nature Sustainability papers.

The repository is actively maintained, and as of now, several updates have been made to improve the package's usability and extend its functionalities. This includes plans for archiving stable releases on Zenodo and providing persistent Digital Object Identifiers (DOIs) for easier citation.

I hope this helps explain the development history. Please let me know if you would like more details.
