### 2025-07-29 #8190 @jromanowska
:wave: @mwaskom, @caimeng2, @dinacmistry, @cbueth, & @williamjameshandley, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
