### 2025-07-29 #8190 @cbueth
Hey @jromanowska, I could be able to fit in the review in the next weeks. It would be my first review for JOSS, but with the guidelines I should get along well.
