### 2025-07-29 #8661 @editorialbot
[REVIEW]: k-diagram: Rethinking Forecasting Uncertainty via Polar-based Visualization
**Submitting author:** @earthai-tech (<a href="http://orcid.org/0000-0001-7259-7254">Kouao Laurent KOUADIO</a>)
**Repository:** https://github.com/earthai-tech/k-diagram
**Branch with paper.md** (empty if default branch):
**Version:** v1.3.4
