### 2025-07-29 000dcf0
Update paper.bib
