+++ paper.bib
@@ -4,7 +4,7 @@
   author = {{Python Software Foundation}},
   title  = {Python Language Reference, version 3.x},
   url    = {https://www.python.org},
-  note   = {Accessed: 2025-04-10} % Or specific version/date used
+  note   = {Accessed: 2025-04-10} 
 }
 
 @article{Murphy1993What,
