### 2025-07-29 #8661 @earthai-tech
> @earthai-tech, please remove the text "% Or specific version/date used" in line 7 in paper.bib (https://github.com/earthai-tech/k-diagram/blob/main/paper.bib) - I think that's why the generating of reference failed.

Done. Thanks
