### 2025-07-29 #8661 @jromanowska
@earthai-tech - I was wrong: I though you were refering to the "Python" citation on page 2 of your manuscript, where there are "??" signs, indicating wrong reference. I can see in the source that you have a typo in paper.md:

`\autoref{fig1:worflow}` should be `\autoref{fig1:workflow}`

And perhaps you want to use the Python citation there as well...?

Additionally, please fix the DOI problems, as indicated by the bot [above](https://github.com/openjournals/joss-reviews/issues/8661#issuecomment-3132326081)
