### 2025-07-29 d2226ee
Update paper.bib
