+++ paper.bib
@@ -8,15 +8,16 @@
 }
 
 @article{Murphy1993What,
-  author    = {Murphy, Allan H. and Winkler, Robert L.},
-  title     = {What is a good forecast? Probability, surprise, consensus, information, and skill in the evaluation of forecasts},
+  author    = {Allan H.  Murphy},
+  title     = {What Is a Good Forecast? An Essay on the Nature of Goodness in Weather Forecasting},
   journal   = {Weather and Forecasting},
   volume    = {8},
   number    = {3},
   pages     = {287-303},
   year      = {1993},
-  doi       = {10.1175/1520-0434(1993)008<0287:WIAGFP>2.0.CO;2},
-  publisher = {American Meteorological Society}
+  doi       = {10.1175/1520-0434(1993)008<0281:WIAGFA>2.0.CO;2},
+  publisher = {American Meteorological Society},
+  url = {https://journals.ametsoc.org/view/journals/wefo/8/2/1520-0434_1993_008_0281_wiagfa_2_0_co_2.xml}
 }
 
 @article{Hunter:2007,
@@ -56,7 +57,7 @@
 
 @article{Liu2024,
     author = {Liu, Jianxin and Liu, Wenxiang and Allechy, Fabrice Blanchard and Zheng, Zhiwen and Liu, Rong and Kouadio, Kouao Laurent},
-    doi = {https://doi.org/10.1016/j.jenvman.2024.120078},
+    doi = {10.1016/j.jenvman.2024.120078},
     journal = {Journal of Environmental Management},
     number = {2024},
     pages = {17},
@@ -64,7 +65,6 @@
     volume = {352},
     year = {2024}
 }
-
 
 @article{harris2020array,
   title={Array programming with NumPy},
@@ -86,7 +86,8 @@
   editor    = {van der Walt, St{\'e}fan and Millman, Jarrod},
   pages     = {56 - 61},
   year      = {2010},
-  publisher = {SciPy}
+  publisher = {SciPy}, 
+  doi={10.25080/majora-92bf1922-00a}, 
 }
 
 @article{reback2020pandas,
