### 2025-07-29 8560e51
Update paper.bib
