+++ paper.bib
@@ -4,7 +4,17 @@
   author = {{Python Software Foundation}},
   title  = {Python Language Reference, version 3.x},
   url    = {https://www.python.org},
-  note   = {Accessed: 2025-04-10} 
+  note   = {Accessed: 2025-04-10}, 
+  doi = {10.5555/1593511}
+}
+@book{10.5555/1593511,
+author = {Van Rossum, Guido and Drake, Fred L.},
+title = {Python 3 Reference Manual},
+year = {2009},
+isbn = {1441412697},
+publisher = {CreateSpace},
+address = {Scotts Valley, CA},
+abstract = {PYTHON 3 Reference Manual (Python Documentation MANUAL Part 2).Python is an easy to learn object-oriented programming language, which combines power with clear syntax. It has modules, classes, exceptions, very high level data types, and dynamic typing. Python is free software. It can be used with GNU (GNU/Linux), Unix, Microsoft Windows and many other systems.This is a printed softcover copy of the official Python documentation from the latest Python 3.0 distribution. For each copy sold $1 will be donated to the Python Software Foundation by the publisher.This book is part of a brand new six-part series of Python documentation books. Searching for "Python Documentation Manual" will show all six available books.ABOUT THE AUTHOR: Guido van Rossum, is the inventor of Python. Fred L. Drake, Jr. is the official editor of the Python documentation.}
 }
 
 @article{Murphy1993What,
