### 2025-07-29 0b70b69
Update paper.md
