+++ paper.md
@@ -73,7 +73,7 @@ to be diagnosed.
 
 # Functionality
 
-`k-diagram` is implemented in Python (\autoref{fig1:worflow}), leveraging core scientific libraries
+`k-diagram` is implemented in Python (\autoref{fig1:workflow}), leveraging core scientific libraries
 including `numpy` [@harris2020array], `pandas` [@mckinney-proc-scipy-2010;
 @reback2020pandas], `matplotlib` [@Hunter:2007], `scipy` [@2020SciPy-NMeth],
 and `scikit-learn` [@scikit-learn]. Its main features include:
