### 2025-07-31 10c531c
Update JOSS paper
