+++ paper.bib
@@ -4,19 +4,32 @@
   author = {{Python Software Foundation}},
   title  = {Python Language Reference, version 3.x},
   url    = {https://www.python.org},
-  note   = {Accessed: 2025-04-10} % Or specific version/date used
+  note   = {Accessed: 2025-04-10}, 
+  url = {https://dl.acm.org/doi/10.5555/1593511}
+}
+
+@book{10.5555/1593511,
+  author = {Van Rossum, Guido and Drake, Fred L.},
+  title = {Python 3 Reference Manual},
+  year = {2009},
+  isbn = {1441412697},
+  publisher = {CreateSpace},
+  address = {Scotts Valley, CA},
+  abstract = {PYTHON 3 Reference Manual (Python Documentation MANUAL Part 2). Python is an easy-to-learn object-oriented programming language, which combines power with clear syntax. It has modules, classes, exceptions, very high-level data types, and dynamic typing. Python is free software. It can be used with GNU (GNU/Linux), Unix, Microsoft Windows, and many other systems. This is a printed softcover copy of the official Python documentation from the latest Python 3.0 distribution. For each copy sold, $1 will be donated to the Python Software Foundation by the publisher. This book is part of a brand new six-part series of Python documentation books. Searching for "Python Documentation Manual" will show all six available books. ABOUT THE AUTHOR: Guido van Rossum, is the inventor of Python. Fred L. Drake, Jr. is the official editor of the Python documentation.},
+  url = {https://dl.acm.org/doi/10.5555/1593511}
 }
 
 @article{Murphy1993What,
-  author    = {Murphy, Allan H. and Winkler, Robert L.},
-  title     = {What is a good forecast? Probability, surprise, consensus, information, and skill in the evaluation of forecasts},
+  author    = {Allan H.  Murphy},
+  title     = {What Is a Good Forecast? An Essay on the Nature of Goodness in Weather Forecasting},
   journal   = {Weather and Forecasting},
   volume    = {8},
   number    = {3},
   pages     = {287-303},
   year      = {1993},
-  doi       = {10.1175/1520-0434(1993)008<0287:WIAGFP>2.0.CO;2},
-  publisher = {American Meteorological Society}
+  doi       = {10.1175/1520-0434(1993)008<0281:WIAGFA>2.0.CO;2},
+  publisher = {American Meteorological Society},
+  url = {https://journals.ametsoc.org/view/journals/wefo/8/2/1520-0434_1993_008_0281_wiagfa_2_0_co_2.xml}
 }
 
 @article{Hunter:2007,
@@ -56,7 +69,7 @@
 
 @article{Liu2024,
     author = {Liu, Jianxin and Liu, Wenxiang and Allechy, Fabrice Blanchard and Zheng, Zhiwen and Liu, Rong and Kouadio, Kouao Laurent},
-    doi = {https://doi.org/10.1016/j.jenvman.2024.120078},
+    doi = {10.1016/j.jenvman.2024.120078},
     journal = {Journal of Environmental Management},
     number = {2024},
     pages = {17},
@@ -64,7 +77,6 @@
     volume = {352},
     year = {2024}
 }
-
 
 @article{harris2020array,
   title={Array programming with NumPy},
@@ -86,7 +98,8 @@
   editor    = {van der Walt, St{\'e}fan and Millman, Jarrod},
   pages     = {56 - 61},
   year      = {2010},
-  publisher = {SciPy}
+  publisher = {SciPy}, 
+  doi={10.25080/majora-92bf1922-00a}, 
 }
 
 @article{reback2020pandas,
@@ -118,9 +131,138 @@
   doi       = {10.1029/2000JD900719}
 }
[…]
