+++ paper.md
@@ -12,71 +12,79 @@ tags:
 authors:
   - name: Kouao Laurent Kouadio
     orcid: 0000-0001-7259-7254 
-    affiliation: "1, 2" 
+    affiliation: "1, 2, 3" 
 
 affiliations:
  - name: School of Geosciences and Info-Physics, Central South University, Changsha, Hunan, 410083, China
    index: 1
  - name: Hunan Key Laboratory of Nonferrous Resources and Geological Hazards Exploration, Changsha, Hunan, 410083, China
    index: 2
+
 date: 10 May 2025 
 bibliography: paper.bib
 ---
 
 # Summary
 
-`k-diagram` is an open-source Python package providing specialized
-diagnostic plots based on polar coordinates for the comprehensive
-evaluation of forecasting models, with a strong emphasis on understanding
-and visualizing predictive uncertainty. It offers a suite of functions
-('k-diagrams') to assess prediction interval coverage, consistency,
-anomaly magnitude, model drift across horizons, feature importance
-patterns, and standard evaluation metrics (e.g., via Taylor diagrams).
-Designed to complement traditional evaluation methods, `k-diagram` aids
-interpretability and facilitates deeper insights into forecast
-reliability, particularly for complex spatiotemporal systems relevant
-to environmental science and geohazard assessment.
+`k-diagram` is an open-source Python package designed for the in-depth 
+interpretation and diagnosis of probabilistic forecasts. Moving beyond 
+traditional aggregate metrics, the package provides a suite of novel 
+visualization tools, primarily based on polar coordinate systems, 
+to dissect the complex spatiotemporal structure of forecast uncertainty. 
+By mapping statistical properties such as sharpness, reliability, and 
+temporal stability onto intuitive geometric representations, `k-diagram` 
+enables researchers and practitioners to identify model biases, understand 
+forecast degradation over time, and communicate uncertainty characteristics 
+more effectively. The package is designed to be a practical extension 
+to the standard forecasting workflow [@Hong2025], providing the visual evidence needed 
+for more robust model evaluation and trustworthy, context-aware 
+decision-making.
 
 # Statement of Need
 
-Evaluating the performance of forecasting models, especially in high-stakes
-domains like environmental hazard prediction (e.g., land subsidence,
-flooding), requires moving beyond simple point-forecast accuracy metrics.
-Understanding the structure, reliability, and limitations of a model's
-predictive uncertainty is critical for risk assessment and informed
-decision-making [@Murphy1993What]. However, standard visualization tools
-often struggle to present multi-faceted uncertainty information
-(e.g., interval width, coverage success/failure, temporal drift)
-simultaneously and intuitively across many data points or locations.
+The evaluation of probabilistic forecasts has a rigorous theoretical 
+foundation based on concepts of calibration and sharpness, which are 
+jointly assessed using proper scoring rules [@Gneiting2007b]. However, 
+these scores are typically aggregated into a single value, which can 
+obscure vital information about a model's performance in heterogeneous 
+spatiotemporal settings. This is a critical limitation, as modern machine 
+learning models, such as Temporal Fusion Transformers [@Lim2021], 
+are increasingly applied to complex, high-dimensional problems where 
+understanding the local and temporal behavior of uncertainty is paramount. 
+A growing body of applied research highlights this challenge in fields as 
+diverse as predictive policing [@Rummens2021], energy forecasting [@Liu2021], 
+climatology [@Baillie2002] and geohazard [@Liu2024].
 
-Existing libraries provide excellent general plotting (e.g., `matplotlib`
[…]
