### 2025-07-31 7fb58e0
Update paper.md
