### 2025-07-31 #8661 @earthai-tech
**Subject:** Minor Revisions to Manuscript - Improved Context and Clarity

Dear Reviewers @ChinniAbburi  and @cbueth,

Thank you for your time dedicated  on my manuscript. To strengthen the paper further,I have made minor revisions focused on:

1. **Refining the Statement of Need** – Better contextualizing our software’s contribution relative to current literature.

2. **Polishing figure captions** – Enhancing clarity and precision.

These updates aim to make the paper’s argument more compelling while addressing broader context and presentation. The revised version is available here:
👉 [Revised Manuscript Link](https://github.com/openjournals/joss-reviews/issues/8661#issuecomment-3139454993)

I'm  grateful for your insights and welcome any suggestions you may have.

Best regards!
