+++ paper.bib
@@ -4,11 +4,12 @@
   author = {{Python Software Foundation}},
   title  = {Python Language Reference, version 3.x},
   url    = {https://www.python.org},
+  year = {2008},
   note   = {Accessed: 2025-04-10}, 
   url = {https://dl.acm.org/doi/10.5555/1593511}
 }
 
-@book{10.5555/1593511,
+@book{pythonbook,
   author = {Van Rossum, Guido and Drake, Fred L.},
   title = {Python 3 Reference Manual},
   year = {2009},
@@ -261,7 +262,7 @@
   doi       = {10.22541/au.175390529.91420978/v1}
 }
 @article{kouadiob2025,
-  author = {Kouadio, Laurent and Liu, Rong and Loukou, Kouam{\'e} Gb{\`e}l{\`e} Hermann and Liu, Jianxin and Liu, Wenxiang},
+  author = {Kouadio, Kouao Laurent and Liu, Rong and Loukou, Kouam{\'e} Gb{\`e}l{\`e} Hermann and Liu, Jianxin and Liu, Wenxiang},
   title = {Analytics Framework for Interpreting Spatiotemporal Probabilistic Forecasts},
   journal = {International Journal of Forecasting},
   year = {2025},
