### 2025-08-07 earthai-tech/k-diagram#14 @cbueth
JOSS Review Feedback
For the https://github.com/openjournals/joss-reviews/issues/8661 review, I have gone through the repository and marked several notes I'd like to have your input on. The points are kept in bullet point format, so let me know if for any of them you're unsure what I am referring to. I have categorized them like so:

### Documentation

- A11y (acessibility): When hovering over an item in the left table of contents, the contrast between font (turning white) and background is not high enough. How about changing the font to black when hovering?
- It seems that docstrings use numpydoc, but the Sphinx documentation does not use the extension (https://numpydoc.readthedocs.io/en/latest/index.html), even though `napoleon_numpy_docstring = True` is set, resulting in Reference pages where the docstrings are not parsed correctly (see e.g. https://k-diagram.readthedocs.io/en/latest/_autosummary/uncertainty/kdiagram.plot.uncertainty.plot_anomaly_magnitude.html)

### Packaging

- Package uses `setup.py` which is not deprecated, but using a `pyproject.toml` is preferable. See https://packaging.python.org/en/latest/guides/modernize-setup-py-project/
  Having `pyproject.toml` also allows more settings to be included there, instead of having separate files:
  - `pytest.ini`
  - `.coveragerc`
  - `.flake8`

- try-except in `setup.py` for the `kdiagram.__version__` makes the package version redundant and unstable. Same in `kdiagram/__init__.py`, where it points to the `_version.py` file, but there, it will always run into an `ImportError`. Please unify this into optimally one place.
- Good having the most important functions exposed by the package in the `__init__.py` file. :+1:
- You supress `SyntaxWarnings` globally in the package. While there might be several reasons to do this, I'd like to know your specific motivation in this case.
- The tests are inside the `k-diagram` package, but are not a subpackage, as the `__init__.py` is missing. "PytestConfigWarning: No files were found in testpaths"

### CI

- Solve the two warnings in the current CI https://github.com/earthai-tech/k-diagram/actions/runs/16663809289. The latter, at least, should be rather simple to solve.
- Code coverage is being measured and uploaded, but not shown in the `README.md`. It should be more visible. Are they reported somewhere else? As of now at 18% code coverage, it can be improved.
  Opening the code tree in https://app.codecov.io/github/earthai-tech/k-diagram/tree/main/kdiagram you can see that the measured coverage is only the `tests/` folder, no any actual source code. This points to that `pytest --cov=kdiagram --cov-report=xml -n auto kdiagram/tests` is not measuring the source code as target.
- You might want to consider running the tests on a more recent version than `3.10` (or both a minimum version and the most current version), as you specify python `3.12` to be supported.
- `flake8` is used for linting in the CI and produces 6067 codestyle warnings.
  At the same time, the `README.md` advertises the adheres to `black` style, which is not the case, as 37 of 41 files in `kdiagram/` would be changed by `black`.
  You can enforce them by making the CI fail if the formatting is not being adhered. Furthermore, using one of the two can suffice applied consequently.

### Development

- The `README.md` suggests `pip install -e .[dev]` from the source directory. The project also has a `environment.yml` file that is not mentioned anywhere, if I'm not mistaken. As this environment has more packages, it might be worth it mentioning this somewhere, or let `pip install -e .[dev]` also install the development dependencies which are not being installed, but are in the `environment.yml` file.
- `pytest.ini`
  - Is the marker `require_skbio` used anywhere?
  - Running just `pytest` passes 266 tests, skips 29, but produces 339 warnings. Please consult these warnings and see if you can fix them.
    - "UserWarning: FigureCanvasAgg is non-interactive" should be
    - "MatplotlibDeprecationWarning"
    - "RuntimeWarning: All-NaN slice encountered"
    - "PytestConfigWarning: No files were found in testpaths"
    - When testing `kdiagram/tests/test_uncertainty.py:241` without the skip marker, I get "'`qlow_cols` and `qup_cols` must be the same length." as warning, but the expected match is "Features missing". While I did not check what is happening internally, I am not sure I understand the initial reason for skipping "Regex pattern did not match. Unable to exactly match the error message."
      I am happy to help with regex matching, if you want to. What catches my eye here is that the reason for skipping does not seem reasonable to me, as the expected and the actual error message is about different things, meaning, the error handling might need to be improved.
- Do you plan to support `numpy~=2` in the future?
- The project still has the PyPI classifier "Development Status :: 3 - Alpha". For a publication I'd assume a later development status of the software should be desirable
- As the `environment.yml` suggest it, have you used code from this project? https://github.com/taipingeric/fusionlab

I hope these points will help. Greetings!
