### 2025-08-07 #8661 @cbueth
Hey @earthai-tech! I have spent some time today on code reviewing your repository and have taken some notes you can find in https://github.com/earthai-tech/k-diagram/issues/14. Thank you for submitting your work. Don't be surprised by the many notes, I think the code has a standard to work with, still there is some potential to improve. I'm excited for your reply and will continue with the review, especially reading the paper, sometime in the coming days.
