+++ kdiagram/plot/uncertainty.py
@@ -73,214 +73,7 @@ def plot_coverage(
     savefig=None,
     verbose=1 
 ):
-    r"""Plot overall coverage scores for forecast intervals or points.
-
-    Computes and visualizes the empirical coverage rate, which is
-    the fraction of times the true values fall within predicted
-    quantile intervals (or match point forecasts). Supports comparing
-    multiple models or prediction sets using various chart types.
-
-    This plot provides a high-level summary of whether prediction
-    intervals are well-calibrated on average across the entire
-    dataset. It's useful for quickly comparing the overall
-    reliability of uncertainty estimates from different models or
-    parameter settings before diving into finer diagnostics.
-
-    Parameters
-    ----------
-    y_true : array-like of shape (n_samples,)
-        The ground truth target values. Must be a 1D array-like
-        object (list, NumPy array, pandas Series).
-
-    *y_preds : array-like
-        Variable number of prediction arrays. Each positional
-        argument should be an array-like object corresponding to a
-        model or prediction set. The shape determines interpretation:
-        - shape (n_samples, n_quantiles): Represents quantile
-          forecasts (e.g., lower, median, upper bounds). Coverage
-          is calculated based on the interval defined by the minimum
-          and maximum quantile columns provided (after sorting based
-          on `q` if given, or by value otherwise). Requires `q`
-          parameter for proper interval definition if quantiles
-          are not ordered.
-        - shape (n_samples,): Represents point forecasts. Coverage
-          is calculated as the proportion of exact matches to
-          `y_true` (typically near zero for continuous data).
-
-    names : list of str, optional
-        Labels for each prediction set provided in `*y_preds`. Used
-        in legends and axis labels. If `None` or shorter than the
-        number of prediction sets, default names like "Model_1",
-        "Model_2", etc., are generated for the missing ones.
-        Default is ``None``.
-
-    q : list of float, optional
-        Quantile levels corresponding to the columns in 2D `y_preds`
-        arrays. Values must be between 0 and 1. Used to identify
-        and sort quantile columns correctly before determining the
-        interval bounds (minimum and maximum specified quantile) for
-        coverage calculation. Required if `y_preds` contains 2D
-        arrays representing unordered quantiles. Default is ``None``.
-
-    kind : {'line', 'bar', 'pie', 'radar'}, optional
-        The type of plot to generate for visualizing coverage scores.
-        Default is ``'line'``.
-
-        - ``'line'``: Line chart connecting coverage scores for
-          each prediction set.
-        - ``'bar'``: Bar chart showing the coverage score for each
-          prediction set.
-        - ``'pie'``: Pie chart where each slice represents a model's
-          coverage score as a fraction of the sum of all scores.
-        - ``'radar'``: Radar (spider) chart where each axis
-          represents a prediction set, and the distance along the
-          axis indicates its coverage score.
-
-    cmap : str, optional
-        Matplotlib colormap name used for coloring slices in the
-        `'pie'` chart or the gradient fill in single-model radar
-        plots when ``cov_fill=True``. Default is ``'viridis'``.
-
-    pie_startangle : float, optional
-        Starting angle in degrees for the first slice in the `'pie'`
[…]
