+++ kdiagram/__init__.py
@@ -30,7 +30,7 @@ def _lazy_import(module_name, alias=None):
 try:
     from ._version import version as __version__
 except ImportError:
-    __version__ = "1.0.3"
+    __version__ = "1.1.0"
 
 # Dependency check
 _required_dependencies = [
