+++ .github/workflows/python-package-conda.yml
@@ -23,10 +23,10 @@ jobs:
       - name: Checkout repository
         uses: actions/checkout@v4
 
-      - name: Set up Miniforge (Mamba)
+      - name: Set up Miniforge (Miniforge3)
         uses: conda-incubator/setup-miniconda@v3
         with:
-          miniforge-variant: Mambaforge
+          miniforge-variant: Miniforge3  # Updated from Mambaforge
           miniforge-version: latest
           use-mamba: true
           activate-environment: kdiagram-dev
@@ -61,7 +61,7 @@ jobs:
 
       - name: Upload coverage to Codecov
         # Run unless it's a PR from a fork AND ensure a token is present
-        # if: ${{ (github.event_name != 'pull_request' || github.event.pull_request.head.repo.fork == false) && secrets.CODECOV_TOKEN != '' }}
+        if: ${{ (github.event_name != 'pull_request' || github.event.pull_request.head.repo.fork == false) && secrets.CODECOV_TOKEN != '' }}
         uses: codecov/codecov-action@v4
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
