+++ .github/workflows/python-package-conda.yml
@@ -61,10 +61,10 @@ jobs:
 
       - name: Upload coverage to Codecov
         # Run unless it's a PR from a fork AND ensure a token is present
-        if: ${{ (github.event_name != 'pull_request' || github.event.pull_request.head.repo.fork == false) && secrets.CODECOV_TOKEN != '' }}
+        if: ${{ github.event_name != 'pull_request' || github.event.pull_request.head.repo.fork == false }}
         uses: codecov/codecov-action@v4
         with:
-          token: ${{ secrets.CODECOV_TOKEN }}
+          token: ${{ secrets.CODECOV_TOKEN }}  # This is the correct way to reference secrets
           files: ./coverage.xml                 # default path from pytest-cov
           name: linux-py${{ matrix.python-version }}
           flags: unittests
