+++ .github/workflows/python-package-conda.yml
@@ -51,11 +51,14 @@ jobs:
           # If you switch to Ruff:
       
           python -m pip install -e .[dev]
-          # Use ruff to check for issues (no --fix argument here)
+          # Use ruff to check for issues
           ruff check .
-          black . --check # Automatically formats the code
-          ruff --fix .  # Automatically fix issues
-
+          # Use ruff to automatically fix issues (e.g., imports)
+          ruff --fix .
+    
+          # Use black to check if the code is formatted
+          black . --check  # Check if code adheres to black formatting
+   
       - name: Test with pytest + coverage
         run: |
           echo "Running pytest with coverage..."
