### 2025-08-09 c573b4b
Apply 157 safe linting fixes
