+++ kdiagram/_version.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 # don't change, don't track in version control
 # __version__ = version = '1.0.0.dev1+g10707dc.d20230709'
 # __version_tuple__ = version_tuple = (1,0,0,  'dev1', 'g10707dc.d20230709')
