+++ tests/test_feature_based_plots.py
@@ -1,7 +1,7 @@
+import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 import pytest
-import matplotlib.pyplot as plt
 
 from kdiagram.plot.feature_based import plot_feature_fingerprint
 
