### 2025-08-10 a7d36bc
py312 tests
