+++ tests/test_decorators.py
@@ -22,9 +22,11 @@ def always_catch_warnings():
         warnings.simplefilter("always")
         yield rec
 
+
 @pytest.fixture
 def tiny_df():
     return pd.DataFrame({"a": [1, 2], "b": [3, 4]})
+
 
 # -----------------------------
 # _is_arraylike_empty coverage
@@ -38,6 +40,7 @@ def test__is_arraylike_empty_variants():
     assert _is_arraylike_empty([1]) is False
     assert _is_arraylike_empty(np.array([1])) is False
 
+
 # -----------------------------
 # check_non_emptiness (no parens)
 # -----------------------------
@@ -45,23 +48,31 @@ def test__is_arraylike_empty_variants():
 def _iden_first_arg(x):
     return x
 
+
 def test_check_non_emptiness_no_parens_raise():
-    with pytest.raises(ValueError, match="Argument 'first positional argument' is empty"):
+    with pytest.raises(
+        ValueError, match="Argument 'first positional argument' is empty"
+    ):
         _iden_first_arg([])
+
 
 def test_check_non_emptiness_no_parens_warn(always_catch_warnings):
     @check_non_emptiness(error="warn")
     def foo(x):
         return x
+
     out = foo([])
     assert out is None
     assert any("is empty" in str(w.message) for w in always_catch_warnings)
+
 
 def test_check_non_emptiness_no_parens_ignore():
     @check_non_emptiness(error="ignore")
     def foo(x):
         return x
+
     assert foo([]) is None  # becomes None, no warning
+
 
 # -----------------------------
 # check_non_emptiness (with parens, named params)
@@ -70,30 +81,39 @@ def test_check_non_emptiness_named_kwarg_warn(always_catch_warnings):
     @check_non_emptiness(params=["df"], error="warn")
     def bar(a, df=None):
         return df
+
     out = bar(1, df=pd.DataFrame())
     assert out is None
-    assert any("Argument 'df' is empty" in str(w.message) for w in always_catch_warnings)
+    assert any(
+        "Argument 'df' is empty" in str(w.message) for w in always_catch_warnings
+    )
+
 
 def test_check_non_emptiness_named_positional_raise():
     @check_non_emptiness(params=["b"], error="raise")
     def baz(a, b):
         return b
-  
+
     assert baz(1, []) == []
 
+
 def test_check_non_emptiness_none_ellipsis_and_include():
-    @check_non_emptiness(params=["x","y","z"], error="warn")
+    @check_non_emptiness(params=["x", "y", "z"], error="warn")
     def f(x, y, z):
         return x, y, z
+
     # None as empty
     x, y, z = f(None, ..., set())
     assert x is None and y is None and z is None
+
     # include filtering: if include=() skip set/dict emptiness detection
-    @check_non_emptiness(params=["u","v"], error="raise", include=())
+    @check_non_emptiness(params=["u", "v"], error="raise", include=())
     def g(u, v):
         return u, v
+
     # empty dict is NOT considered empty when include=()
     assert g({}, set()) == ({}, set())
+
 
 # -----------------------------
 # isdf decorator
@@ -103,14 +123,17 @@ def accept_df(data, columns=None):
     assert isinstance(data, pd.DataFrame)
     return data
 
+
 def test_isdf_function_from_list_and_columns():
     df = accept_df([[1, 2], [3, 4]], columns=["x", "y"])
     assert list(df.columns) == ["x", "y"]
+
 
 def test_isdf_function_from_array_no_columns():
[…]
