+++ kdiagram/api/summary.py
@@ -28,7 +28,7 @@ import copy
 from .bunch import FlexDict
 from .util import beautify_dict, get_table_size, to_camel_case, to_snake_case
 
-__all__ = ['ResultSummary'] 
+__all__ = ["ResultSummary"]
 
 
 class ResultSummary(FlexDict):
@@ -37,7 +37,7 @@ class ResultSummary(FlexDict):
     results in a structured format. The class allows for optional customization
     of the display settings such as padding of keys and maximum character limits
     for value display.
-    
+
     Parameters
     ----------
     name : str, optional
@@ -50,16 +50,16 @@ class ResultSummary(FlexDict):
         The maximum number of characters that a value can have before being
         truncated. Defaults to ``100``.
     flatten_nested_dicts : bool, optional
-        Determines whether nested dictionaries within the results should be 
+        Determines whether nested dictionaries within the results should be
         displayed in a flattened, one-line format.
         When set to ``True``, nested dictionaries are presented as a compact single
         line, which might be useful for brief overviews or when space is limited.
-        If set to ``False``, nested dictionaries are displayed with full 
-        indentation and key alignment, which improves readability for complex 
+        If set to ``False``, nested dictionaries are displayed with full
+        indentation and key alignment, which improves readability for complex
         structures. Defaults to ``True``.
-    mute_note: bool, default=False 
-       Skip displaying the note after result formatage. 
-       
+    mute_note: bool, default=False
+       Skip displaying the note after result formatage.
+
     Examples
     --------
     >>> from kdiagram.api.summary import ResultSummary
@@ -77,18 +77,25 @@ class ResultSummary(FlexDict):
       }
     )
     """
-    def __init__(self, name=None, pad_keys=None, max_char=None,
-                 flatten_nested_dicts =True, mute_note=False, 
-                 **kwargs):
+
+    def __init__(
+        self,
+        name=None,
+        pad_keys=None,
+        max_char=None,
+        flatten_nested_dicts=True,
+        mute_note=False,
+        **kwargs,
+    ):
 
         super().__init__(**kwargs)
         self.name = name or "Result"
         self.pad_keys = pad_keys
         self.max_char = max_char or get_table_size()
-        self.flatten_nested_dicts = flatten_nested_dicts 
-        self.mute_note=mute_note
+        self.flatten_nested_dicts = flatten_nested_dicts
+        self.mute_note = mute_note
         self.results = {}
-        
+
     def add_results(self, results):
         """
         Adds results to the summary and dynamically creates attributes for each
@@ -115,33 +122,33 @@ class ResultSummary(FlexDict):
         """
         if not isinstance(results, dict):
             raise TypeError("results must be a dictionary")
-    
-        # Deep copy to ensure that changes to input dictionary 
+
+        # Deep copy to ensure that changes to input dictionary
         # do not affect internal state
         self.results = copy.deepcopy(results)
-    
+
         # Apply snake_case to dictionary keys and set attributes
         for name in list(self.results.keys()):
             snake_name = to_snake_case(name)
             setattr(self, snake_name, self.results[name])
-            
-        return self 
+
+        return self
 
     def __str__(self):
         """
         Return a formatted string representation of the results dictionary.
         """
         _name = to_camel_case(self.name)
-        result_title = _name + '(\n  {\n'
+        result_title = _name + "(\n  {\n"
         formatted_results = []
-        
+
         # Determine key padding if auto pad_keys is specified
         if self.pad_keys == "auto":
             max_key_length = max(len(key) for key in self.results.keys())
             key_padding = max_key_length
         else:
             key_padding = 0  # No padding by default
-    
+
         # Construct the formatted result string
         for key, value in self.results.items():
             if self.pad_keys == "auto":
@@ -149,48 +156,56 @@ class ResultSummary(FlexDict):
             else:
[…]
