### 2025-08-10 669038b
py312 tests
