+++ README.md
@@ -9,7 +9,6 @@
 ![GitHub License](https://img.shields.io/github/license/earthai-tech/k-diagram?style=flat-square&logo=apache&color=purple)
 [![Code Style: Black](https://img.shields.io/badge/code%20style-Black-000000.svg?style=flat-square)](https://github.com/psf/black)
 [![Contributions Welcome](https://img.shields.io/badge/Contributions-Welcome-brightgreen?style=flat-square)](https://github.com/earthai-tech/k-diagram/blob/main/CONTRIBUTING.md)
-<!-- [![Code Coverage](https://img.shields.io/codecov/c/github/earthai-tech/k-diagram?style=flat-square)](https://app.codecov.io/gh/earthai-tech/k-diagram) -->
 [![codecov](https://codecov.io/gh/earthai-tech/k-diagram/branch/main/graph/badge.svg)](https://codecov.io/gh/earthai-tech/k-diagram)
 
 🧭 **_Navigate the complexities of forecast uncertainty and model behavior with `k-diagram`'s specialized polar visualizations_**
