### 2025-08-10 4cc873e
Ready to attempt unsafe fixes
