+++ .github/workflows/python-package-conda.yml
@@ -43,7 +43,7 @@ jobs:
         run: |
           python -m pip install . --no-deps --force-reinstall
 
-      - name:Lint (ruff and black)
+      - name: Lint (ruff and black)
         run: |
           echo "Running flake8 linter..."
           # flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
