+++ kdiagram/__init__.py
@@ -69,10 +69,10 @@ if _missing_dependencies:
     )
 
 # Re-export config helpers
-from .config import configure_warnings, warnings_config  # noqa: F401
+from .config import configure_warnings, warnings_config  # noqa: F401, E402
 
 # from . import datasets
-from .plot import (
+from .plot import (  # noqa: E402
     plot_actual_vs_predicted,
     plot_anomaly_magnitude,
     plot_coverage,
