### 2025-08-11 0661cea
resolve np.float error in Numpy 2.xx
