+++ tests/test_utils_validator.py
@@ -1,0 +1,475 @@
+import numpy as np
+import pandas as pd
+import pytest
+import scipy.sparse as sp
+
+from kdiagram.utils import validator as V
+
+
+# -----------------------
+# validate_length_range
+# -----------------------
+def test_validate_length_range_happy_and_sorting_and_errors():
+    # auto-sorts when sorted_values=True
+    assert V.validate_length_range((202, 25)) == (25, 202)
+    # already sorted
+    assert V.validate_length_range((1, 3)) == (1, 3)
+    # equality after sorting -> error
+    with pytest.raises(ValueError):
+        V.validate_length_range((5, 5))
+    # non-numeric -> error
+    with pytest.raises(ValueError):
+        V.validate_length_range(("a", 2))
+    # sorted_values=False keeps order and allows first>second here (no extra checks)
+    assert V.validate_length_range((9, 1), 
+                                   sorted_values=False) == (9, 1)
+
+
+# ------------
+# validate_yy
+# ------------
+def test_validate_yy_shapes_types_and_expected_type():
+    # flatten=True
+    y_true = np.array([[0.0], [1.0], [2.0]])
+    y_pred = np.array([[0.0], [1.0], [2.0]])
+    yt, yp = V.validate_yy(y_true, y_pred, flatten=True)
+    assert yt.ndim == yp.ndim == 1
+
+    # auto flatten: "auto" string path
+    yt2, yp2 = V.validate_yy(y_true, y_pred, flatten="auto")
+    assert yt2.ndim == yp2.ndim == 1
+
+    # wrong ndim -> error (suggests flatten=True)
+    with pytest.raises(ValueError):
+        V.validate_yy(np.array([[1, 2]]), np.array([[1, 2]]), flatten=False)
+
+    # length mismatch through check_consistent_length -> error
+    with pytest.raises(ValueError):
+        V.validate_yy([0, 1], [0], flatten=True)
+
+    # expected_type mismatch -> error
+    with pytest.raises(ValueError):
+        V.validate_yy([0.1, 0.2], [0.1, 0.3], 
+                      expected_type="binary", flatten=True)
+
+    # expected_type correct (continuous)
+    yt3, yp3 = V.validate_yy([0.1, 0.2], [0.1, 0.3], 
+                             expected_type="continuous", flatten=True)
+    assert yt3.shape == yp3.shape == (2,)
+
+
+# ------------------------
+# contains_nested_objects
+# ------------------------
+def test_contains_nested_objects_strict_and_non_strict():
+    nested = [{1, 2}, [3, 4], {"k": "v"}]
+    mixed = [1, 2, [3]]
+    flat = [1, 2, 3]
+    assert V.contains_nested_objects(nested) is True
+    assert V.contains_nested_objects(nested, strict=True) is True
+    assert V.contains_nested_objects(mixed) is True
+    assert V.contains_nested_objects(mixed, strict=True) is False
+    assert V.contains_nested_objects(flat) is False
+
+
+# ------------------------
+# check_consistent_length / _num_samples
+# ------------------------
+class _HasFit:
+    def fit(self):  # pragma: no cover - behavior is in _num_samples
+        return self
+
+
+def test_check_consistent_length_and_num_samples_errors():
+    # mismatch
+    with pytest.raises(ValueError):
+        V.check_consistent_length([1, 2, 3], [1])
+    # estimator-like with .fit -> TypeError in _num_samples
+    with pytest.raises(TypeError):
+        V._num_samples(_HasFit())
+    # 0-dim array -> TypeError
+    with pytest.raises(TypeError):
+        V._num_samples(np.array(1.0))
+    # normal paths
+    assert V._num_samples(np.array([[1, 2], [3, 4]])) == 2
+    assert V._num_samples([1, 2, 3]) == 3
+
+
[…]
