+++ kdiagram/config.py
@@ -41,7 +41,7 @@ def configure_warnings(
     modules: Optional[Iterable[Union[str, Pattern[str]]]] = None,
     clear: bool = False,
 ) -> None:
-    """
+    r"""
     Configure warning filters for callers (tests/docs/apps). No effect unless called.
 
     Parameters
@@ -88,7 +88,7 @@ def configure_warnings(
 
 @contextmanager
 def warnings_config(*args, **kwargs):
-    """
+    r"""
     Context manager that temporarily applies `configure_warnings(...)`.
 
     Example
@@ -106,7 +106,7 @@ def warnings_config(*args, **kwargs):
 
 # Backward-compatible shim for the old API (now opt-in & deprecated)
 def suppress_warnings(suppress: bool = True) -> None:
-    """
+    r"""
     DEPRECATED: Use `configure_warnings(...)` instead.
 
     - suppress=True  -> ignore SyntaxWarning (was the previous default)
