### 2025-08-11 9dbcc5d
add more tests
