+++ tests/test_io.py
@@ -4,20 +4,21 @@ from __future__ import annotations
 import builtins
 import sys
 import types
-
 from pathlib import Path
 
 import pytest
 
 import kdiagram.utils.io as io_mod
 
-
 # ---- helpers ---------------------------------------------------------------
+
 
 class FakeResponse:
     """Context-manager mock for requests.get(..., stream=True)."""
-    def __init__(self, data: bytes, status: int = 200,
-                 content_length: int | None = None):
+
+    def __init__(
+        self, data: bytes, status: int = 200, content_length: int | None = None
+    ):
         self._data = data
         self.status_code = status
         clen = len(data) if content_length is None else content_length
@@ -41,6 +42,7 @@ class FakeResponse:
 
 
 # ---- fancier_downloader tests ---------------------------------------------
+
 
 def test_fancier_downloader_requires_requests(monkeypatch):
     monkeypatch.setattr(io_mod, "HAS_REQUESTS", False)
@@ -88,7 +90,9 @@ def test_fancier_downloader_success_with_progress(tmp_path, monkeypatch):
     if "tqdm" not in sys.modules:
         fake_tqdm_mod = types.SimpleNamespace(
             tqdm=lambda **kw: types.SimpleNamespace(
-                update=lambda *a, **k: None, close=lambda: None))
+                update=lambda *a, **k: None, close=lambda: None
+            )
+        )
         sys.modules["tqdm"] = types.SimpleNamespace(tqdm=fake_tqdm_mod.tqdm)
 
     out = io_mod.fancier_downloader("http://x", "ok.bin", verbose=False)
@@ -103,19 +107,24 @@ def test_fancier_downloader_check_size_mismatch(monkeypatch, tmp_path, mode):
     monkeypatch.chdir(tmp_path)
     data = b"x" * 50
     monkeypatch.setattr(
-        io_mod.requests, "get", lambda *a, **k: FakeResponse(
-            data, content_length=100)
+        io_mod.requests, "get", lambda *a, **k: FakeResponse(data, content_length=100)
     )
 
     # Provide minimal tqdm stub if missing
     if "tqdm" not in sys.modules:
         sys.modules["tqdm"] = types.SimpleNamespace(
             tqdm=lambda **kw: types.SimpleNamespace(
-                update=lambda *a, **k: None, close=lambda: None)
+                update=lambda *a, **k: None, close=lambda: None
+            )
         )
 
-    kwargs = dict(url="http://x", filename="mismatch.bin", 
-                  check_size=True, error=mode, verbose=False)
+    kwargs = dict(
+        url="http://x",
+        filename="mismatch.bin",
+        check_size=True,
+        error=mode,
+        verbose=False,
+    )
 
     if mode == "raise":
         with pytest.raises(RuntimeError, match="Downloaded file size"):
@@ -132,33 +141,33 @@ def test_fancier_downloader_check_size_mismatch(monkeypatch, tmp_path, mode):
 def test_fancier_downloader_moves_to_dst(monkeypatch, tmp_path):
     data = b"hello world"
     monkeypatch.chdir(tmp_path)
-    monkeypatch.setattr(io_mod.requests, "get", 
-                        lambda *a, **k: FakeResponse(data))
+    monkeypatch.setattr(io_mod.requests, "get", lambda *a, **k: FakeResponse(data))
     if "tqdm" not in sys.modules:
         sys.modules["tqdm"] = types.SimpleNamespace(
             tqdm=lambda **kw: types.SimpleNamespace(
-                update=lambda *a, **k: None, close=lambda: None)
+                update=lambda *a, **k: None, close=lambda: None
+            )
         )
 
     dest = tmp_path / "dest"
[…]
