+++ tests/test_compat_sklearn.py
@@ -1,9 +1,10 @@
 # tests/test_compat_sklearn.py
 from __future__ import annotations
 
+import inspect
 import sys
 import types
-import inspect
+from numbers import Integral
 from types import SimpleNamespace
 
 import numpy as np
@@ -28,30 +29,27 @@ def test_interval_inclusive_removed_branch(monkeypatch):
         if obj is skl_compat.sklearn_Interval.__init__:
             # build a signature that does NOT include 'inclusive'
             return inspect.Signature(
-            parameters=[
-                inspect.Parameter(
-                    "self", inspect.Parameter.POSITIONAL_ONLY),
-                inspect.Parameter(
-                    "types", inspect.Parameter.POSITIONAL_OR_KEYWORD),
-                inspect.Parameter(
-                    "left", inspect.Parameter.POSITIONAL_OR_KEYWORD),
-                inspect.Parameter(
-                    "right", inspect.Parameter.POSITIONAL_OR_KEYWORD),
-                inspect.Parameter(
-                    "closed", inspect.Parameter.POSITIONAL_OR_KEYWORD),
-            ]
-        )
+                parameters=[
+                    inspect.Parameter("self", inspect.Parameter.POSITIONAL_ONLY),
+                    inspect.Parameter("types", inspect.Parameter.POSITIONAL_OR_KEYWORD),
+                    inspect.Parameter("left", inspect.Parameter.POSITIONAL_OR_KEYWORD),
+                    inspect.Parameter("right", inspect.Parameter.POSITIONAL_OR_KEYWORD),
+                    inspect.Parameter(
+                        "closed", inspect.Parameter.POSITIONAL_OR_KEYWORD
+                    ),
+                ]
+            )
         return real_sig
 
     monkeypatch.setattr(skl_compat.inspect, "signature", fake_signature)
     # Should NOT error even if we pass inclusive; shim must drop it.
-    obj = skl_compat.Interval(int, 0, 1, closed="left", inclusive=True)
+    obj = skl_compat.Interval(Integral, 0, 1, closed="left", inclusive=True)
     assert isinstance(obj, skl_compat.sklearn_Interval)
 
 
 def test_interval_inclusive_supported_branch():
     # On modern sklearn, passing inclusive should be accepted
-    obj = skl_compat.Interval(int, 0, 1, closed="left", inclusive=True)
+    obj = skl_compat.Interval(Integral, 0, 1, closed="left", inclusive=True)
     assert isinstance(obj, skl_compat.sklearn_Interval)
 
 
@@ -96,8 +94,7 @@ def test_validate_params_with_and_without_flag(monkeypatch):
     params = {"x": [int], "y": [str]}
 
     # Case 1: Signature contains prefer_skip_nested_validation (real modern sklearn)
-    dec = skl_compat.validate_params(
-        params, prefer_skip_nested_validation=False)
+    dec = skl_compat.validate_params(params, prefer_skip_nested_validation=False)
 
     @dec
     def fn(x, y):
@@ -112,8 +109,11 @@ def test_validate_params_with_and_without_flag(monkeypatch):
 
     def fake_sig(_):
         # Drop the flag from the signature
-        pars = [p for p in real_sig.parameters.values(
-            ) if p.name != "prefer_skip_nested_validation"]
+        pars = [
+            p
+            for p in real_sig.parameters.values()
+            if p.name != "prefer_skip_nested_validation"
+        ]
         return real_sig.replace(parameters=pars)
 
[…]
