### 2025-08-11 d317f2b
Fix colormap issue in matplotlib 3.10
