+++ tests/test_plot_reliability_diagram.py
@@ -132,8 +132,9 @@ def test_2d_probabilities_with_and_without_class_index(rng):
     P = np.column_stack([1 - p_pos, p_pos]).astype(float)
 
     # class_index omitted -> last column used; expect a warning per docs
-    with warnings.catch_warnings(record=True) as rec:
-        warnings.simplefilter("always")
+    # with warnings.catch_warnings(record=True) as rec:
+    #     warnings.simplefilter("always")
+    with pytest.warns(UserWarning):
         ax1 = plot_reliability_diagram(
             y,
             P,
@@ -144,11 +145,11 @@ def test_2d_probabilities_with_and_without_class_index(rng):
             return_data=False,
         )
     # at least one warning about class_index or similar behavior
-    assert any(
-        ("matplotlib is" in str(w.message).lower())
-        or ("figurecanvasagg" in str(w.message).lower())
-        for w in rec
-    )
+    # assert any(
+    #     ("matplotlib is" in str(w.message).lower())
+    #     or ("figurecanvasagg" in str(w.message).lower())
+    #     for w in rec
+    # )
     _close(ax1)
 
     # explicit class_index=0 uses negative-class prob; should still run
