### 2025-08-11 1ad0253
revise matplotlib.compat
