+++ kdiagram/compat/matplotlib.py
@@ -19,127 +19,122 @@ _MPL_VERSION = parse(matplotlib.__version__)
 __all__ = ["get_cmap", "is_valid_cmap"]
 
 
-def _get_cmap(name="viridis", lut=None):
+def get_cmap(name="viridis", lut=None, allow_none=False):
+    """Get a Matplotlib colormap with version compatibility.
+
+    This function acts as a robust wrapper to retrieve a colormap
+    object, ensuring consistent behavior across different
+    Matplotlib versions. It handles the API change from
+    `matplotlib.cm.get_cmap` to `matplotlib.colormaps.get`.
+
+    Parameters
+    ----------
+    name : str or None, optional
+        The name of the colormap to retrieve. Defaults to "viridis".
+    lut : int, optional
+        The number of colors in the lookup table. This parameter
+        is only used for Matplotlib versions older than 3.6 and
+        is ignored in newer versions. Defaults to None.
+    allow_none : bool, optional
+        If True, passing `name=None` will return `None` without
+        raising an error. If False (default), passing `name=None`
+        will raise a `ValueError`. This is the primary control for
+        preventing unexpected `TypeError` exceptions downstream.
+
+    Returns
+    -------
+    matplotlib.colors.Colormap or None
+        The requested colormap instance. Returns `None` only if
+        `name` is None and `allow_none` is True.
+
+    Raises
+    ------
+    ValueError
+        If `name` is `None` and `allow_none` is False, if `name` is
+        not a string (and not None), or if the named colormap
+        does not exist.
     """
-    A compatibility wrapper for getting a colormap.
+    if name is None:
+        if allow_none:
+            return None
+        else:
+            raise ValueError(
+                "Colormap `name` cannot be None when `allow_none` is False."
+            )
 
-    Handles the deprecation of `matplotlib.cm.get_cmap` in favor of
-    `matplotlib.colormaps.get()` in Matplotlib v3.7+.
+    if not isinstance(name, str):
+        raise ValueError(f"Colormap name must be a string, not {type(name)}.")
 
-    Args:
-        name (str, optional): The name of the colormap. Defaults to "viridis".
-        lut (int, optional): The number of colors in the lookup table.
-
-    Returns:
-        A Matplotlib colormap instance.
-    """
-    # The new API was introduced in 3.6 but get_cmap was deprecated in 3.7.
-    # We check for >= 3.6 for safety.
-    if _MPL_VERSION >= parse("3.6"):
-        # Use the new, recommended API
-        return matplotlib.colormaps.get(name, None)
-    else:
-        # Use the old, deprecated API for older versions
-        return matplotlib.cm.get_cmap(name, lut)
-
-
-def get_cmap(name="viridis", lut=None):
-    """
-    A compatibility wrapper for getting a colormap that consistently
-    raises a ValueError for invalid names across Matplotlib versions.
-
-    Handles the deprecation of `matplotlib.cm.get_cmap` in favor of
-    `matplotlib.colormaps.get()` in Matplotlib v3.7+.
-
-    Args:
-        name (str, optional): The name of the colormap. Defaults to "viridis".
-        lut (int, optional): The number of colors in the lookup table.
-                             (Note: This is ignored in Matplotlib >= 3.6)
-
-    Returns:
-        A Matplotlib colormap instance.
-
-    Raises:
-        ValueError: If the colormap name is not valid.
-    """
     if _MPL_VERSION >= parse("3.6"):
         # The new API raises KeyError for invalid names. We catch it and
         # re-raise as ValueError to match the old API's behavior. This
         # ensures that try/except ValueError blocks in the main code
         # continue to work as expected.
+        
         try:
             return matplotlib.colormaps.get(name)
         except (TypeError, KeyError, ValueError) as err:
-            # Re-raise as ValueError for consistent exception handling.
[…]
