### 2025-08-11 6e9f340
fix matplotlib compatibility
