+++ kdiagram/compat/matplotlib.py
@@ -19,7 +19,7 @@ _MPL_VERSION = parse(matplotlib.__version__)
 __all__ = ["get_cmap", "is_valid_cmap"]
 
 
-def get_cmap(name="viridis", lut=None, allow_none=False):
+def _get_cmap(name="viridis", lut=None, allow_none=False):
     """Get a Matplotlib colormap with version compatibility.
 
     This function acts as a robust wrapper to retrieve a colormap
@@ -82,8 +82,8 @@ def get_cmap(name="viridis", lut=None, allow_none=False):
         
         return matplotlib.cm.get_cmap(name, lut)
 
-def is_valid_cmap(cmap, default="viridis", error="raise"):
-    """Check if a colormap name is valid by attempting to retrieve it.
+def _is_valid_cmap(cmap, default="viridis", error="raise"):
+    """Legacy check if a colormap name is valid by attempting to retrieve it.
 
     This utility leverages the robust `get_cmap` wrapper to handle
     different Matplotlib versions and input types seamlessly.
@@ -117,7 +117,7 @@ def is_valid_cmap(cmap, default="viridis", error="raise"):
         # We call get_cmap with its safe default (allow_none=False)
         # because this function's purpose is to validate names.
         # It will raise a ValueError for any invalid input.
-        get_cmap(cmap, allow_none=False)
+        _get_cmap(cmap, allow_none=False)
         return cmap
     except ValueError:
         # The name is invalid, so handle it based on the 'error' flag.
@@ -138,3 +138,142 @@ def is_valid_cmap(cmap, default="viridis", error="raise"):
                 "'warn', or 'ignore'."
             )
             
+def is_valid_cmap(
+    cmap,  allow_none=False, 
+    **kw
+    ): # **for future extension 
+    r"""Check if a colormap identifier is valid.
+
+    This function purely validates whether a given identifier can be
+    resolved to a Matplotlib colormap. It does not retrieve the
+    colormap object itself.
+
+    Parameters
+    ----------
+    cmap : any
+        The colormap identifier to validate, typically a string.
+    allow_none : bool, optional
+        If True, `None` is considered a valid input and the
+        function will return True. If False (default), `None` is
+        considered invalid.
+
+    Returns
+    -------
+    bool
+        True if the `cmap` is a valid, retrievable colormap name
+        or if `cmap` is None and `allow_none` is True. Otherwise,
+        returns False.
+    """
+    if cmap is None:
+        return allow_none
+
+    if not isinstance(cmap, str):
+        return False
+
+    try:
+        # The most reliable way to check for existence is to try
+        # getting it, using the modern API first.
+        if _MPL_VERSION >= parse("3.6"):
+            matplotlib.colormaps.get(cmap)
+        else:
+            matplotlib.cm.get_cmap(cmap)
+        return True
+    except (ValueError, KeyError):
+        # ValueError is for old MPL, KeyError for new MPL.
+        return False
+    
+
+def get_cmap(
+    name, default='viridis', allow_none=False, 
+    error=None,
+    failsafe='continuous',
+    **kw
+   ):
+    r"""Robustly retrieve a Matplotlib colormap with fallbacks.
+
+    This function ensures a valid colormap object is always returned,
+    preventing runtime errors from invalid names. It uses a
+    cascading fallback system.
+
+    Parameters
+    ----------
+    name : str or None
+        The desired colormap name.
+    default : str, optional
+        The fallback colormap if `name` is invalid.
+        Defaults to 'viridis'.
+    allow_none : bool, optional
+        If True, a `name` of `None` will return `None` without
+        any warnings or errors. Defaults to False.
+    failsafe : {'continuous', 'discrete'}, optional
[…]
