+++ tests/test_uncertainty_plots.py
@@ -17,10 +17,8 @@ from kdiagram.plot.uncertainty import (
     plot_velocity,
 )
 
-
 def _cleanup():
     plt.close("all")
-
 
 # ---------- plot_coverage ----------
 def test_plot_coverage_line_bar_pie_radar(tmp_path):
@@ -66,7 +64,8 @@ def test_plot_coverage_line_bar_pie_radar(tmp_path):
     out3 = tmp_path / "cov_pie.png"
     preds_zero = np.ones_like(y_true) * 42
     plot_coverage(
-        y_true, preds_zero, names=["Zero"], kind="pie", title="pie", savefig=str(out3)
+        y_true, preds_zero, names=["Zero"], kind="pie",
+        title="pie", savefig=str(out3)
     )
     assert out3.exists()
 
@@ -94,8 +93,6 @@ def test_plot_coverage_bad_q_raises():
 
 
 # ---------- plot_model_drift ----------
-
-
 def test_plot_model_drift_half_circle_and_colors(tmp_path):
     # 3 horizons, widths positive
     n = 30
@@ -133,10 +130,7 @@ def test_plot_model_drift_half_circle_and_colors(tmp_path):
     assert not gx and not gy
     _cleanup()
 
-
 # ---------- plot_velocity ----------
-
-
 def test_plot_velocity_paths_and_warnings(tmp_path):
     # Build df with 3 time steps; include theta_col to trigger warning
     n = 40
@@ -194,7 +188,6 @@ def test_plot_velocity_missing_cols_raises():
 
 # ---------- plot_interval_consistency ----------
 
-
 def test_plot_interval_consistency_cv_and_std(tmp_path):
     n = 35
     df = pd.DataFrame(
@@ -247,9 +240,7 @@ def test_plot_interval_consistency_length_mismatch_raises():
         plot_interval_consistency(df, ["a"], ["b", "b2"])
 
 
-# ---------- plot_anomaly_magnitude ----------
-
-
+# ---------- plot_anomaly_magnitude ---------- 
 def test_plot_anomaly_magnitude_under_over_and_cbar(tmp_path):
     n = 60
     df = pd.DataFrame(
