+++ environment.yml
@@ -26,17 +26,6 @@ dependencies:
   - seaborn
   - scikit-learn
 
-  # Docs toolchain (from docs requirements + pyproject dev)
-  - sphinx>=7.0,<8.0
-  - furo
-  - sphinx-copybutton
-  - myst-parser
-  - sphinx-design
-  - sphinx-gallery
-  - sphinx-mathjax
-  - sphinx-issues
-  - sphinx-prompt
-
   # Testing / coverage
   - pytest>=8.0
   - pytest-xdist
@@ -61,6 +50,19 @@ dependencies:
   # - tensorflow
   # - pytorch
   # - jax
-
+  
+  # Docs toolchain (from docs requirements + pyproject dev)
+  # Pip-only packages (especially Sphinx extensions)
+  - pip:
+    - sphinx>=7.0,<8.0
+    - furo
+    - sphinx-copybutton
+    - myst-parser
+    - sphinx-design
+    - sphinx-gallery
+    - sphinx-mathjax-offline  
+    - sphinx-issues
+    - sphinx-prompt
+    
 # (Intentionally NOT installing the package here via `pip -e .`;
 # the workflow does `pip install . --no-deps` after the env is created.)
