+++ tests/test_compat_matplotlib.py
@@ -1,13 +1,13 @@
 # tests/test_compat_matplotlib.py
 from __future__ import annotations
 
+import warnings
+
 import matplotlib
 import pytest
 from packaging.version import parse
 
 from kdiagram.compat import matplotlib as mpl_compat
-
-import warnings
 
 
 def test_get_cmap_valid_returns_colormap():
@@ -40,35 +40,33 @@ def test_get_cmap_invalid_name_falls_back_to_valid_default_with_warning():
 def test_get_cmap_invalid_name_and_default_uses_failsafe_continuous():
     with warnings.catch_warnings(record=True) as w:
         warnings.simplefilter("always")
-        cmap = mpl_compat.get_cmap("nope", default="also_nope", 
-                                   failsafe="continuous")
+        cmap = mpl_compat.get_cmap("nope", default="also_nope", failsafe="continuous")
     # Current implementation falls back to viridis
     assert cmap.name.lower() == "viridis"
     # Two-step fallback: to default, then to failsafe
     msgs = [str(ww.message) for ww in w]
     assert any("Colormap 'nope' not found" in m for m in msgs)
-    assert any("Default colormap 'also_nope' also not found" 
-               in m for m in msgs)
+    assert any("Default colormap 'also_nope' also not found" in m for m in msgs)
 
 
 def test_get_cmap_invalid_name_and_default_uses_failsafe_discrete():
     with warnings.catch_warnings(record=True) as w:
         warnings.simplefilter("always")
-        cmap = mpl_compat.get_cmap("nope", default="also_nope",
-                                   failsafe="discrete")
+        cmap = mpl_compat.get_cmap("nope", default="also_nope", failsafe="discrete")
     # Implementation currently always returns viridis;
     # accept either viridis or tab10 to be forward-compatible.
     assert cmap.name.lower() in {"viridis", "tab10"}
     msgs = [str(ww.message) for ww in w]
     assert any("Colormap 'nope' not found" in m for m in msgs)
-    assert any("Default colormap 'also_nope' also not found"
-               in m for m in msgs)
+    assert any("Default colormap 'also_nope' also not found" in m for m in msgs)
+
 
 # def test_get_cmap_deprecated_error_param_emits_futurewarning():
 #     with pytest.warns(FutureWarning, match=re.compile(
 #             "deprecated.*error.*ignored", re.I)):
 #         cmap = mpl_compat.get_cmap("viridis", error="raise")
 #     assert cmap.name.lower() == "viridis"
+
 
 def test_is_valid_cmap_various_inputs():
     assert mpl_compat.is_valid_cmap("viridis") is True
@@ -93,7 +91,10 @@ def test_private_is_valid_cmap_modes():
     with warnings.catch_warnings(record=True) as w:
         warnings.simplefilter("always")
         out = mpl_compat._is_valid_cmap(
-            "no_such_map", error="warn", default="viridis",)
+            "no_such_map",
+            error="warn",
+            default="viridis",
+        )
     assert out == "viridis"
     assert any("Falling back to 'viridis'" in str(ww.message) for ww in w)
 
@@ -101,7 +102,8 @@ def test_private_is_valid_cmap_modes():
     with warnings.catch_warnings(record=True) as w:
         warnings.simplefilter("always")
         out = mpl_compat._is_valid_cmap(
-            "no_such_map", error="ignore", default="viridis")
+            "no_such_map", error="ignore", default="viridis"
+        )
     assert out == "viridis"
     assert len(w) == 0
 
@@ -115,6 +117,7 @@ def test_get_cmap_valid_new_api(monkeypatch):
 
     assert isinstance(cm, Colormap)
[…]
