+++ pyproject.toml
@@ -12,7 +12,7 @@ license = { text = "Apache-2.0" }
 requires-python = ">=3.9"
 keywords = ["forecasting", "uncertainty visualization", "model evaluation"]
 classifiers = [
-  "Development Status :: 3 - Alpha",
+  "Development Status :: 4 - Beta",
   "Intended Audience :: Science/Research",
   "Intended Audience :: Developers",
   "Topic :: Software Development",
