+++ joss_replies.md
@@ -1,230 +1,0 @@
-Thanks so much for the thoughtful, detailed review and for taking the time to go through the repo 
-so carefully. Your notes were spot-on and very actionable. Below I respond point-by-point 
-for the first two sections, including exactly what I changed.
-
-# Documentation
-
-***1) A11y (acessibility): When hovering over an item in the left table of contents, the contrast between font (turning white) and background is not high enough. How about changing the font to black when hovering?***
-
-**Action taken** — I reworked the sidebar hover/active styles to remove the Furo gradient and use a solid background with high-contrast text. Hover and active states now set a solid `background-color: var(--color-brand-primary)` and force the link color to `var(--color-background-primary)` (white on dark teal). This yields AA+ contrast and is much clearer than the previous gradient. Relevant CSS:
-
-```css
-.sidebar-drawer .reference:hover {
-  background-image: none !important;
-  background-color: var(--color-brand-primary) !important;
-  color: var(--color-background-primary) !important;
-  padding-left: 1.1em;
-}
-.sidebar-drawer li.current > a.reference {
-  background-image: none !important;
-  background-color: var(--color-brand-primary) !important;
-  color: var(--color-background-primary) !important;
-  font-weight: 600;
-}
-```
-
-***2) It seems that docstrings use numpydoc, but the Sphinx documentation does not use the extension ([https://numpydoc.readthedocs.io/en/latest/index.html](https://numpydoc.readthedocs.io/en/latest/index.html)), even though napoleon\_numpy\_docstring = True is set, resulting in Reference pages where the docstrings are not parsed correctly (see e.g. [https://k-diagram.readthedocs.io/en/latest/\_autosummary/uncertainty/kdiagram.plot.uncertainty.plot\_anomaly\_magnitude.html](https://k-diagram.readthedocs.io/en/latest/_autosummary/uncertainty/kdiagram.plot.uncertainty.plot_anomaly_magnitude.html))***
-
-**Action taken** — I installed and enabled the **numpydoc** extension in `conf.py` (alongside Napoleon), then fixed a few docstring issues (duplicate “Notes” sections and indentation) that were causing parse errors. Autosummary/API pages like `plot_anomaly_magnitude` now render Parameters/Notes/Examples and math blocks correctly.
-
-```diff
-extensions = [
-+   'numpydoc',
-    'sphinx.ext.autodoc',
-    'sphinx.ext.autosummary',
-    'sphinx.ext.napoleon',
-    ...
-]
-```
-
-# Packaging
-
-***1) Package uses setup.py which is not deprecated, but using a pyproject.toml is preferable. See [https://packaging.python.org/en/latest/guides/modernize-setup-py-project/](https://packaging.python.org/en/latest/guides/modernize-setup-py-project/)***
-
-**Action taken** — I migrated to a PEP 621 `pyproject.toml` and removed all metadata from `setup.py`. Builds now use the PEP 517 backend (`setuptools.build_meta`). If necessary for legacy workflows, I can keep a tiny shim `setup.py` that just calls `setuptools.setup()` with no metadata.
-
-***2) Having pyproject.toml also allows more settings to be included there, instead of having separate files:
-pytest.ini
-.coveragerc
-.flake8***
-
-**Action taken** — I consolidated tool configuration into `pyproject.toml`:
-
-* `pytest` under `[tool.pytest.ini_options]` (including `testpaths`, `addopts`, and targeted `filterwarnings` for docs/CI),
-* coverage under `[tool.coverage.run]` / `[tool.coverage.report]`,
-* formatting under `[tool.black]`,
-* linting under `[tool.ruff]`.
-  This removes the separate `pytest.ini`, `.coveragerc`, and `.flake8` files.
-
[…]
