+++ tests/test_datasets_make.py
@@ -1,0 +1,241 @@
+import numpy as np
+import pandas as pd
+import pytest
+
+from kdiagram.datasets import make as mk
+
+
+# ------------------------------
+# make_cyclical_data
+# ------------------------------
+
+def test_make_cyclical_data_bunch_and_frame_shapes():
+    n = 24
+    s = 2
+    bunch = mk.make_cyclical_data(
+        n_samples=n, n_series=s, seed=7, as_frame=False
+    )
+    assert hasattr(bunch, "frame")
+    assert bunch.frame.shape[0] == n
+    # expected columns: y_true, time_step, s prediction cols
+    assert len(bunch.prediction_columns) == s
+    assert set(bunch.feature_names) == {"time_step"}
+    assert set(bunch.target_names) == {"y_true"}
+    # as_frame=True ordering check
+    df = mk.make_cyclical_data(
+        n_samples=n, n_series=s, seed=7, as_frame=True
+    )
+    assert list(df.columns[:2]) == ["y_true", "time_step"]
+    assert sum(c.startswith("model_") for c in df.columns) == s
+
+
+def test_make_cyclical_data_series_and_param_validation_errors():
+    # series_names length must match n_series
+    with pytest.raises(
+            ValueError, match=r"Length of series_names .* must match"):
+        mk.make_cyclical_data(n_samples=10, 
+                              n_series=2, series_names=["only_one"])
+    # list length mismatch for scalar/list params
+    with pytest.raises(ValueError,
+                       match=r"Length of 'pred_bias'"):
+        mk.make_cyclical_data(n_samples=10, 
+                              n_series=3, pred_bias=[0.0, 1.0])
+    # wrong type for param
+    with pytest.raises(TypeError, 
+                       match=r"'pred_noise_factor' must be float or list"):
+        mk.make_cyclical_data(
+            n_samples=10, n_series=2, pred_noise_factor=(1.0, 1.5))
+
+
+def test_make_cyclical_data_seed_reproducibility():
+    a = mk.make_cyclical_data(n_samples=16, seed=123, as_frame=True)
+    b = mk.make_cyclical_data(n_samples=16, seed=123, as_frame=True)
+    np.testing.assert_allclose(
+        a["y_true"].to_numpy(), b["y_true"].to_numpy())
+
+
+# ------------------------------
+# make_fingerprint_data
+# ------------------------------
+
+def test_make_fingerprint_data_bunch_and_frame_and_validations():
+    # basic bunch
+    bunch = mk.make_fingerprint_data(
+        n_layers=4, n_features=6, seed=3, as_frame=False
+    )
+    assert bunch.importances.shape == (4, 6)
+    assert bunch.frame.shape == (4, 6)
+    assert list(bunch.frame.index) == bunch.layer_names
+    assert list(bunch.frame.columns) == bunch.feature_names
+
+    # as_frame
+    df = mk.make_fingerprint_data(
+        n_layers=2, n_features=3, seed=4, as_frame=True)
+    assert df.shape == (2, 3)
+
+    # validations
+    with pytest.raises(ValueError,
+                       match="sparsity must be between 0.0 and 1.0"):
+        mk.make_fingerprint_data(sparsity=1.5)
+    with pytest.raises(ValueError, 
+                       match=r"value_range must be a tuple"):
+        mk.make_fingerprint_data(value_range=(2.0, 1.0))
+    with pytest.raises(ValueError, 
+                       match=r"feature_names .* must match"):
+        mk.make_fingerprint_data(n_features=3, feature_names=["F1"])
+    with pytest.raises(ValueError, 
[…]
