+++ tests/test_uncertainty.py
@@ -129,11 +129,6 @@ def sample_data_velocity():
     return {"df": df, "q50_cols": q50_cols}
 
 
-# --- Test Functions ---
-
-# == Tests for plot_coverage ==
-
-
 @pytest.mark.parametrize("kind", ["line", "bar", "pie", "radar"])
 def test_plot_coverage_single_model_quantile(sample_data_coverage, kind):
     """Test plot_coverage with one quantile model for various kinds."""
@@ -198,12 +193,6 @@ def test_plot_coverage_invalid_q(sample_data_coverage):
 
     with pytest.raises(ValueError, match="1D list or array"):
         plot_coverage(data["y_true"], data["y_pred_q1"], q=[[0.1], [0.9]])
-
-
-# Note: Test for q vs pred dimension mismatch is skipped as the
-# check is commented out in the source code provided.
-
-# == Tests for plot_model_drift ==
 
 
 @pytest.mark.parametrize(
@@ -287,10 +276,6 @@ def test_plot_model_drift_mismatched_lengths(sample_data_drift):
             q90_cols=data["q90_cols"],
             horizons=data["horizons"],
         )
-
-
-# == Tests for plot_velocity ==
-
 
 @pytest.mark.parametrize("normalize", [True, False])
 @pytest.mark.parametrize("use_abs_color", [True, False])
@@ -448,12 +433,6 @@ def sample_data_anomaly(request):
         "has_anomalies": generate_anomalies,  # Flag for assertions
     }
 
-
-# --- Test Functions ---
-
-# == Tests for plot_interval_consistency ==
-
-
 @pytest.mark.parametrize("use_cv", [True, False])
 @pytest.mark.parametrize("q50_provided", [True, False])
 @pytest.mark.parametrize("acov", ["default", "half_circle"])
@@ -551,10 +530,6 @@ def test_plot_interval_consistency_zero_mean_warning(sample_data_consistency):
             use_cv=True,  # Warning occurs during CV calculation
         )
 
-
-# == Tests for plot_anomaly_magnitude ==
-
-
 @pytest.mark.parametrize("theta_col_provided", [True, False])
 @pytest.mark.parametrize("acov", ["default", "quarter_circle"])
 @pytest.mark.parametrize("cbar", [True, False])
@@ -639,18 +614,6 @@ def test_plot_anomaly_magnitude_missing_cols(sample_data_anomaly):
 def test_plot_anomaly_magnitude_theta_col_warning(sample_data_anomaly):
     """Test warnings related to theta_col."""
     data = sample_data_anomaly
-    # df_mod = data["df"].copy()
-
-    # # Case 1: theta_col exists but has non-numeric data (after NaN drop)
-    # df_mod["order_feature_str"] = df_mod["order_feature"].astype(str)
-    # with pytest.warns(UserWarning):
-    #     plot_anomaly_magnitude(
-    #         df=df_mod,
-    #         actual_col=data["actual_col"],
-    #         q_cols=data["q_cols"],
-    #         theta_col="order_feature_str",  # Non-numeric
-    #     )
-
     # Case 2: theta_col does not exist
     with pytest.raises(KeyError):
         plot_anomaly_magnitude(
@@ -799,7 +762,6 @@ def sample_data_temporal():
     }
 
 
-# == Tests for plot_uncertainty_drift ==
 @pytest.mark.parametrize("dt_labels_provided", [True, False])
 @pytest.mark.parametrize("acov", ["default", "quarter_circle"])
 @pytest.mark.parametrize("mask_angle", [True, False])
@@ -899,38 +861,49 @@ def test_plot_uncertainty_drift_width_warnings(sample_data_drift_uncertainty):
             df=df_zero, qlow_cols=data["qlow_cols"], qup_cols=data["qup_cols"]
         )
 
[…]
