+++ tests/test_cli_.py
@@ -1,0 +1,288 @@
+import sys
+import csv
+import numpy as np
+import pandas as pd
+import matplotlib.pyplot as plt
+import pytest
+
+import kdiagram.cli as cli
+
+
+@pytest.fixture(autouse=True)
+def _matplotlib_agg_backend(monkeypatch):
+    # keep GUI backends out of CI
+    import matplotlib
+
+    matplotlib.use("Agg", force=True)
+    monkeypatch.setattr(plt, "show", lambda *a, **k: None)
+
+
+def _write_csv(path, rows):
+    path = str(path)
+    with open(path, "w", newline="") as f:
+        w = csv.writer(f)
+        if (
+            isinstance(rows, (list, tuple))
+            and rows
+            and not isinstance(rows[0], (list, tuple))
+        ):
+            for r in rows:
+                w.writerow([r])
+        elif isinstance(rows, (list, tuple)):
+            w.writerows(rows)
+        else:
+            f.write(str(rows))
+    return path
+
+
+def _stub_calls(sink):
+    def _stub(*args, **kwargs):
+        # ensure there's a figure to save
+        plt.figure()
+        sink.append({"args": args, "kwargs": kwargs})
+
+    return _stub
+
+
+def test_read_csv_to_df_success_and_errors(tmp_path, capsys):
+    p = tmp_path / "data.csv"
+    pd.DataFrame({"a": [1, 2], "b": [3, 4]}).to_csv(p, index=False)
+
+    df = cli.read_csv_to_df(str(p))
+    assert isinstance(df, pd.DataFrame)
+    assert df.shape == (2, 2)
+
+    # not found
+    df2 = cli.read_csv_to_df(str(tmp_path / "missing.csv"))
+    assert df2 is None
+    _, err = capsys.readouterr()
+    assert "File not found" in err
+
+    # passing a directory trips the generic exception branch
+    df3 = cli.read_csv_to_df(str(tmp_path))
+    assert df3 is None
+
+
+def test_read_csv_to_numpy_scalar_array_and_notfound(tmp_path, capsys):
+    p_scalar = _write_csv(tmp_path / "val.csv", [5])
+    arr_s = cli.read_csv_to_numpy(p_scalar)
+    assert isinstance(arr_s, np.ndarray)
+    assert arr_s.ndim == 1 and arr_s.size == 1
+
+    p_arr = _write_csv(tmp_path / "arr.csv", [[1, 2, 3], [4, 5, 6]])
+    arr = cli.read_csv_to_numpy(p_arr)
+    assert arr.shape == (2, 3)
+
+    arr2 = cli.read_csv_to_numpy(str(tmp_path / "nope.csv"))
+    assert arr2 is None
+    _, err = capsys.readouterr()
+    assert "File not found" in err
+
+
+def test_handle_figsize_valid_and_invalid(capsys):
+    assert cli._handle_figsize("8,6", (1, 1)) == (8.0, 6.0)
+
+    # non-numeric
+    out = cli._handle_figsize("bad", (9, 9))
+    assert out == (9, 9)
+    _, err = capsys.readouterr()
+    assert "Invalid format for figsize" in err
+
+    # wrong arity
+    out = cli._handle_figsize("1,2,3", (7, 7))
+    assert out == (7, 7)
+
+
+def test_handle_savefig_show_writes_file_and_handles_show_error(
+    tmp_path, monkeypatch, capsys
+):
+    # save path creates a file
+    plt.figure()
+    out = tmp_path / "plot.png"
+    cli._handle_savefig_show(str(out))
+    assert out.exists() and out.stat().st_size > 0
+
+    # simulate show() failure -> friendly stderr note
[…]
