+++ tests/test_datasets_.py
@@ -78,7 +78,8 @@ def test_load_uncertainty_data():
 
 def test_load_uncertainty_data_anomalies():
     """Test synthetic data with anomalies."""
-    result = load_uncertainty_data(n_samples=10, anomaly_frac=0.2, as_frame=True)
+    result = load_uncertainty_data(
+        n_samples=10, anomaly_frac=0.2, as_frame=True)
     
     # Check if anomalies are generated by verifying values 
     # outside the expected range
