+++ tests/test_q_utils.py
@@ -1,7 +1,7 @@
-
 import re
-import pandas as pd
+
 import numpy as np
+import pandas as pd
 import pytest
 
 from kdiagram.utils import q_utils
@@ -40,8 +40,7 @@ def long_df():
 # ---------- reshape_quantile_data ----------
 def test_reshape_quantile_data_basic(wide_df):
     out = q_utils.reshape_quantile_data(
-        wide_df, value_prefix="subs", 
-        spatial_cols=["lon", "lat"], dt_col="year"
+        wide_df, value_prefix="subs", spatial_cols=["lon", "lat"], dt_col="year"
     )
     # expected columns exist
     assert set(["lon", "lat", "year"]).issubset(out.columns)
@@ -64,11 +63,9 @@ def test_reshape_quantile_data_missing_spatial_warns_and_continues(wide_df):
 
 
 def test_reshape_quantile_data_no_matching_prefix(wide_df):
-    with pytest.warns(
-            UserWarning, match=r"No columns found with prefix 'other'"):
+    with pytest.warns(UserWarning, match=r"No columns found with prefix 'other'"):
         out = q_utils.reshape_quantile_data(
-            wide_df, value_prefix="other", 
-            spatial_cols=["lon", "lat"], error="warn"
+            wide_df, value_prefix="other", spatial_cols=["lon", "lat"], error="warn"
         )
     assert isinstance(out, pd.DataFrame) and out.empty
 
@@ -83,7 +80,8 @@ def test_reshape_quantile_data_invalid_pattern_returns_empty():
         }
     )
     out = q_utils.reshape_quantile_data(
-        df, value_prefix="subs", spatial_cols=["lon", "lat"])
+        df, value_prefix="subs", spatial_cols=["lon", "lat"]
+    )
     assert out.empty
 
 
@@ -122,8 +120,7 @@ def test_melt_q_data_filter_quantiles(wide_df):
 
 
 def test_melt_q_data_no_match_prefix_warns_and_empty(wide_df):
-    with pytest.warns(
-            UserWarning, match=r"No columns found with prefix 'other'"):
+    with pytest.warns(UserWarning, match=r"No columns found with prefix 'other'"):
         out = q_utils.melt_q_data(
             wide_df, value_prefix="other", dt_name="year", error="warn"
         )
@@ -132,10 +129,7 @@ def test_melt_q_data_no_match_prefix_warns_and_empty(wide_df):
 
 def test_melt_q_data_invalid_spatial_cols_raises(wide_df):
     # spatial col 'y' does not exist
-    with pytest.raises(ValueError, match=re.escape ( 
-            "The following spatial_cols are"
-            " not present in the dataframe: {'y', 'x'}"
-            )):
+    with pytest.raises(ValueError):
         q_utils.melt_q_data(
             wide_df,
             value_prefix="subs",
@@ -180,9 +174,7 @@ def test_pivot_q_data_basic(long_df):
     # Must have lon/lat and reconstructed quantile/year columns
     assert set(["lon", "lat"]).issubset(out.columns)
     assert any(
-        c.startswith(
-            "subs_2022_q") or c.startswith("subs_2023_q"
-                                           ) for c in out.columns
+        c.startswith("subs_2022_q") or c.startswith("subs_2023_q") for c in out.columns
     )
     # check a known value location (rough presence check)
     # there should be at least these two reconstructed columns
@@ -204,10 +196,8 @@ def test_pivot_q_data_missing_dt_col_warns_and_empty(long_df):
 
 
 def test_pivot_q_data_no_quantile_columns_warns_and_empty():
-    df = pd.DataFrame({"lon": [0, 0], "lat": [0, 0], 
[…]
