+++ tests/test_datasets_.py
@@ -1,105 +1,125 @@
-
 import os
-import pytest
-from kdiagram.datasets.load import load_zhongshan_subsidence, load_uncertainty_data
-from kdiagram.api.bunch import Bunch
-import pandas as pd
+
 import numpy as np
+import pandas as pd
+import pytest
+
+from kdiagram.api.bunch import Bunch
+from kdiagram.datasets.load import load_uncertainty_data, load_zhongshan_subsidence
+
 
 @pytest.fixture
 def mock_load_zhongshan_subsidence(mocker):
     # Mock get_data to return the correct directory
-    mocker.patch("kdiagram.datasets.load.get_data",
-                 return_value=os.path.join(
-                     os.path.dirname(__file__),
-                    'kdiagram', 'datasets', 'data'))
+    mocker.patch(
+        "kdiagram.datasets.load.get_data",
+        return_value=os.path.join(
+            os.path.dirname(__file__), "kdiagram", "datasets", "data"
+        ),
+    )
 
     # Mock the file download path
-    mocker.patch("kdiagram.datasets.load.download_file_if",
-                 return_value=os.path.join(
-                     os.path.dirname(__file__), 'kdiagram', 
-                     'datasets', 'data', 'min_zhongshan.csv'))
-    
+    mocker.patch(
+        "kdiagram.datasets.load.download_file_if",
+        return_value=os.path.join(
+            os.path.dirname(__file__),
+            "kdiagram",
+            "datasets",
+            "data",
+            "min_zhongshan.csv",
+        ),
+    )
+
     # Mock the package resource path
-    mocker.patch("kdiagram.datasets.load.resources.files", 
-                 return_value=os.path.join(os.path.dirname(__file__),
-                                           'kdiagram', 'datasets', 'data')
-                 )
+    mocker.patch(
+        "kdiagram.datasets.load.resources.files",
+        return_value=os.path.join(
+            os.path.dirname(__file__), "kdiagram", "datasets", "data"
+        ),
+    )
+
+
 def test_load_zhongshan_subsidence_local():
     """Test loading Zhongshan subsidence dataset from local cache."""
     result = load_zhongshan_subsidence(as_frame=True)
-    
+
     # Check that the result is a DataFrame
     assert isinstance(result, pd.DataFrame)
-    
+
     # Check that specific columns exist (e.g., 'longitude' and 'latitude')
     assert "longitude" in result.columns
     assert "latitude" in result.columns
+
 
 def test_load_zhongshan_subsidence_invalid_years():
     """Test loading Zhongshan subsidence dataset with invalid years."""
     # Requesting invalid years (e.g., years not present in dataset)
     result = load_zhongshan_subsidence(years=[2030, 2040], as_frame=True)
-    
+
     # Ensure warning is raised, and the DataFrame is returned
     assert isinstance(result, pd.DataFrame)
     # Check if the correct years are filtered in the dataset
     assert not all(year in result.columns for year in [2022, 2023])
 
+
 def test_load_zhongshan_subsidence_with_quantiles():
     """Test loading Zhongshan subsidence dataset with specific quantiles."""
     result = load_zhongshan_subsidence(quantiles=[0.1, 0.9], as_frame=True)
-    
+
     # Check if the expected quantile columns are present
-    assert 'subsidence_2022_q0.1' in result.columns
-    assert 'subsidence_2022_q0.9' in result.columns
+    assert "subsidence_2022_q0.1" in result.columns
+    assert "subsidence_2022_q0.9" in result.columns
+
 
 def test_load_zhongshan_subsidence_no_coords():
     """Test loading Zhongshan subsidence dataset excluding coordinates."""
[…]

[…]
