### 2025-08-12 0b3e390
fix compat issue py39 and py312
