+++ tests/test_datasets__property.py
@@ -1,0 +1,260 @@
+
+import shutil
+from pathlib import Path
+
+import pytest
+
+from kdiagram.datasets import _property as prop
+
+
+# ---------- helpers ----------
+
+def _touch(p: Path, data: bytes = b"data"):
+    p.parent.mkdir(parents=True, exist_ok=True)
+    p.write_bytes(data)
+    return p
+
+
+class _PathCtx:
+    """Simple context manager to mimic importlib.resources.path."""
+    def __init__(self, p: Path):
+        self._p = p
+
+    def __enter__(self):
+        return self._p
+
+    def __exit__(self, exc_type, exc, tb):
+        return False
+
+
+# ---------- get_data / remove_data ----------
+
+def test_get_data_env_and_custom(tmp_path, monkeypatch):
+    # env variable path takes precedence
+    env_dir = tmp_path / "envdir"
+    monkeypatch.setenv("KDIAGRAM_DATA", str(env_dir))
+    got = prop.get_data()
+    assert Path(got) == env_dir
+    assert env_dir.exists()
+
+    # explicit path overrides env
+    custom = tmp_path / "custom"
+    got2 = prop.get_data(str(custom))
+    assert Path(got2) == custom
+    assert custom.exists()
+
+
+def test_remove_data_deletes_dir(tmp_path):
+    d = tmp_path / "delme"
+    # create via get_data to mimic normal behavior
+    prop.get_data(str(d))
+    assert d.exists()
+    prop.remove_data(str(d))
+    assert not d.exists()
+
+
+# ---------- download_file_if : package resource path ----------
+
+def test_download_from_package_copies_to_cache(tmp_path, monkeypatch):
+    cache_dir = tmp_path / "cache"
+    filename = "sample.txt"
+    pkg_file = tmp_path / "pkg" / filename
+    _touch(pkg_file, b"pkgdata")
+
+    # Mock importlib.resources APIs used by the module
+    monkeypatch.setattr(prop.resources, "is_resource",
+                        lambda mod, name: name == filename)
+    monkeypatch.setattr(prop.resources, "path",
+                        lambda mod, name: _PathCtx(pkg_file))
+
+    # Ensure downloader is never called in this path
+    monkeypatch.setattr(prop, "fancier_downloader",
+                        lambda **k: pytest.fail("downloader should not run"))
+
+    out = prop.download_file_if(filename, data_home=str(cache_dir),
+                                download_if_missing=False, verbose=False)
+    assert out is not None
+    out_path = Path(out)
+    assert out_path.parent == cache_dir
+    assert out_path.read_bytes() == b"pkgdata"
+
+
+def test_download_uses_cache_if_already_present(tmp_path, monkeypatch):
+    cache_dir = tmp_path / "cache"
+    filename = "already.txt"
+    cached = cache_dir / filename
+    _touch(cached, b"cached")
+
+    # Pretend package also has it; function should still return cache path
+    monkeypatch.setattr(prop.resources, "is_resource",
+                        lambda mod, name: True)
+    monkeypatch.setattr(prop.resources, "path",
+                        lambda mod, name: _PathCtx(tmp_path / "other" / name))
+
+    out = prop.download_file_if(filename, data_home=str(cache_dir),
+                                download_if_missing=False, verbose=False)
+    assert Path(out) == cached
+    assert Path(out).read_bytes() == b"cached"
+
+
+# ---------- download_file_if : forced download ----------
+
+def test_force_download_calls_downloader_and_writes_file(
+        tmp_path, monkeypatch):
+    cache_dir = tmp_path / "cache"
+    filename = "force.txt"
+
+    # No package resource
[…]
