### 2025-08-12 083d27e
safe fixes
