+++ tests/test_q_utils.py
@@ -6,6 +6,7 @@ import pytest
 
 from kdiagram.utils import q_utils
 
+
 # ---------- Fixtures ----------
 @pytest.fixture
 def wide_df():
