+++ kdiagram/datasets/_property.py
@@ -71,6 +71,7 @@ __all__ = [
     "remove_data",
 ]
 
+
 # --- Function Definitions ---
 def get_data(data_home: str | None = None) -> str:
     r"""Get the path to the k-diagram data cache directory.
