### 2025-08-12 25c0151
Update cli_usage.rst
