+++ docs/source/cli_usage.rst
@@ -112,20 +112,20 @@ The basic structure of a `k-diagram` CLI command is:
    k-diagram <plot_command> [input_files...] [options...]
 
 * **`<plot_command>`:** The name of the plot you want to generate
-    (e.g., `plot_anomaly_magnitude`, `plot_coverage`).
+  (e.g., `plot_anomaly_magnitude`, `plot_coverage`).
 * **`[input_files...]`:** One or more positional arguments specifying
-    the path(s) to your input CSV data file(s). The number and meaning
-    of these files depend on the specific command (check its `--help`).
+  the path(s) to your input CSV data file(s). The number and meaning
+  of these files depend on the specific command (check its `--help`).
 * **`[options...]`:** Flags and arguments starting with `--` or `-`
-    used to specify necessary information (like column names) or customize
-    the plot appearance. Common options include:
-    * `--actual_col`, `--q_cols`, `--q10_cols`, etc.: Specify which
-        columns in your CSV contain the relevant data. **These are often
-        required.**
-    * `--title`: Set a custom plot title.
-    * `--savefig`: Save the output plot to a file instead of displaying it.
-    * Options specific to the plot type (e.g., `--acov`, `--cmap`,
-        `--normalize`).
+  used to specify necessary information (like column names) or customize
+  the plot appearance. Common options include:
+
+  * `--actual_col`, `--q_cols`, `--q10_cols`, etc.: Specify which
+    columns in your CSV contain the relevant data. **These are often
+    required.**
+  * `--title`: Set a custom plot title.
+  * `--savefig`: Save the output plot to a file instead of displaying it.
+  * Options specific to the plot type (e.g., `--acov`, `--cmap`, `--normalize`).
 
 Command Examples
 ------------------
@@ -205,8 +205,8 @@ prediction models against true values.
 
 * `true_values.csv`: File with actual values (first positional argument).
 * `model_A_preds.csv model_B_preds.csv`: Files with predictions for
-    each model (subsequent positional arguments for `y_preds_files`).
-    *Note: Assumes these CSVs contain just the lower and upper bounds needed.*
+  each model (subsequent positional arguments for `y_preds_files`).
+  *Note: Assumes these CSVs contain just the lower and upper bounds needed.*
 * `--q 0.05 0.95`: Specifies the quantile levels used for the interval (90%).
 * `--names ...`: Provides labels for the models in the legend.
 * `--kind radar`: Selects the radar chart type.
@@ -258,9 +258,9 @@ changes across forecast horizons defined in `multi_horizon_preds.csv`.
 
 * `multi_horizon_preds.csv`: The input data file.
 * `--q10_cols ...`: Lists the columns containing the lower quantile
-    for each horizon.
+  for each horizon.
 * `--q90_cols ...`: Lists the columns containing the upper quantile
-    for each horizon.
+  for each horizon.
 * `--horizons ...`: Provides labels for each horizon shown on the plot.
 * `--acov quarter_circle`: Restricts the plot to a 90-degree arc.
 * `--savefig ...`: Saves the plot as an SVG file.
@@ -285,7 +285,7 @@ Colors points by velocity and saves as SVG.
 * `median_predictions.csv`: Positional argument for input CSV.
 * `--q50-cols ...`: Specifies the Q50 columns for consecutive periods.
 * `--use-velocity-color`: Colors points by the calculated velocity
-    value (positive/negative change) instead of absolute Q50 magnitude.
+  value (positive/negative change) instead of absolute Q50 magnitude.
 * `--cmap coolwarm`: Uses a diverging colormap suitable for velocity.
 * `--title "..."`: Custom plot title.
 * `--cbar`: Adds a color bar (representing velocity).
[…]
