### 2025-08-15 earthai-tech/k-diagram#22 @cbueth
References
The documentation pages are well structured and informative. As there are quite some mathematical and theoretical concepts it would be good to reference these. Specifically for each of the uncertainty diagnostics, the reader can be curious if they are developed by you for this package or if they have been developed in earlier publications, either by you or other scientist. I suggest adding them in the paper as well as these and other references inside the documentation, you can add the [sphinxcontrib-bibtex](https://sphinxcontrib-bibtex.readthedocs.io/en/latest/usage.html) package for easy referencing, inside the documentation, as well as the docstrings inside the code.

In the abstract of the paper you cite a reference. This might be possible if you really want, but is uncommon practice.

https://github.com/openjournals/joss-reviews/issues/8661
