### 2025-08-16 c1f2b8f
Add radial ring plot
