+++ kdiagram/utils/hist.py
@@ -1,0 +1,246 @@
+import matplotlib.pyplot as plt
+from typing import Optional, Tuple, Union
+import numpy as np
+import pandas as pd
+
+from .handlers import columns_manager
+from .plot import (
+    prepare_data_for_kde,
+    setup_plot_axes,
+    add_kde_to_plot,
+    add_histogram_to_plot,
+    normalize_pdf,
+    set_axis_grid,
+)
+
+def plot_hist_kde(
+    data: Union[np.ndarray, pd.Series, pd.DataFrame],
+    column: Optional[str] = None,
+    *,
+    bins: int = 50,
+    x_label: Optional[str] = None, 
+    title: str = "Distribution (Histogram + KDE)",
+    bandwidth: Optional[float] = None,
+    show_kde: bool = True,
+    savefig: Optional[str] = None,
+    dpi: int = 300,
+    figsize: Tuple[float, float] = (8, 6),
+    kde_color: str = "orange",
+    hist_color: str = "skyblue",
+    hist_edge_color: str = "white",
+    kde_line_width: float = 2,
+    hist_alpha: float = 0.7,
+    normalize_kde: bool = False,
+    show_grid: bool = True,
+    grid_props: Optional[dict] = None,
+    **hist_kws, # Pass extra keywords to the histogram
+) -> Tuple[np.ndarray, np.ndarray]:
+    # Ensure the data is a valid type and convert it to np.ndarray
+    if isinstance(data, ( pd.DataFrame, pd.Series)):
+        if isinstance(data, pd.DataFrame):
+            columns = columns_manager(column, empty_as_none=True)
+            if columns is None:
+                raise ValueError(
+                    "If a DataFrame is provided, the 'column' "
+                    "parameter must specify the series to plot."
+                )
+            column = columns[0]
+            series_data = data[column]
+        else: 
+            series_data = data
+
+        # Auto-set x_label from series name if not provided by user
+        if x_label is None and series_data.name:
+            x_label = str(series_data.name)
+            
+        data = series_data.values
+      
+    # Default x_label if still not set
+    if x_label is None:
+        x_label = "Value"
+        
+    data = np.asarray(data)
+
+    # Prepare the data for KDE
+    grid, pdf = prepare_data_for_kde(data, bandwidth=bandwidth)
+
+    if normalize_kde:
+        pdf = normalize_pdf(pdf)
+
+    # Create the plot axes
+    ax = setup_plot_axes(figsize=figsize, title=title, x_label= x_label)
+
+    # Set grid properties
+    set_axis_grid(ax, show_grid=show_grid, grid_props=grid_props)
+
+    add_histogram_to_plot(
+        data,
+        ax,
+        bins=bins,
+        hist_color=hist_color,
+        hist_edge_color=hist_edge_color,
+        hist_alpha=hist_alpha,
+        **hist_kws
+    )
+
+    # Add KDE to the plot if requested
+    if show_kde:
+        add_kde_to_plot(
+            grid,
+            pdf,
+            ax,
+            color=kde_color,
+            line_width=kde_line_width,
+        )
+
+    # Customize axis labels and title
+    ax.set_xlabel(x_label, fontsize=12)
+    ax.set_ylabel("Density", fontsize=12)
+    
+    # Let legend be added after all elements are plotted
+    ax.legend()
+
+    # Save or display the plot
+    if savefig:
+        plt.savefig(savefig, dpi=dpi, bbox_inches="tight")
+        plt.close()
+    else:
+        plt.show()
+
+    return grid, pdf
+
+
+plot_hist_kde.__doc__=r"""
+Plot histogram and Kernel Density Estimate (KDE) for uncertainty 
+evaluation.
+
[…]
