+++ kdiagram/utils/__init__.py
@@ -8,10 +8,13 @@ from .q_utils import (
     reshape_quantile_data,
 )
 
+from .hist import plot_hist_kde 
+
 __all__ = [
     "reshape_quantile_data",
     "melt_q_data",
     "pivot_q_data",
     "detect_quantiles_in",
     "build_q_column_names",
+    "plot_hist_kde"
 ]
