+++ kdiagram/utils/plot.py
@@ -2,10 +2,16 @@
 #   Author: LKouadio <etanoyau@gmail.com>
 
 import re
-from typing import Optional
+from typing import Optional, Tuple
+from scipy.stats import gaussian_kde
+import numpy as np
+import matplotlib.pyplot as plt
 
+from .generic_utils import get_valid_kwargs 
 
-def set_axis_grid(ax, show_grid: bool = True, grid_props: dict = None) -> None:
+def set_axis_grid(
+        ax, show_grid: bool = True, grid_props: dict = None
+        ) -> None:
     """Robustly set grid properties on one or more matplotlib axes."""
     # Ensure grid_props is a dictionary.
     grid_props = (
@@ -149,3 +155,335 @@ def is_valid_kind(
             raise ValueError(f"Invalid plot type '{kind}'. Allowed: {allowed}")
 
     return final_kind
+
+
+def prepare_data_for_kde(
+    data: np.ndarray,
+    bandwidth: Optional[float] = None
+) -> Tuple[np.ndarray, np.ndarray]:
+    r"""
+    Prepares the data for Kernel Density Estimate (KDE) calculation.
+    
+    This function cleans the data by removing non-finite values, then creates
+    a grid for KDE evaluation. It then computes the Kernel Density Estimate 
+    using a Gaussian kernel.
+
+    Parameters
+    ----------
+    data : np.ndarray
+        The data to calculate the KDE from.
+    bandwidth : float, optional
+        Bandwidth for the KDE. Default is None, which uses Silverman's rule 
+        of thumb for bandwidth estimation.
+
+    Returns
+    -------
+    grid : np.ndarray
+        The x-values grid for KDE evaluation.
+    pdf : np.ndarray
+        The estimated probability density function (PDF) from the KDE.
+
+    Notes
+    -----
+    The KDE is computed by finding the optimal bandwidth (if not provided)
+    and applying a Gaussian kernel to the data. The bandwidth is chosen 
+    using Silverman's rule of thumb when not explicitly provided.
+
+    Examples
+    --------
+    >>> data = np.random.normal(0, 1, 1000)
+    >>> grid, pdf = prepare_data_for_kde(data)
+    >>> print(grid, pdf)
+
+    See Also
+    --------
+    _kde_pdf : Helper function that calculates the KDE.
+
+    References
+    ----------
+    [1] Silverman, B. W. (1986). Density Estimation for Statistics and Data 
+        Analysis. CRC Press.
+    """
+    data = np.asarray(data, dtype=float)
+    data = data[np.isfinite(data)]
+    
+    if data.size == 0:
+        raise ValueError("`data` is empty after removing non-finite values.")
+    
+    lo, hi = np.min(data), np.max(data)
+    grid = np.linspace(lo, hi, 512)
+    
+    pdf = _kde_pdf(data, grid, bandwidth=bandwidth)
+    
+    return grid, pdf
+
+
+def _kde_pdf(
+    x: np.ndarray,
+    grid: np.ndarray,
+    bandwidth: Optional[float] = None
+) -> np.ndarray:
+    r"""
+    Helper function to calculate the Kernel Density Estimate (KDE).
+
+    This function computes the KDE for a given dataset and returns the 
+    probability density function (PDF) evaluated at the grid points.
+    
+    Parameters
+    ----------
+    x : np.ndarray
+        The data values to estimate the KDE from.
+    grid : np.ndarray
+        The grid points at which to evaluate the KDE.
+    bandwidth : float, optional
+        Bandwidth for the KDE. Default is None, which uses Silverman's rule 
+        of thumb.
+
+    Returns
+    -------
+    np.ndarray
[…]
