### 2025-08-16 9f6ceab
Add errors module
