+++ kdiagram/plot/uncertainty.py
@@ -56,7 +56,9 @@ __all__ = [
     "plot_temporal_uncertainty",
     "plot_uncertainty_drift",
     "plot_velocity",
-    "plot_radial_density_ring"
+    "plot_radial_density_ring", 
+    "plot_polar_heatmap", 
+    "plot_polar_quiver"
 ]
 
 
@@ -5164,6 +5166,358 @@ Examples
 ... )
 """
 
+
+@check_non_emptiness
+@isdf
+def plot_polar_quiver(
+    df: pd.DataFrame,
+    r_col: str,
+    theta_col: str,
+    u_col: str,
+    v_col: str,
+    *,
+    color_col: Optional[str] = None,
+    theta_period: Optional[float] = None,
+    title: Optional[str] = None,
+    figsize: Tuple[float, float] = (8, 8),
+    cmap: str = "viridis",
+    mask_angle: bool = False,
+    mask_radius: bool = False,
+    show_grid: bool = True,
+    grid_props: Optional[Dict[str, Any]] = None,
+    savefig: Optional[str] = None,
+    dpi: int = 300,
+    **quiver_kws,
+):
+    
+    required_cols = [r_col, theta_col, u_col, v_col]
+    if color_col:
+        required_cols.append(color_col)
+    exist_features(df, features=required_cols)
+    
+    data = df[required_cols].dropna()
+    if data.empty:
+        warnings.warn(
+            "DataFrame is empty after dropping NaNs in "
+            "required columns. Cannot plot.",
+            UserWarning,
+            stacklevel=2,
+        )
+        return None
+
+    r_data = data[r_col].to_numpy()
+    theta_data = data[theta_col].to_numpy()
+    u_data = data[u_col].to_numpy()  # Radial component
+    v_data = data[v_col].to_numpy()  # Angular component
+
+    if theta_period:
+        theta_rad = (theta_data % theta_period) / theta_period * (
+            2 * np.pi
+        )
+    else:
+        min_theta, max_theta = theta_data.min(), theta_data.max()
+        if (max_theta - min_theta) > 1e-9:
+            theta_rad = (
+                (theta_data - min_theta) / (max_theta - min_theta)
+            ) * 2 * np.pi
+        else:
+            theta_rad = np.zeros_like(theta_data)
+
+    if color_col:
+        color_data = data[color_col].to_numpy()
+        cbar_label = color_col
+    else:
+        # Default to coloring by vector magnitude
+        color_data = np.sqrt(u_data**2 + v_data**2)
+        cbar_label = "Vector Magnitude"
+
+    norm = Normalize(
+        vmin=np.min(color_data), vmax=np.max(color_data)
+    )
+    cmap_obj = get_cmap(cmap, default="viridis")
+    colors = cmap_obj(norm(color_data))
+
+    fig, ax = plt.subplots(
+        figsize=figsize, subplot_kw={"projection": "polar"}
+    )
+    
+    # Matplotlib's quiver on a polar axis expects:
+    # (theta, r, angular_change, radial_change)
+    ax.quiver(
+        theta_rad, r_data, v_data, u_data,
+        color=colors, **quiver_kws
+    )
+
+    cbar = plt.colorbar(
+        plt.cm.ScalarMappable(norm=norm, cmap=cmap_obj),
+        ax=ax, pad=0.1, shrink=0.75
+    )
+    cbar.set_label(cbar_label, fontsize=10)
+
+    ax.set_title(
+        title or f"Quiver Plot of ({u_col}, {v_col})"
+    )
+    set_axis_grid(ax, show_grid=show_grid, grid_props=grid_props)
+    if mask_angle:
+        ax.set_xticklabels([])  
+
[…]
