+++ kdiagram/plot/errors.py
@@ -1,0 +1,752 @@
+# License: Apache 2.0
+# Author: LKouadio <etanoyau@gmail.com>
+
+import warnings 
+import numpy as np
+import pandas as pd
+import matplotlib.pyplot as plt
+from matplotlib.colors import Normalize
+from typing import Optional, Tuple, Dict, Any, List
+from scipy.stats import gaussian_kde
+
+from ..decorators import check_non_emptiness, isdf
+from ..utils.validator import exist_features
+from ..compat.matplotlib import get_cmap
+from ..utils.plot import set_axis_grid
+
+__all__ = [
+    "plot_error_ellipses","plot_error_bands" , "plot_error_violins"
+  ]
+
+@check_non_emptiness
+@isdf
+def plot_error_violins(
+    df: pd.DataFrame,
+    *error_cols: str,
+    names: Optional[List[str]] = None,
+    title: Optional[str] = None,
+    figsize: Tuple[float, float] = (9, 9),
+    cmap: str = "viridis",
+    show_grid: bool = True,
+    grid_props: Optional[Dict[str, Any]] = None,
+    savefig: Optional[str] = None,
+    dpi: int = 300,
+    **violin_kws,
+):
+    if not error_cols:
+        raise ValueError("At least one error column must be provided.")
+    exist_features(df, features=list(error_cols))
+
+    if names and len(names) != len(error_cols):
+        warnings.warn(
+            f"Number of names ({len(names)}) does not match number of "
+            f"error columns ({len(error_cols)}). Using default names.",
+            UserWarning,
+            stacklevel=2,
+        )
+        names = None
+    if not names:
+        names = [f"Model {i+1}" for i in range(len(error_cols))]
+
+    # Prepare data and KDEs for each model
+    violin_data = []
+    all_errors = np.concatenate(
+        [df[col].dropna().to_numpy() for col in error_cols]
+    )
+    r_min, r_max = all_errors.min(), all_errors.max()
+    grid = np.linspace(r_min, r_max, 200)
+
+    for col in error_cols:
+        errors = df[col].dropna().to_numpy()
+        if len(errors) < 2:
+            violin_data.append(None) # Cannot compute KDE
+            continue
+        
+        kde = gaussian_kde(errors)
+        density = kde(grid)
+        violin_data.append(density / density.max()) # Normalize density
+
+    # Plot setup
+    fig, ax = plt.subplots(
+        figsize=figsize, subplot_kw={"projection": "polar"}
+    )
+    num_violins = len(error_cols)
+    angles = np.linspace(0, 2 * np.pi, num_violins, endpoint=False)
+    cmap_obj = get_cmap(cmap, default="viridis")
+    colors = cmap_obj(np.linspace(0, 1, num_violins))
+    
+    # Draw violins
+    for i, (angle, density) in enumerate(zip(angles, violin_data)):
+        if density is None:
+            continue
+            
+        # Width of the violin slice
+        width = (2 * np.pi / num_violins) * 0.8
+        
+        # Create the path for the violin polygon
+        x = np.concatenate([
+            -density * width / 2,
+            np.flip(density * width / 2)
+        ])
+        y = np.concatenate([grid, np.flip(grid)])
+        
+        # Rotate and translate path to the correct angle
+        theta = x + angle
+        r = y
+        
+        ax.fill(theta, r, color=colors[i], label=names[i],
+                alpha=violin_kws.pop('alpha', 0.6), **violin_kws)
+
+    # Add zero-error reference line
+    ax.plot(np.linspace(0, 2 * np.pi, 100), np.zeros(100),
[…]
