+++ kdiagram/__init__.py
@@ -2,10 +2,11 @@
 # Author: L. Kouadio <etanoyau@gmail.com>
 
 """
-K-Diagram: Rethinking Forecasting Uncertainty via Polar-Based Visualization
-============================================================================
-`k-diagram` is a Python package designed to provide specialized diagnostic polar plots,
-called "k-diagrams," for comprehensive model evaluation and forecast analysis.
+K-Diagram: Polar Diagnostics for Forecast Uncertainty 
+=======================================================
+`k-diagram` is a Python package designed to provide 
+specialized diagnostic polar plots, called "k-diagrams, 
+for comprehensive model evaluation and forecast analysis.
 """
 import importlib
 import logging
@@ -27,7 +28,6 @@ def _lazy_import(module_name, alias=None):
         globals()[alias] = _lazy_loader
     else:
         globals()[module_name] = _lazy_loader
-
 
 # Version (single source)
 try:
@@ -71,7 +71,6 @@ if _missing_dependencies:
 # Re-export config helpers
 from .config import configure_warnings, warnings_config  # noqa: F401, E402
 
-# from . import datasets
 from .plot import (  # noqa: E402
     plot_actual_vs_predicted,
     plot_anomaly_magnitude,
@@ -91,7 +90,14 @@ from .plot import (  # noqa: E402
     taylor_diagram,
     plot_radial_density_ring, 
     plot_reliability_diagram, 
-    plot_horizon_metrics
+    plot_horizon_metrics, 
+    plot_polar_heatmap, 
+    plot_polar_quiver, 
+    plot_error_bands, 
+    plot_error_ellipses, 
+    plot_error_violins, 
+    plot_radial_density_ring, 
+    
 )
 
 __all__ = [
@@ -116,5 +122,11 @@ __all__ = [
     "plot_model_comparison",
     "plot_radial_density_ring", 
     "plot_reliability_diagram", 
-    "plot_horizon_metrics"
+    "plot_horizon_metrics", 
+    "plot_horizon_metrics",  
+    "plot_polar_heatmap", 
+    "plot_polar_quiver", 
+    "plot_error_bands", 
+    "plot_error_ellipses", 
+    "plot_error_violins"
 ]
