+++ paper.md
@@ -30,9 +30,9 @@ bibliography: paper.bib
 interpretation and diagnosis of probabilistic forecasts. Moving beyond 
 traditional aggregate metrics, the package provides a suite of novel 
 visualization tools, primarily based on polar coordinate systems, 
-to dissect the complex spatiotemporal structure of forecast uncertainty. 
-By mapping statistical properties such as sharpness, reliability, and 
-temporal stability onto intuitive geometric representations, `k-diagram` 
+to dissect the complex spatiotemporal structure of forecast uncertainty and 
+errors. By mapping statistical properties such as sharpness, reliability, 
+and temporal stability onto intuitive geometric representations, `k-diagram` 
 enables researchers and practitioners to identify model biases, understand 
 forecast degradation over time, and communicate uncertainty characteristics 
 more effectively. The package is designed to be a practical extension 
@@ -81,42 +81,71 @@ and gain a deeper, more actionable understanding of their forecasting models.
 # Functionality
 
 `k-diagram` is implemented in Python [@python3; @pythonbook], leveraging core scientific libraries
-including `numpy` [@harris2020array], `pandas` [@mckinney-proc-scipy-2010;
-@reback2020pandas], `matplotlib` [@Hunter:2007], `scipy` [@2020SciPy-NMeth],
+including `numpy` [@harris2020array], `pandas` [@mckinney-proc-scipy-2010; @reback2020pandas], 
+`matplotlib` [@Hunter:2007], `scipy` [@2020SciPy-NMeth],
 and `scikit-learn` [@scikit-learn]. The core functionality is organized around diagnosing 
 key aspects of forecast quality \autoref{fig1:workflow}:
 
-* **Uncertainty Diagnostics:** A collection of polar plots under
-    `kdiagram.plot.uncertainty` for visualizing:
-    * Prediction interval coverage (point-wise diagnostics and overall scores).
-    * Anomaly magnitude (severity and type of interval failures).
-    * Interval width (magnitude and consistency over time/samples).
-    * Model drift (average uncertainty increase over forecast horizons).
-    * Uncertainty drift (evolution of uncertainty patterns using concentric rings).
-    * Prediction velocity (rate of change of central forecasts).
-* **Model Evaluation:** Functions under `kdiagram.plot.evaluation` for
-    generating standard and enhanced Taylor Diagrams [@Taylor2001] to compare
-    models based on correlation, standard deviation, and RMSD relative to a
-    reference.
-* **Feature Importance:** Radar charts under `kdiagram.plot.feature_based`
-    for visualizing and comparing feature importance profiles ("fingerprints")
-    across different models or contexts.
-* **Relationship Visualization:** Polar scatter plots under
-    `kdiagram.plot.relationship` mapping true values to angle and
-    (normalized) predictions to radius.
-* **Data Utilities:** Helper functions under `kdiagram.utils` for detecting,
-    validating, and reshaping pandas DataFrames containing quantile-based
-    forecast data, facilitating data preparation for the plotting functions.
-* **Command-Line Interface (CLI):** A `k-diagram` command allowing users
-    to generate key plots directly from CSV files without writing Python scripts.
-* **Customization:** Plots offer various parameters for adjusting appearance,
-    angular coverage (`acov`), colormaps, normalization, and labeling.
+# Functionality
 
+`k-diagram` is implemented in Python [@python3; @pythonbook], leveraging
+core scientific libraries including `numpy` [@harris2020array],
+`pandas` [@mckinney-proc-scipy-2010; @reback2020pandas],
+`matplotlib` [@Hunter:2007], `scipy` [@2020SciPy-NMeth], and
+`scikit-learn` [@scikit-learn]. The core functionality is organized
+around a cohesive API for diagnosing key aspects of forecast quality
+(see \autoref{fig1:workflow}):
+
[…]
