+++ kdiagram/plot/uncertainty.py
@@ -312,7 +312,7 @@ This plot provides a high-level summary of whether prediction
 intervals are well-calibrated on average across the entire
 dataset. It's useful for quickly comparing the overall
 reliability of uncertainty estimates from different models or
-parameter settings before diving into finer diagnostics.
+parameter settings before diving into finer diagnostics [1]_.
 
 Parameters
 ----------
@@ -471,7 +471,7 @@ intervals [1]_.
 **Interpretation:** Compare the calculated coverage score(s) to
 the nominal coverage rate implied by the quantiles (e.g., 80%
 for a Q10-Q90 interval). Scores far from the nominal rate suggest
-miscalibration (intervals too wide or too narrow on average).
+miscalibration (intervals too wide or too narrow on average) [2]_.
 
 **Radar Plot Fill (`cov_fill=True`):**
 
@@ -483,7 +483,12 @@ miscalibration (intervals too wide or too narrow on average).
 
 References
 ----------
-.. [1] Gneiting, T., & Raftery, A. E. (2007). Strictly Proper
+
+.. [1] Kouadio, K. L., Liu, R., Loukou, K. G. H., Liu, J., & Liu, W. (2025).
+   Analytics Framework for Interpreting Spatiotemporal Probabilistic Forecasts.
+   *International Journal of Forecasting*. Manuscript submitted.
+   
+.. [2] Gneiting, T., & Raftery, A. E. (2007). Strictly Proper
        Scoring Rules, Prediction, and Estimation. *Journal of the
        American Statistical Association*, 102(477), 359–378.
        *(General reference on probabilistic forecast evaluation)*
@@ -672,11 +677,16 @@ def plot_model_drift(
 
     Radii may be scaled relative to the maximum radius if a
     restricted angular coverage (``acov`` is not 'default') is used,
-    to better fit the visual sector.
+    to better fit the visual sector [2]_.
 
     References
     ----------
-    .. [1] Gama, J., Žliobaitė, I., Bifet, A., Pechenizkiy, M., &
+
+    .. [1] Kouadio, K. L., Liu, R., Loukou, K. G. H., Liu, J., & Liu, W. (2025).
+       Analytics Framework for Interpreting Spatiotemporal Probabilistic 
+       Forecasts. *International Journal of Forecasting*. Manuscript submitted.
+       
+    .. [2] Gama, J., Žliobaitė, I., Bifet, A., Pechenizkiy, M., &
            Bouchachia, A. (2014). A survey on concept drift
            adaptation. *ACM Computing Surveys (CSUR)*, 46(4), 1-37.
 
@@ -972,9 +982,7 @@ def plot_velocity(
       DataFrame index, mapped linearly onto the angular range defined
       by `acov`. The `theta_col` parameter is not used for positioning
       in the current implementation, which might be revised in future
-      versions. A warning is issued if `theta_col` is provided.
-    - Input validation relies on decorators `@isdf` and
-      `@check_non_emptiness` from the `gofast` library.
+      versions. A warning is issued if `theta_col` is provided [2]_.
 
     Let :math:`\mathbf{Q}` be the data matrix extracted from `df` using
     columns `q50_cols`, with shape :math:`(N, M)`, where :math:`N` is
@@ -1033,7 +1041,12 @@ def plot_velocity(
 
     References
     ----------
-    .. [1] Hunter, J. D. (2007). Matplotlib: A 2D graphics environment.
+
+    .. [1] Kouadio, K. L., Liu, R., Loukou, K. G. H., Liu, J., & Liu, W. (2025).
+       Analytics Framework for Interpreting Spatiotemporal Probabilistic Forecasts.
+       *International Journal of Forecasting*. Manuscript submitted.
+       
+    .. [2] Hunter, J. D. (2007). Matplotlib: A 2D graphics environment.
[…]
