### 2025-08-17 6e35f5b
Joss review: Updated References and Paper
