+++ paper.bib
@@ -296,3 +296,42 @@
   publisher = {Chapman and Hall/CRC},
   isbn      = {9780412246203}
 }
+
+@article{Biswas2025,
+    doi = {10.21105/joss.08336},
+    url = {https://doi.org/10.21105/joss.08336},
+    year = {2025},
+    publisher = {The Open Journal},
+    volume = {10},
+    number = {111},
+    pages = {8336},
+    author = {Biswas, Anubhab and Nespoli, Lorenzo and Azzimonti, Dario and Zambon, Lorenzo and Rubattu, Nicolò and Corani, Giorgio},
+    title = {BayesReconPy: A Python package for forecast reconciliation},
+    journal = {Journal of Open Source Software}
+}
+
+@article{Brady2021,
+    doi = {10.21105/joss.02781},
+    url = {https://doi.org/10.21105/joss.02781},
+    year = {2021},
+    publisher = {The Open Journal},
+    volume = {6},
+    number = {59},
+    pages = {2781},
+    author = {Brady, Riley X. and Spring, Aaron},
+    title = {climpred: Verification of weather and climate forecasts},
+    journal = {Journal of Open Source Software}
+}
+
+@article{Caulk2022,
+    doi = {10.21105/joss.04864},
+    url = {https://doi.org/10.21105/joss.04864},
+    year = {2022},
+    publisher = {The Open Journal},
+    volume = {7},
+    number = {80},
+    pages = {4864},
+    author = {Caulk, Robert A. and Törnquist, Elin and Voppichler, Matthias and Lawless, Andrew R. and McMullan, Ryan and Santos, Wagner Costa and Pogue, Timothy C. and van der Vlugt, Johan and Gehring, Stefan P. and Schmidt, Pascal},
+    title = {FreqAI: generalizing adaptive modeling for chaotic time-series market forecasts},
+    journal = {Journal of Open Source Software}
+}
