+++ paper.bib
@@ -3,10 +3,9 @@
 @misc{python3,
   author = {{Python Software Foundation}},
   title  = {Python Language Reference, version 3.x},
-  url    = {https://www.python.org},
-  year = {2008},
-  note   = {Accessed: 2025-04-10}, 
-  url = {https://dl.acm.org/doi/10.5555/1593511}
+  year   = {2008},
+  note   = {Accessed: 2025-04-10},
+  url    = {https://www.python.org}
 }
 
 @book{pythonbook,
@@ -79,20 +78,7 @@
     year = {2024}
 }
 
-@article{harris2020array,
-  title={Array programming with NumPy},
-  author={Charles R. Harris and K. Jarrod Millman and St\'efan J. van der Walt and Ralf Gommers and Pauli Virtanen and David Cournapeau and Eric Wieser and Julian Taylor and Stephan Memmesheimer and Tom Backx and others},
-  year=2020,
-  month=feb,
-  publisher={Nature Publishing Group},
-  volume={585},
-  number={7825},
-  pages={357--362},
-  doi = {10.1038/s41586-020-2649-2},
-  journal={Nature}
-}
-
-@article{mckinney-proc-scipy-2010,
+@inproceedings{mckinney-proc-scipy-2010,
   author    = {McKinney, Wes},
   title     = {{Data Structures for Statistical Computing in Python}},
   booktitle = {Proceedings of the 9th Python in Science Conference},
@@ -100,14 +86,7 @@
   pages     = {56 - 61},
   year      = {2010},
   publisher = {SciPy}, 
-  doi={10.25080/majora-92bf1922-00a}, 
-}
-
-@article{reback2020pandas,
-  title={pandas 1.0, a fundamental rewrite of the library},
-  author={Jeff Reback and Joris Van den Bossche and Wes McKinney and Terry J. Reedy and Garrett Bergstrom and Tom Augspurger and Martin Brooke and Simone Ciampi and Antony Lee and Irv Lustig and others},
-  journal={arXiv preprint arXiv:2005.03268},
-  year={2020}
+  doi       = {10.25080/majora-92bf1922-00a}, 
 }
 
 @article{2020SciPy-NMeth,
@@ -132,33 +111,12 @@
   doi       = {10.1029/2000JD900719}
 }
 
-@article{Hunter:2007,
-  author    = {Hunter, J. D.},
-  title     = {Matplotlib: A 2D graphics environment},
-  journal   = {Computing in Science \& Engineering},
-  volume    = {9},
-  number    = {3},
-  pages     = {90-95},
-  year      = {2007},
-  doi       = {10.1109/MCSE.2007.55}
-}
-
-@article{Waskom2021,
-  author    = {Waskom, Michael L.},
-  title     = {Seaborn: statistical data visualization},
-  journal   = {Journal of Open Source Software},
-  volume    = {6},
-  number    = {60},
-  pages     = {3021},
-  year      = {2021},
-  doi       = {10.21105/joss.03021}
-}
-
 @article{harris2020array,
   title={Array programming with {NumPy}},
   author={Charles R. Harris and K. Jarrod Millman and St{\'e}fan J. van der Walt and Ralf Gommers and Pauli Virtanen and David Cournapeau and Eric Wieser and Julian Taylor and Stephan Memmesheimer and Tom Backx and others},
   year={2020},
   journal={Nature},
+  publisher={Nature Publishing Group},
   volume={585},
   number={7825},
   pages={357--362},
