+++ kdiagram/datasets/__init__.py
@@ -6,7 +6,6 @@ Datasets submodule for k-diagram, including data generation tools
 and loading APIs.
 """
 from .load import (
-    # load_synthetic_uncertainty_data, #  renamed to load_uncertainty_data
     load_uncertainty_data,
     load_zhongshan_subsidence,
 )
