### 2025-08-18 73093dc
Fist step of duplicate citations fixed
