+++ kdiagram/datasets/load.py
@@ -459,7 +459,6 @@ def load_zhongshan_subsidence(
 
         return Bunch(**bunch_dict)
 
-
 def load_uncertainty_data(
     *,
     as_frame: bool = False,
