+++ kdiagram/utils/__init__.py
@@ -2,13 +2,12 @@ from .diagnose_q import (
     build_q_column_names,
     detect_quantiles_in,
 )
+from .hist import plot_hist_kde
 from .q_utils import (
     melt_q_data,
     pivot_q_data,
     reshape_quantile_data,
 )
-
-from .hist import plot_hist_kde 
 
 __all__ = [
     "reshape_quantile_data",
@@ -16,5 +15,5 @@ __all__ = [
     "pivot_q_data",
     "detect_quantiles_in",
     "build_q_column_names",
-    "plot_hist_kde"
+    "plot_hist_kde",
 ]
