### 2025-08-18 aea811e
Fix sphinx warnings: Duplicated citations
