+++ kdiagram/plot/relationship.py
@@ -47,165 +47,6 @@ def plot_relationship(
     z_label=None,
     savefig=None,
 ):
-    r"""
-    Visualize the relationship between `y_true` and multiple `y_preds`
-    using a circular or polar plot. The function allows flexible
-    configurations such as angular coverage, z-values for replacing
-    angle labels, and customizable axis labels.
-
-    Parameters
-    -------------
-    y_true : array-like
-        The true values. Must be numeric, one-dimensional, and of the
-        same length as the values in `y_preds`.
-
-    y_preds : array-like (one or more)
-        Predicted values from one or more models. Each `y_pred` must
-        have the same length as `y_true`.
-
-    names : list of str, optional
-        A list of model names corresponding to each `y_pred`. If not
-        provided or if fewer names than predictions are given, the
-        function assigns default names as ``"Model_1"``, ``"Model_2"``,
-        etc. For instance, if `y_preds` has three predictions and
-        `names` is `["SVC", "RF"]`, the names will be updated to
-        `["SVC", "RF", "Model_3"]`.
-
-    title : str, optional
-        The title of the plot. If `None`, the title defaults to
-        `"Relationship Visualization"`.
-
-    theta_offset : float, default=0
-        Angular offset in radians to rotate the plot. This allows
-        customization of the orientation of the plot.
-
-    theta_scale : {'proportional', 'uniform'}, default='proportional'
-        Determines how `y_true` values are mapped to angular
-        coordinates (`theta`):
-
-        - ``'proportional'``: Maps `y_true` proportionally to the
-          angular range (e.g., 0 to 360° or a subset defined by
-          `acov`).
-        - ``'uniform'``: Distributes `y_true` values uniformly around
-          the angular range.
-
-    acov : {'default', 'half_circle', 'quarter_circle', 'eighth_circle'},
-        default='default'
-
-        Specifies the angular coverage of the plot:
-
-        - ``'default'``: Full circle (360°).
-        - ``'half_circle'``: Half circle (180°).
-        - ``'quarter_circle'``: Quarter circle (90°).
-        - ``'eighth_circle'``: Eighth circle (45°).
-        The angular span is automatically restricted to the selected
-        portion of the circle.
-
-    figsize : tuple of float, default=(8, 8)
-        The dimensions of the figure in inches.
-
-    cmap : str, default='viridis'
-        Colormap for the scatter points. Refer to Matplotlib
-        documentation for a list of supported colormaps.
-
-    s : float, default=50
-        Size of scatter points representing predictions.
-
-    alpha : float, default=0.7
-        Transparency level for scatter points. Valid values range
-        from 0 (completely transparent) to 1 (fully opaque).
-
-    legend : bool, default=True
-        Whether to display a legend indicating the model names.
-
-    show_grid : bool, default=True
-        Whether to display a grid on the polar plot.
-
-    color_palette : list of str, optional
-        A list of colors to use for the scatter points. If not
-        provided, the default Matplotlib color palette (`tab10`) is
-        used.
-
-    xlabel : str, optional
-        Label for the radial axis (distance from the center). Defaults
-        to `"Normalized Predictions (r)"`.
-
-    ylabel : str, optional
[…]
