+++ kdiagram/plot/uncertainty.py
@@ -301,7 +301,8 @@ This plot provides a high-level summary of whether prediction
 intervals are well-calibrated on average across the entire
 dataset. It's useful for quickly comparing the overall
 reliability of uncertainty estimates from different models or
-parameter settings before diving into finer diagnostics [1]_.
+parameter settings before diving into finer diagnostics 
+:footcite:p:`kouadiob2025`.
 
 Parameters
 ----------
@@ -425,15 +426,14 @@ TypeError
 
 See Also
 --------
-kdiagram.plot.uncertainty.plot_coverage_diagnostic : Plot point-wise
-    coverage status.
+plot_coverage_diagnostic : Plot point-wise coverage status.
 numpy.mean : Used to calculate the average coverage rate.
 
 Notes
 -----
 The empirical coverage rate is a key metric for assessing the
 calibration of probabilistic forecasts, particularly prediction
-intervals [1]_.
+intervals :footcite:p:`kouadiob2025`.
 
 **Calculation:**
 
@@ -447,6 +447,7 @@ intervals [1]_.
      \text{Coverage} = \frac{1}{N}\sum_{i=1}^{N}
      \mathbf{1}\{\hat{y}_{i}^{(\ell)} \leq y_i
      \leq \hat{y}_{i}^{(u)}\}
+     
   where :math:`\mathbf{1}\{\cdot\}` is 1 if true, 0 otherwise.
 
 - For point forecasts (1D `y_preds`), coverage is the proportion
@@ -455,12 +456,14 @@ intervals [1]_.
   .. math::
      \text{Coverage} = \frac{1}{N}\sum_{i=1}^{N}
      \mathbf{1}\{ \hat{y}_i = y_i \}
+     
   This is typically very low for continuous :math:`y_true`.
 
 **Interpretation:** Compare the calculated coverage score(s) to
 the nominal coverage rate implied by the quantiles (e.g., 80%
 for a Q10-Q90 interval). Scores far from the nominal rate suggest
-miscalibration (intervals too wide or too narrow on average) [2]_.
+miscalibration (intervals too wide or too narrow on average)
+:footcite:p:`Gneiting2007b`.
 
 **Radar Plot Fill (`cov_fill=True`):**
 
@@ -469,18 +472,6 @@ miscalibration (intervals too wide or too narrow on average) [2]_.
   reference line at the coverage score.
 - Multiple models: Fills the polygon defined by the coverage
   scores for all models with a semi-transparent color.
-
-References
-----------
-
-.. [1] Kouadio, K. L., Liu, R., Loukou, K. G. H., Liu, J., & Liu, W. (2025).
-   Analytics Framework for Interpreting Spatiotemporal Probabilistic Forecasts.
-   *International Journal of Forecasting*. Manuscript submitted.
-   
-.. [2] Gneiting, T., & Raftery, A. E. (2007). Strictly Proper
-       Scoring Rules, Prediction, and Estimation. *Journal of the
-       American Statistical Association*, 102(477), 359–378.
-       *(General reference on probabilistic forecast evaluation)*
 
 Examples
 --------
@@ -506,6 +497,10 @@ Examples
 ...               title='Coverage Comparison (Radar)',
 ...               cov_fill=True, cmap='Set2')
 
+References
+----------
+.. footbibliography::
+    
 """
 
 
@@ -825,300 +820,6 @@ def plot_velocity(
     cbar: bool = True,
     mask_angle: bool = False,
 ):
-    r"""Polar plot visualizing average velocity across locations.
-
-    Generates a polar scatter plot where each point represents a
-    unique location or observation from the input DataFrame. The
-    radial distance (`r`) of each point corresponds to the average
-    rate of change (velocity) of the median prediction (Q50) over
[…]
