### 2025-08-19 989390d
Ruff check passed and formatted code
