+++ kdiagram/plot/errors.py
@@ -303,10 +303,12 @@ def plot_error_bands(
 
     # Calculate stats per bin
     # stats = data.groupby("theta_bin")[error_col].agg(["mean", "std"]).reset_index()
-    stats = data.groupby('theta_bin', observed=False)[error_col].agg(
-        ['mean', 'std']
-        ).reset_index()
-    
+    stats = (
+        data.groupby("theta_bin", observed=False)[error_col]
+        .agg(["mean", "std"])
+        .reset_index()
+    )
+
     stats["std"] = stats["std"].fillna(0)  # Handle bins with one sample
 
     # Create the plot
@@ -555,8 +557,7 @@ def plot_error_ellipses(
         ax.fill(theta_path, r_path, color=colors[i], **ellipse_kws)
 
     cbar = plt.colorbar(
-        plt.cm.ScalarMappable(norm=norm, cmap=cmap_obj), 
-        ax=ax, pad=0.1, shrink=0.75
+        plt.cm.ScalarMappable(norm=norm, cmap=cmap_obj), ax=ax, pad=0.1, shrink=0.75
     )
     cbar.set_label(cbar_label, fontsize=10)
 
