+++ tests/test_datasets_.py
@@ -1,5 +1,6 @@
 import os
 import re
+
 import numpy as np
 import pandas as pd
 import pytest
@@ -58,16 +59,15 @@ def test_load_zhongshan_subsidence_invalid_years():
     match_pattern = r"Requested years not available: .* Available: .*"
 
     with pytest.warns(UserWarning, match=match_pattern):
-        result = load_zhongshan_subsidence(
-            years=[2030, 2040], as_frame=True
-        )
+        result = load_zhongshan_subsidence(years=[2030, 2040], as_frame=True)
 
     # Assertions remain the same
     assert isinstance(result, pd.DataFrame)
     # The resulting dataframe should be empty of data columns since all
     # requested years were invalid.
-    data_cols = [c for c in result.columns if c.startswith('subsidence')]
+    data_cols = [c for c in result.columns if c.startswith("subsidence")]
     assert len(data_cols) == 0
+
 
 def test_load_zhongshan_subsidence_with_quantiles():
     """Test loading Zhongshan subsidence dataset with specific quantiles."""
@@ -130,7 +130,7 @@ def test_load_uncertainty_data_invalid_parameters():
 
 @pytest.mark.network
 def test_load_zhongshan_subsidence_force_download(
-    mock_load_zhongshan_subsidence, 
+    mock_load_zhongshan_subsidence,
 ):
     """
     Force-download the Zhongshan subsidence dataset
@@ -139,14 +139,13 @@ def test_load_zhongshan_subsidence_force_download(
     try:
         # Use the context manager to capture all warnings that occur
         with pytest.warns(UserWarning) as record:
-            result = load_zhongshan_subsidence(
-                force_download=True, as_frame=True)
-            
+            result = load_zhongshan_subsidence(force_download=True, as_frame=True)
+
             # If the function succeeds, it must have found a fallback resource.
             assert isinstance(result, pd.DataFrame)
 
         # Now, check the captured warnings
-        assert len(record) >= 1 # At least one warning was caught
+        assert len(record) >= 1  # At least one warning was caught
         # Check that the "download failed" message is in one of the warnings
         assert any("Forced download failed" in str(w.message) for w in record)
         # Optionally, check for the copy failure as well
@@ -156,8 +155,8 @@ def test_load_zhongshan_subsidence_force_download(
         # This is the expected final outcome if all fallbacks fail.
         # We skip the test as it confirms the dataset is truly unavailable.
         pytest.skip(f"Dataset unavailable in test env, as expected: {e}")
-        
-    
+
+
 def test_load_uncertainty_data_bunch():
     """Test if the synthetic uncertainty data is returned as Bunch."""
     result = load_uncertainty_data(n_samples=10, as_frame=False)
