### 2025-08-19 8b5e200
Ruff to exclude auto-generated file _scm_version.py
