+++ pyproject.toml
@@ -135,7 +135,7 @@ target-version = ["py39","py310","py311","py312"]
 [tool.ruff]
 line-length = 88
 target-version = "py39"
-exclude = ["docs"]
+exclude = ["docs", "kdiagram/_scm_version.py"]
 
 [tool.ruff.lint]
 select = ["E","F","I","B","UP"]
