### 2025-08-19 27680fa
Fix matplotlib user warning for nympy v2
