+++ tests/test_uncertainty.py
@@ -7,9 +7,10 @@ Pytest suite for testing uncertainty visualization functions in
 kdiagram.plot.uncertainty.
 """
 import re
-import sys
 from datetime import datetime, timedelta
 from unittest.mock import patch
+from packaging.version import parse as parse_version
+
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -66,7 +67,6 @@ def sample_data_coverage():
         "names": names,
     }
 
-
 @pytest.fixture
 def sample_data_drift():
     """Provides sample DataFrame for plot_model_drift tests."""
@@ -119,14 +119,15 @@ def sample_data_velocity():
     return {"df": df, "q50_cols": q50_cols}
 
 
-@pytest.mark.parametrize("kind", ["line", "bar", "pie", "radar"])
+@pytest.mark.parametrize("kind", ["line", "bar", "radar"])
 def test_plot_coverage_single_model_quantile(sample_data_coverage, kind):
     """Test plot_coverage with one quantile model for various kinds."""
     data = sample_data_coverage
 
     try:
+        if kind =="pie" and parse_version(
+                np.__version__) >= parse_version("2.0.0"):
         # The warning only happens for polar plots in newer library versions
-        if kind in ("pie", "radar") and sys.version_info >= (3, 11):
             # On newer versions, expect and catch the warning
             with pytest.warns(UserWarning, match=r"result dtype changed"):
                 plot_coverage(
@@ -161,44 +162,17 @@ def test_plot_coverage_single_model_quantile(sample_data_coverage, kind):
 def test_plot_coverage_multi_model_quantile_radar(sample_data_coverage):
     """Test plot_coverage with multiple quantile models (radar)."""
     data = sample_data_coverage
-
-    try:
-        # The Matplotlib/NumPy dtype warning can affect polar plots like 'radar'
-        # in newer library versions (Python 3.11+).
-        if sys.version_info >= (3, 11):
-            # On newer versions, expect and catch the warning
-            with pytest.warns(UserWarning, match=r"result dtype changed"):
-                plot_coverage(
-                    data["y_true"],
-                    data["y_pred_q1"],
-                    data["y_pred_q2"],  # Two models
-                    names=data["names"][:2],
-                    q=data["q"],
-                    kind="radar",
-                    cov_fill=True,
-                    figsize=(6, 6),
-                )
-        else:
-            # On older versions, run without expecting a warning
-            plot_coverage(
-                data["y_true"],
-                data["y_pred_q1"],
-                data["y_pred_q2"],
-                names=data["names"][:2],
-                q=data["q"],
-                kind="radar",
-                cov_fill=True,
-                figsize=(6, 6),
-            )
-
-        assert len(plt.get_fignums()) > 0, "Plot should be created"
-
-    except Exception as e:
-        pytest.fail(f"plot_coverage raised an exception: {e}")
-
-    finally:
-        plt.close("all")
-
+    plot_coverage(
+        data["y_true"],
+        data["y_pred_q1"],
+        data["y_pred_q2"],
+        names=data["names"][:2],
+        q=data["q"],
+        kind="radar",
+        cov_fill=True,
+        figsize=(6, 6),
+    )
+    assert len(plt.get_fignums()) > 0, "Plot should be created"
 
 def test_plot_coverage_single_model_point(sample_data_coverage):
     """Test plot_coverage with a single point forecast model."""
@@ -1025,10 +999,6 @@ def test_plot_interval_width_negative_width_warning(sample_data_iw):
     df_neg.loc[0:10, data["q_cols"][1]] = df_neg.loc[0:10, data["q_cols"][0]] - 1
[…]
