+++ joss_reviews.md
@@ -1,38 +1,103 @@
 Dear Reviewer Cubeth ,
 
-Thank you very much for your thoughtful and constructive feedback on the `k-diagram` package. I truly appreciate 
-your positive comments on the documentation structure and your excellent suggestion to integrate scientific 
-references for the visualization methods. This is a crucial point that significantly enhances the scientific 
-rigor and context of the project, and I am grateful you brought it to my attention.
+Thank you very much for your thoughtful and constructive feedback on the 
+`k-diagram` package. I truly appreciate  your positive comments on 
+the documentation structure and your excellent suggestion to integrate 
+scientific references for the visualization methods. This is a crucial 
+point that significantly enhances the scientific rigor and context 
+of the project, and I am grateful you brought it to my attention.
 
-You are absolutely right that readers would benefit from understanding the origins of each diagnostic plot—whether 
-they are established methods or novel contributions from this work. I have now completed a comprehensive update to 
-address this across the JOSS paper, the documentation, and the code's docstrings.
+You are absolutely right that readers would benefit from understanding 
+the origins of each diagnostic plot—whether they are established methods 
+or novel contributions from this work. I have now completed a comprehensive 
+update to address this across the JOSS paper, the documentation, and the 
+code's docstrings.
 
 Here is a summary of the actions taken:
 
 1.  **JOSS Paper (`paper.md`)**:
-    * The **Functionality** section has been revised to provide clear scientific context. I have added citations 
-    for established, foundational methods like Taylor Diagrams `[@Taylor2001]`, Reliability Diagrams `[@Jolliffe2012]`, 
+
+    * The **Functionality** section has been revised to provide clear 
+    scientific context. I have added citations for established, foundational 
+    methods like Taylor Diagrams `[@Taylor2001]`, Reliability Diagrams `[@Jolliffe2012]`, 
     and Kernel Density Estimation `[@Silverman1986]`.
-    * For the novel visualizations developed as part of this research (e.g., polar error bands, quiver plots, feature 
-    fingerprints), I have cited our accompanying research paper `[@kouadiob2025]` to properly attribute their origin.
+    * For the novel visualizations developed as part of this research 
+    (e.g., polar error bands, quiver plots, feature fingerprints), I have 
+    cited our accompanying research paper `[@kouadiob2025]` to properly 
+    attribute their origin.
+    
+    However it you want me to insert some equations, Im willing to do it so.
 
 2.  **Full Documentation (User Guide & Docstrings)**:
-    * I have integrated the `sphinxcontrib-bibtex` extension into the documentation build process to handle citations
-    professionally. A central `references.bib` file has been created.
-    * The **User Guide** pages (`uncertainty.rst`, `errors.rst`, `comparison.rst`, etc.) have been updated with `:cite:` 
-    directives, linking the descriptions of the plots directly to their foundational papers.
-    * Following your suggestion, the **docstrings** for the relevant functions have also been updated to include 
-    a `References` section with plain-text citations, making the scientific context accessible directly from the code.
-    * I also added an admonition box at the top of the uncertainty user guide, which directs readers to our research 
+
+    * I have integrated the `sphinxcontrib-bibtex` extension into the 
+    documentation build process to handle citations professionally. 
+    A central `references.bib` file has been created.
+    * The **User Guide** pages (`uncertainty.rst`, `errors.rst`, 
+    `comparison.rst`, etc.) have been updated with `:footcite:` directives( 
+    ``footsize``  to avoid duplicated citations: https://sphinxcontrib-bibtex.readthedocs.io/en/latest/usage.html ), 
+    linking the descriptions of the plots directly to their foundational papers.
+    
[…]
