### 2025-08-19 c6a0236
Fix matplotlib user warning for Numpy v2
