### 2025-08-19 f32a72a
Remove if __name__=='__main__': in tests
