+++ tests/test_io.py
@@ -15,7 +15,10 @@ class FakeResponse:
     """Context-manager mock for requests.get(..., stream=True)."""
 
     def __init__(
-        self, data: bytes, status: int = 200, content_length: int | None = None
+        self,
+        data: bytes,
+        status: int = 200,
+        content_length: int | None = None,
     ):
         self._data = data
         self.status_code = status
@@ -55,7 +58,9 @@ def test_fancier_downloader_fallback_without_tqdm(monkeypatch, tmp_path):
     # make tqdm import raise ImportError
     real_import = builtins.__import__
 
-    def _failing_import(name, globals=None, locals=None, fromlist=(), level=0):
+    def _failing_import(
+        name, globals=None, locals=None, fromlist=(), level=0
+    ):
         if name == "tqdm" or (name == "tqdm" and fromlist):
             raise ImportError("no tqdm")
         return real_import(name, globals, locals, fromlist, level)
@@ -71,7 +76,9 @@ def test_fancier_downloader_fallback_without_tqdm(monkeypatch, tmp_path):
     monkeypatch.setattr(io_mod, "download_file", fake_download)
 
     with pytest.warns(UserWarning, match="tqdm is not installed"):
-        out = io_mod.fancier_downloader("http://x", "f.bin", dstpath=str(tmp_path))
+        out = io_mod.fancier_downloader(
+            "http://x", "f.bin", dstpath=str(tmp_path)
+        )
     assert out == "returned-path"
     assert called["args"] == ("http://x", "f.bin", str(tmp_path))
 
@@ -79,7 +86,9 @@ def test_fancier_downloader_fallback_without_tqdm(monkeypatch, tmp_path):
 def test_fancier_downloader_success_with_progress(tmp_path, monkeypatch):
     monkeypatch.chdir(tmp_path)
     data = b"abcdef" * 10  # 60 bytes
-    monkeypatch.setattr(io_mod.requests, "get", lambda *a, **k: FakeResponse(data))
+    monkeypatch.setattr(
+        io_mod.requests, "get", lambda *a, **k: FakeResponse(data)
+    )
 
     # provide a tiny tqdm stand-in so import succeeds (if tqdm not installed)
     if "tqdm" not in sys.modules:
@@ -102,7 +111,9 @@ def test_fancier_downloader_check_size_mismatch(monkeypatch, tmp_path, mode):
     monkeypatch.chdir(tmp_path)
     data = b"x" * 50
     monkeypatch.setattr(
-        io_mod.requests, "get", lambda *a, **k: FakeResponse(data, content_length=100)
+        io_mod.requests,
+        "get",
+        lambda *a, **k: FakeResponse(data, content_length=100),
     )
 
     # Provide minimal tqdm stub if missing
@@ -136,7 +147,9 @@ def test_fancier_downloader_check_size_mismatch(monkeypatch, tmp_path, mode):
 def test_fancier_downloader_moves_to_dst(monkeypatch, tmp_path):
     data = b"hello world"
     monkeypatch.chdir(tmp_path)
-    monkeypatch.setattr(io_mod.requests, "get", lambda *a, **k: FakeResponse(data))
+    monkeypatch.setattr(
+        io_mod.requests, "get", lambda *a, **k: FakeResponse(data)
+    )
     if "tqdm" not in sys.modules:
         sys.modules["tqdm"] = types.SimpleNamespace(
             tqdm=lambda **kw: types.SimpleNamespace(
@@ -157,7 +170,9 @@ def test_fancier_downloader_moves_to_dst(monkeypatch, tmp_path):
 def test_fancier_downloader_move_error_handling(monkeypatch, tmp_path):
     data = b"data"
     monkeypatch.chdir(tmp_path)
[…]
