+++ docs/source/release_notes/v1.2.0.rst
@@ -4,46 +4,20 @@
 Version 1.2.0
 ----------------
 
-*(Released: 2025-08-16)*
+*(Released: 2025-08-20)*
 
-This is a major feature release that significantly expands the `k-diagram`
-library's capabilities for error analysis and distribution visualization.
-It introduces a new `errors` module with advanced polar plots for
-diagnosing systemic vs. random errors, comparing multiple models, and
+This is a major feature release that expands the ``k-diagram``
+library's capabilities for **error analysis** and **distribution visualization**.
+It introduces a new `error analysis module <https://k-diagram.readthedocs.io/en/latest/user_guide/errors.html>`_ 
+with advanced polar plots for diagnosing systemic vs. random errors, comparing multiple models, and
 visualizing 2D uncertainty. Additionally, histogram and density
 plotting utilities have been centralized and enhanced.
-
-Highlights
-~~~~~~~~~~~~~
-
-* |Feature| **New Error Analysis Module** (``kdiagram.plot.errors``):
-  A dedicated suite of plots for deep error diagnostics.
-  
-  - :func:`~kdiagram.plot.errors.plot_error_bands` to separate
-    systemic bias from random error.
-  - :func:`~kdiagram.plot.errors.plot_error_violins` to compare
-    full error distributions of multiple models.
-  - :func:`~kdiagram.plot.errors.plot_error_ellipses` for
-    visualizing 2D positional uncertainty.
-
-* |Feature| **New Advanced Uncertainty Plots**:
-
-  - :func:`~kdiagram.plot.uncertainty.plot_polar_heatmap` to find
-    correlations between cyclical features and magnitudes.
-  - :func:`~kdiagram.plot.uncertainty.plot_polar_quiver` to visualize
-    vector fields like forecast revisions.
-
-* |Feature| **New Histogram & Density Utilities**:
-
-  - Centralized :func:`~kdiagram.utils.plot_hist_kde` for
-    robust 1D distribution plotting.
-  - New :func:`~kdiagram.plot.uncertainty.plot_radial_density_ring`
-    for visualizing 1D distributions in a polar format.
 
 New Features
 ~~~~~~~~~~~~~
 
-* |New| New module ``kdiagram.plot.errors`` with three powerful plots:
+* |Feature| **Error Analysis Module** (``kdiagram.plot.errors``):
+  A dedicated suite of plots for deep error diagnostics.
 
   - :func:`~kdiagram.plot.errors.plot_error_bands`: Aggregates errors
     into angular bins to plot mean error (bias) and a standard
@@ -55,7 +29,7 @@ New Features
     2D uncertainty by drawing confidence ellipses based on radial and
     angular standard deviations.
 
-* |New| New plots added to ``kdiagram.plot.uncertainty``:
+* |Feature| **Advanced Uncertainty Plots** (``kdiagram.plot.uncertainty``):
 
   - :func:`~kdiagram.plot.uncertainty.plot_polar_heatmap`: Creates a
     2D density plot on a polar grid to identify "hot spots" where
@@ -64,19 +38,22 @@ New Features
     arrows on a polar grid to show magnitude and direction, ideal for
     visualizing forecast updates or error vectors.
 
-* |New| New utility module ``kdiagram.utils.hist``:
-  :func:`~kdiagram.utils.plot_hist_kde`: A robust, standalone
-  function for plotting histograms with optional KDE overlays,
-  supporting pandas DataFrames, Series, and NumPy arrays.
+* |Feature| **Model Comparison Plot** (``kdiagram.plot.comparison``):
 
-API Changes
-~~~~~~~~~~~~~
+  - :func:`~kdiagram.plot.comparison.plot_horizon_metrics`: Creates a polar 
+    bar chart to compare key metrics (like interval width and median value) 
[…]
