### 2025-08-19 912094f
Safe doc build : release v1.2.0
