### 2025-08-20 fa8bccb
Update documentation: citing.rst
