### 2025-08-20 9e93ee3
Format kdiagram.__init__.py
