+++ kdiagram/__init__.py
@@ -33,7 +33,7 @@ def _lazy_import(module_name, alias=None):
 try:
     from ._scm_version import version as __version__
 except Exception:
-    # Fallback version 
+    # Fallback version
     __version__ = "1.2.0"
 
 # Dependency check
@@ -66,6 +66,7 @@ if _missing_dependencies:
 
 # Re-export config helpers
 from .config import configure_warnings, warnings_config  # noqa: F401, E402
+
 # from . import datasets # noqa: E402
 from .plot import (  # noqa: E402
     plot_actual_vs_predicted,
