### 2025-08-21 dd24e0d
Update documentation : fix reference in hist1d page
