### 2025-08-22 93ab17e
Update paper.md
