+++ paper.md
@@ -83,14 +83,6 @@ and gain a deeper, more actionable understanding of their forecasting models.
 
 # Functionality
 
-`k-diagram` is implemented in Python [@python3; @pythonbook], leveraging core scientific libraries
-including `numpy` [@harris2020array], `pandas` [@mckinney-proc-scipy-2010; @reback2020pandas], 
-`matplotlib` [@Hunter:2007], `scipy` [@2020SciPy-NMeth],
-and `scikit-learn` [@scikit-learn]. The core functionality is organized around diagnosing 
-key aspects of forecast quality \autoref{fig1:workflow}:
-
-# Functionality
-
 `k-diagram` is implemented in Python [@python3; @pythonbook], leveraging
 core scientific libraries including `numpy` [@harris2020array],
 `pandas` [@mckinney-proc-scipy-2010; @reback2020pandas],
