+++ paper.md
@@ -81,13 +81,6 @@ answers to key diagnostic questions (e.g., "Where is a forecast least certain?",
 for any researcher or practitioner seeking to move beyond aggregate metrics 
 and gain a deeper, more actionable understanding of their forecasting models.
 
-# Functionality
-
-`k-diagram` is implemented in Python [@python3; @pythonbook], leveraging core scientific libraries
-including `numpy` [@harris2020array], `pandas` [@mckinney-proc-scipy-2010; @reback2020pandas], 
-`matplotlib` [@Hunter:2007], `scipy` [@2020SciPy-NMeth],
-and `scikit-learn` [@scikit-learn]. The core functionality is organized around diagnosing 
-key aspects of forecast quality \autoref{fig1:workflow}:
 
 # Functionality
 
@@ -151,63 +144,59 @@ around a cohesive API for diagnosing key aspects of forecast quality
 
 The package is designed for ease of use and customization, allowing
 users to control plot aesthetics, angular coverage (`acov`), and color mapping
-to tailor the visualizations for their specific domain. \autoref{fig2:performance} 
-demonstrates some of these key diagnostic plots.
+to tailor the visualizations for their specific domain. 
 
-![Figure 2: Model performance evaluation. (a) Coverage Evaluation: radial plot comparing empirical coverage against nominal quantile levels (average coverage = 0.811). (b) Model Error Distributions: The radial axis represents the error value, with the dashed circle indicating zero error. The width of each violin shows the density of errors, revealing that the "Good Model" is unbiased and consistent, the "Biased Model" consistently under-predicts, and the "Inconsistent Model" has high variance. (c) Forecast Horizon Drift: radial bar chart of uncertainty width (Q90–Q10) for forecast years 2023–2026, illustrating increasing prediction uncertainty.\label{fig2:performance}](docs/source/_static/paper_fig2.png)
+# Illustrative Diagnostics and Application
 
-### Installation 
+The practical utility of `k-diagram` is best demonstrated through its key 
+diagnostic plots, which reveal model behaviors often hidden by aggregate
+metrics (see \autoref{fig2:performance}).
 
-The latest stable release of `k-diagram` is available on the Python Package 
-Index (PyPI) and can be installed using pip:
+The **Coverage Evaluation** plot (\autoref{fig2:performance}a) provides a 
+point-wise diagnostic of interval performance. While the overall coverage of 
+81.1% is close to the nominal 80% for a Q10-Q90 interval, the plot allows 
+for a deeper analysis of where the interval failures occur. The 
+**Model Error Distributions** plot (\autoref{fig2:performance}b) offers 
+a powerful comparative view, using polar violins to contrast the
+ full error profiles of multiple models. It clearly distinguishes 
+a "Good Model" (unbiased, low variance) from a "Biased Model" 
+(consistently under-predicting) and an "Inconsistent Model" (high variance), 
+revealing performance trade-offs that a single error score would miss. 
+Finally, the **Forecast Horizon Drift** plot (\autoref{fig2:performance}c) 
+visualizes how uncertainty evolves over time. The increasing height of the 
+bars from 2023 to 2026 provides an immediate and intuitive confirmation 
+that the model's uncertainty grows as it forecasts further into the future.
 
-```bash
-pip install k-diagram
-```
-Detailed installation instructions are available in the documentation.
+These visualization methods were developed alongside research applying 
+advanced deep learning models, such as physics-informed deep learning[^1], 
+to complex environmental forecasting challenges [@kouadiob2025]. Specifically, 
+these polar diagnostics were utilized to analyze and interpret the uncertainty 
+associated with land subsidence predictions using an Extreme Temporal Fusion 
[…]
