### 2025-08-23 74ea29f
Update paper.md
