### 2025-08-23 76f82be
Update paper: Acknowledgments.md
