### 2025-08-24 e142b5d
Revise module probabilistic
