+++ kdiagram/plot/__init__.py
@@ -2,15 +2,28 @@ from .comparison import (
     plot_horizon_metrics,
     plot_model_comparison,
     plot_reliability_diagram,
+    plot_polar_reliability
 )
-from .errors import plot_error_bands, plot_error_ellipses, plot_error_violins
+from .errors import ( 
+    plot_error_bands, 
+    plot_error_ellipses, 
+    plot_error_violins
+)
 from .evaluation import (
     plot_taylor_diagram,
     plot_taylor_diagram_in,
     taylor_diagram,
 )
-from .feature_based import plot_feature_fingerprint
-from .relationship import plot_relationship
+from .feature_based import ( 
+    plot_feature_fingerprint, 
+    plot_feature_interaction
+)
+from .relationship import ( 
+    plot_relationship, 
+    plot_error_relationship, 
+    plot_residual_relationship, 
+    plot_conditional_quantiles
+)
 from .uncertainty import (
     plot_actual_vs_predicted,
     plot_anomaly_magnitude,
@@ -25,6 +38,13 @@ from .uncertainty import (
     plot_temporal_uncertainty,
     plot_uncertainty_drift,
     plot_velocity,
+)
+from .probabilistic import ( 
+    plot_crps_comparison, 
+    plot_pit_histogram, 
+    plot_polar_sharpness, 
+    plot_credibility_bands, 
+    plot_calibration_sharpness
 )
 
 __all__ = [
@@ -43,13 +63,24 @@ __all__ = [
     "plot_taylor_diagram_in",
     "taylor_diagram",
     "plot_feature_fingerprint",
+    "plot_feature_interaction", 
     "plot_relationship",
     "plot_radial_density_ring",
     "plot_reliability_diagram",
+    "plot_polar_reliability", 
     "plot_horizon_metrics",
     "plot_polar_heatmap",
     "plot_polar_quiver",
     "plot_error_bands",
     "plot_error_ellipses",
     "plot_error_violins",
+    "plot_crps_comparison",
+    "plot_pit_histogram",
+    "plot_polar_sharpness", 
+    "plot_credibility_bands",
+    "plot_calibration_sharpness", 
+    "plot_error_relationship", 
+    "plot_residual_relationship", 
+    "plot_conditional_quantiles", 
+    
 ]
