+++ kdiagram/plot/feature_based.py
@@ -1,6 +1,8 @@
 # License: Apache 2.0
 # Author: LKouadio <etanoyau@gmail.com>
 
+from __future__ import annotations 
+
 import warnings
 from typing import Any, Optional, Union
 
@@ -9,12 +11,257 @@ import numpy as np
 import pandas as pd
 
 from ..compat.matplotlib import get_cmap
-from ..decorators import check_non_emptiness
+from ..decorators import check_non_emptiness, isdf
 from ..utils.handlers import columns_manager
-from ..utils.validator import ensure_2d
+from ..utils.validator import ensure_2d, exist_features
+from ..utils.plot import set_axis_grid
 
-__all__ = ["plot_feature_fingerprint"]
+__all__ = ["plot_feature_fingerprint", "plot_feature_interaction"]
 
+
+@check_non_emptiness(params=["df"])
+@isdf
+def plot_feature_interaction(
+    df: pd.DataFrame,
+    theta_col: str,
+    r_col: str,
+    color_col: str,
+    *,
+    statistic: str = "mean",
+    theta_period: Optional[float] = None,
+    theta_bins: int = 24,
+    r_bins: int = 10,
+    title: Optional[str] = None,
+    figsize: tuple[float, float] = (8, 8),
+    cmap: str = "viridis",
+    show_grid: bool = True,
+    grid_props: dict[str, Any] | None = None,
+    mask_radius: bool=False, 
+    savefig: Optional[str] = None,
+    dpi: int = 300,
+):
+    # --- Input Validation ---
+    required_cols = [theta_col, r_col, color_col]
+    exist_features(df, features=required_cols)
+    data = df[required_cols].dropna().copy()
+    if data.empty:
+        warnings.warn(
+            "DataFrame is empty after dropping NaNs.", UserWarning)
+        return None
+
+    # --- Binning and Aggregation ---
+    if theta_period:
+        data['theta_rad'] = (
+            (data[theta_col] % theta_period) / theta_period) * 2 * np.pi
+    else:
+        min_theta, max_theta = data[theta_col].min(), data[theta_col].max()
+        if (max_theta - min_theta) > 1e-9:
+            data['theta_rad'] = ((data[theta_col] - min_theta) /
+                                 (max_theta - min_theta)) * 2 * np.pi
+        else:
+            data['theta_rad'] = 0
+
+    data['theta_bin'] = pd.cut(data['theta_rad'], bins=theta_bins)
+    data['r_bin'] = pd.cut(data[r_col], bins=r_bins)
+
+    # Group by both bins and calculate the statistic
+    heatmap_data = data.groupby(
+        ['theta_bin', 'r_bin'], observed=False
+        )[color_col].agg(statistic).unstack()
+    
+    # Get bin edges for plotting
+    theta_edges = np.linspace(0, 2 * np.pi, theta_bins + 1)
+    r_edges = np.linspace(data[r_col].min(), data[r_col].max(), r_bins + 1)
+    
+    # --- Plotting ---
+    fig, ax = plt.subplots(figsize=figsize, subplot_kw={"projection": "polar"})
+    
+    T, R = np.meshgrid(theta_edges, r_edges)
+    
+    # Note: pcolormesh expects C with shape (num_r_bins, num_theta_bins)
+    # The unstacked heatmap_data has shape (num_theta_bins, num_r_bins)
+    pcm = ax.pcolormesh(T, R, heatmap_data.T.values, cmap=cmap, shading="auto")
+
+    # --- Formatting ---
+    cbar = fig.colorbar(pcm, ax=ax, pad=0.1, shrink=0.75)
[…]

[…]
