+++ tests/test_mathext_.py
@@ -1,0 +1,146 @@
+import pytest
+import numpy as np
+
+from kdiagram.utils.mathext import (
+    compute_coverage_score,
+    compute_winkler_score,
+    build_cdf_interpolator,
+    compute_pinball_loss,
+    calculate_calibration_error,
+)
+
+# --- Fixture for generating test data ---
+
+@pytest.fixture(scope="module")
+def forecast_data():
+    """Provides consistent data for testing math utilities."""
+    np.random.seed(0)
+    y_true = np.array([1, 2, 3, 4, 5, 6, 7, 8, 9, 10])
+    # A forecast that is sometimes too narrow
+    y_pred_lower = np.array([0, 2, 2, 5, 4, 7, 6, 9, 8, 9])
+    y_pred_upper = np.array([2, 3, 4, 6, 6, 8, 8, 10, 11, 11])
+    
+    # Probabilistic data
+    quantiles = np.array([0.1, 0.5, 0.9])
+    preds_quantiles = np.array([
+        [8, 10, 12], # y_true = 10 -> PIT = 0.66
+        [0, 1, 2],   # y_true = 1 -> PIT = 0.66
+        [4, 5, 6],   # y_true = 5 -> PIT = 0.66
+    ])
+    y_true_prob = np.array([10, 1, 5])
+
+    return {
+        "y_true": y_true,
+        "lower": y_pred_lower,
+        "upper": y_pred_upper,
+        "y_true_prob": y_true_prob,
+        "preds_quantiles": preds_quantiles,
+        "quantiles": quantiles
+    }
+
+# --- Tests for compute_coverage_score ---
+
+def test_compute_coverage_score_within(forecast_data):
+    """Test standard 'within' coverage calculation."""
+    # Expected: [F, T, T, F, T, T, T, T, T, T] -> 8/10 = 0.8
+    score = compute_coverage_score(
+        forecast_data["y_true"], forecast_data["lower"], forecast_data["upper"]
+    )
+    assert score == pytest.approx(0.8)
+
+def test_compute_coverage_score_above_and_below(forecast_data):
+    """Test 'above' and 'below' methods."""
+    # y_true > upper: [F, F, F, F, F, F, F, F, F, F] -> 0
+    above_score = compute_coverage_score(
+        forecast_data["y_true"], forecast_data["lower"], forecast_data["upper"], method='above'
+    )
+    assert above_score == pytest.approx(0.0)
+    
+    # y_true < lower: [F, F, F, F, F, F, F, F, F, F] -> This is wrong, let's re-check
+    # y_true: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
+    # lower:  [0, 2, 2, 5, 4, 7, 6, 9, 8, 9]
+    # below: [F, F, F, T, F, T, F, T, F, F] -> 3/10 = 0.3... wait, y_true < lower
+    # 1<0(F), 2<2(F), 3<2(F), 4<5(T), 5<4(F), 6<7(T), 7<6(F), 8<9(T), 9<8(F), 10<9(F)
+    # Ah, the logic was based on a different example. Let's re-calculate.
+    # y_true: [1, 2, 3, 4, 5, 6]
+    # lower:  [2, 2, 4, 4, 6, 6]
+    # upper:  [3, 3, 5, 5, 7, 7]
+    # y_true < lower -> [T, F, T, F, T, F] -> 3/6 = 0.5
+    y_true = np.array([1, 2, 3, 4, 5, 6])
+    lower = np.array( [2, 2, 4, 4, 6, 6])
+    upper = np.array( [3, 3, 5, 5, 7, 7])
+    
+    below_score = compute_coverage_score(y_true, lower, upper, method='below')
[…]
