### 2025-08-24 36236e7
Format Math formulation in paper.md
