### 2025-08-24 9950dc4
Update paper: Check math to render propertly
