### 2025-08-25 78fee75
Add context plots and tests
