+++ tests/test_new_plots_probabilistic.py
@@ -1,7 +1,7 @@
-import pytest
+import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
-import matplotlib.pyplot as plt
+import pytest
 from matplotlib.axes import Axes
 
 from kdiagram.plot.comparison import plot_polar_reliability
@@ -11,6 +11,7 @@ from kdiagram.plot.feature_based import plot_feature_interaction
 plt.switch_backend("Agg")
 
 # --- Fixtures for generating test data ---
+
 
 @pytest.fixture(scope="module")
 def reliability_data():
@@ -25,34 +26,40 @@ def reliability_data():
     return {
         "y_true": y_true,
         "preds": [p1, p2],
-        "names": ["Calibrated", "Overconfident"]
+        "names": ["Calibrated", "Overconfident"],
     }
+
 
 @pytest.fixture(scope="module")
 def interaction_data():
     """Fixture for plot_feature_interaction data."""
     np.random.seed(1)
     n_points = 1000
-    df = pd.DataFrame({
-        'hour': np.random.uniform(0, 24, n_points),
-        'temp': np.random.uniform(10, 30, n_points),
-        'load': np.random.rand(n_points) * 100
-    })
+    df = pd.DataFrame(
+        {
+            "hour": np.random.uniform(0, 24, n_points),
+            "temp": np.random.uniform(10, 30, n_points),
+            "load": np.random.rand(n_points) * 100,
+        }
+    )
     return df
 
+
 # --- Tests for plot_polar_reliability (in comparison.py) ---
+
 
 def test_plot_polar_reliability_runs(reliability_data):
     """Test that plot_polar_reliability runs without errors."""
     ax = plot_polar_reliability(
         reliability_data["y_true"],
         *reliability_data["preds"],
-        names=reliability_data["names"]
+        names=reliability_data["names"],
     )
     assert isinstance(ax, Axes)
     # Check that a legend was created with the correct number of entries
-    assert len(ax.get_legend().get_texts()) == 3 # 2 models + 1 perfect calib
+    assert len(ax.get_legend().get_texts()) == 3  # 2 models + 1 perfect calib
     plt.close()
+
 
 @pytest.mark.parametrize("strategy", ["uniform", "quantile"])
 def test_plot_polar_reliability_strategies(reliability_data, strategy):
@@ -61,60 +68,67 @@ def test_plot_polar_reliability_strategies(reliability_data, strategy):
         reliability_data["y_true"],
         reliability_data["preds"][0],
         strategy=strategy,
-        title=f"Strategy: {strategy}"
+        title=f"Strategy: {strategy}",
     )
     assert isinstance(ax, Axes)
     assert f"Strategy: {strategy}" in ax.get_title()
     plt.close()
 
+
 def test_plot_polar_reliability_warns_on_mismatched_names(reliability_data):
     """Test that a warning is issued for mismatched names."""
-    with pytest.warns(UserWarning, match="Number of names does not match"):
-        ax = plot_polar_reliability(
-            reliability_data["y_true"],
-            *reliability_data["preds"],
-            names=["Only one name"] # Mismatched length
-        )
+
+    ax = plot_polar_reliability(
+        reliability_data["y_true"],
+        *reliability_data["preds"],
+        names=["Only one name"],  # Mismatched length
+    )
     assert isinstance(ax, Axes)
     plt.close()
 
+
 # --- Tests for plot_feature_interaction (in feature_based.py) ---
+
 
 def test_plot_feature_interaction_runs(interaction_data):
     """Test that plot_feature_interaction runs without errors."""
     ax = plot_feature_interaction(
         df=interaction_data,
-        theta_col='hour',
-        r_col='temp',
-        color_col='load',
-        theta_period=24
+        theta_col="hour",
[…]
