+++ kdiagram/utils/mathext.py
@@ -1,32 +1,30 @@
 #   License: Apache 2.0
 #   Author: LKouadio <etanoyau@gmail.com>
 
-from __future__ import annotations 
+from __future__ import annotations
 
-from typing import (
-    overload, Literal, List, Tuple, Union, Optional, 
-    Callable
-)
-from scipy.stats import kstest, uniform
+from typing import Callable, Literal, Union, overload
+
 import numpy as np
 import pandas as pd
+from scipy.stats import kstest, uniform
 
-from .handlers import columns_manager 
-from .validator import validate_length_range
-from .validator import validate_yy, exist_features
+from .handlers import columns_manager
+from .validator import exist_features, validate_length_range, validate_yy
 
+FrameLike = Union[np.ndarray, pd.DataFrame, pd.Series]
 
 __all__ = [
-    "minmax_scaler", 
-    "compute_coverage_score", 
-    "compute_winkler_score", 
-    "build_cdf_interpolator", 
+    "minmax_scaler",
+    "compute_coverage_score",
+    "compute_winkler_score",
+    "build_cdf_interpolator",
     "calculate_calibration_error",
-    "compute_pinball_loss", 
-    "compute_pit", 
-    "compute_crps", 
-    "get_forecast_arrays"
-  ]
+    "compute_pinball_loss",
+    "compute_pit",
+    "compute_crps",
+    "get_forecast_arrays",
+]
 
 
 @overload
@@ -46,8 +44,7 @@ def get_forecast_arrays(
     ensure_numeric: bool = ...,
     coerce_numeric: bool = ...,
     copy: bool = ...,
-) -> np.ndarray:
-    ...
+) -> np.ndarray: ...
 
 
 @overload
@@ -67,15 +64,14 @@ def get_forecast_arrays(
     ensure_numeric: bool = ...,
     coerce_numeric: bool = ...,
     copy: bool = ...,
-) -> np.ndarray:
-    ...
+) -> np.ndarray: ...
 
 
 @overload
 def get_forecast_arrays(
     df: pd.DataFrame,
     actual_col: None,
-    pred_cols: List[str],
+    pred_cols: list[str],
     *,
     drop_na: bool = ...,
     na_policy: Literal["any", "all", "none"] = ...,
@@ -88,15 +84,14 @@ def get_forecast_arrays(
     ensure_numeric: bool = ...,
     coerce_numeric: bool = ...,
     copy: bool = ...,
-) -> np.ndarray:
-    ...
+) -> np.ndarray: ...
 
 
 @overload
 def get_forecast_arrays(
     df: pd.DataFrame,
     actual_col: str,
-    pred_cols: Union[str, List[str]],
+    pred_cols: str | list[str],
     *,
     drop_na: bool = ...,
     na_policy: Literal["any", "all", "none"] = ...,
@@ -109,8 +104,7 @@ def get_forecast_arrays(
     ensure_numeric: bool = ...,
     coerce_numeric: bool = ...,
     copy: bool = ...,
-) -> Tuple[np.ndarray, np.ndarray]:
-    ...
+) -> tuple[np.ndarray, np.ndarray]: ...
 
 
 @overload
@@ -130,8 +124,7 @@ def get_forecast_arrays(
     ensure_numeric: bool = ...,
     coerce_numeric: bool = ...,
     copy: bool = ...,
-) -> pd.Series:
-    ...
+) -> pd.Series: ...
 
 
 @overload
@@ -151,15 +144,14 @@ def get_forecast_arrays(
     ensure_numeric: bool = ...,
     coerce_numeric: bool = ...,
     copy: bool = ...,
-) -> pd.Series:
-    ...
+) -> pd.Series: ...
 
 
[…]

[…]
