+++ kdiagram/utils/__init__.py
@@ -10,6 +10,17 @@ from .forecast_utils import (
     pivot_forecasts_long,
 )
 from .hist import plot_hist_kde
+from .mathext import ( 
+    build_cdf_interpolator,
+    calculate_calibration_error,
+    compute_coverage_score,
+    compute_pinball_loss,
+    compute_winkler_score,
+    compute_crps, 
+    compute_pit, 
+    get_forecast_arrays, 
+    minmax_scaler
+)
 from .q_utils import (
     melt_q_data,
     pivot_q_data,
@@ -28,4 +39,15 @@ __all__ = [
     "calculate_probabilistic_scores",
     "bin_by_feature",
     "compute_interval_width",
+    "minmax_scaler",
+    "compute_coverage_score",
+    "compute_winkler_score",
+    "build_cdf_interpolator",
+    "calculate_calibration_error",
+    "compute_pinball_loss",
+    "compute_pit",
+    "compute_crps",
+    "get_forecast_arrays",
+    "minmax_scaler"
+    
 ]
