+++ docs/source/_autosummary/uncertainty/kdiagram.plot.uncertainty.plot_radial_density_ring.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.plot.uncertainty.plot\_radial\_density\_ring
-=====================================================
-
-.. currentmodule:: kdiagram.plot.uncertainty
-
-.. autofunction:: plot_radial_density_ring

[…]
