### 2025-08-26 6e4aa09
revise the new CLI configuration
