+++ kdiagram/cli/_utils.py
@@ -1,0 +1,822 @@
+# Author: LKouadio <etanoyau@gmail.com>
+# License: Apache License 2.0
+
+from __future__ import annotations
+
+import argparse
+import io
+import re
+import warnings
+from dataclasses import dataclass
+from collections.abc import Iterable, Mapping, Sequence
+from pathlib import Path
+from typing import Any, Union
+
+import numpy as np  # noqa
+import pandas as pd
+
+from ..core._io_utils import _normalize_ext
+from ..core.io import read_data, write_data
+
+PathLike = Union[str, Path]
+
+FileLike = io.IOBase
+
+
+__all__ = [
+    "load_df",
+    "save_df",
+    "ensure_columns",
+    "ensure_numeric",
+    "parse_list",
+    "parse_pairs",
+    "parse_quantiles",
+    "normalize_acov",
+    "expand_prefix_cols",
+    "detect_quantile_columns",
+    "natural_sort_key",
+]
+
+
+# ---------- IO helpers -------------------------------------------------
+
+
+def load_df(
+    src: PathLike | FileLike,
+    *,
+    format: str | None = None,
+    storage_options: Mapping[str, Any] | None = None,
+    errors: str = "raise",
+    verbose: int = 0,
+    index_col: str | Iterable[str] | None = None,
+    sort_index: bool = False,
+    drop_na: str | Iterable[str] | None = None,
+    fillna: Any | Mapping[str, Any] | None = None,
+    **kwargs: Any,
+) -> pd.DataFrame:
+    """Thin wrapper around core.read_data."""
+    df = read_data(
+        src,
+        format=format,
+        storage_options=storage_options,
+        errors=errors,
+        verbose=verbose,
+        index_col=index_col,
+        sort_index=sort_index,
+        drop_na=drop_na,
+        fillna=fillna,
+        **kwargs,
+    )
+    if isinstance(df, pd.DataFrame):
+        return df
+    # If chunked/HTML-list slipped through, force a DataFrame
+    return (
+        pd.concat(list(df), ignore_index=True)
+        if isinstance(df, Iterable)
+        else pd.DataFrame()
+    )
+
+
+def save_df(
+    df: pd.DataFrame,
+    dest: str | Path | io.IOBase,
+    *,
+    format: str | None = None,
+    errors: str = "raise",
+    overwrite: bool = False,
+    mkdirs: bool = True,
+    index: bool | None = None,
+    storage_options: Mapping[str, Any] | None = None,
+    **kwargs: Any,
+):
+    ext = _normalize_ext(dest, explicit=format)
+
+    # Default CSV writes to *no* index unless user specifies one
+    if index is None and ext == ".csv":
+        index = False
+
+    return write_data(
+        df,
+        dest,
+        format=ext,
+        errors=errors,
+        overwrite=overwrite,
+        mkdirs=mkdirs,
+        index=index,
+        storage_options=storage_options,
+        **kwargs,
+    )
+
+
+# ---------- Schema / validation ---------------------------------------
+
+
+def ensure_columns(
+    df: pd.DataFrame,
+    cols: Sequence[str],
+    *,
+    error: str = "raise",
+    name: str | None = None,
+) -> None:
+    """Validate required columns exist."""
+    need = list(cols)
+    miss = [c for c in need if c not in df.columns]
+    if miss and error == "raise":
+        lbl = f" for {name}" if name else ""
+        raise ValueError(f"Missing columns{lbl}: {miss}")
+    if miss and error == "warn":
+        import warnings
+
+        warnings.warn(f"Missing columns: {miss}", stacklevel=2)
+
+
+def ensure_numeric(
[…]
