### 2025-08-26 a2748fe
revise the new CLI configuration
