### 2025-08-26 34aaa94
Revise CLI configuration
