+++ tests/test_cli.py
@@ -1,594 +1,0 @@
-from __future__ import annotations
-
-import os
-import sys
-
-import matplotlib
-import numpy as np
-import pandas as pd
-import pytest
-
-# # Use a non-interactive backend
-# matplotlib.use("Agg")
-import kdiagram.cli as cli
-
-
-@pytest.fixture(autouse=True)
-def _mpl_agg(monkeypatch):
-    monkeypatch.setenv("MPLBACKEND", "Agg")
-
-
-@pytest.fixture
-def stub_plots(monkeypatch):
-    # stub every imported plot function to a sentinel
-    called = {}
-
-    def make_stub(name):
-        def f(*a, **k):
-            called[name] = (a, k)
-
-        return f
-
-    for fn in [
-        "plot_coverage",
-        "plot_model_drift",
-        "plot_velocity",
-        "plot_interval_consistency",
-        "plot_anomaly_magnitude",
-        "plot_uncertainty_drift",
-        "plot_actual_vs_predicted",
-        "plot_coverage_diagnostic",
-        "plot_interval_width",
-        "plot_temporal_uncertainty",
-        "taylor_diagram",
-        "plot_taylor_diagram_in",
-        "plot_taylor_diagram",
-        "plot_feature_fingerprint",
-        "plot_relationship",
-    ]:
-        if hasattr(cli, fn):
-            monkeypatch.setattr(cli, fn, make_stub(fn), raising=False)
-    # stub IO helpers
-    monkeypatch.setattr(
-        cli,
-        "read_csv_to_df",
-        lambda p: pd.DataFrame({"a": [1, 2], "b": [3, 4]}),
-    )
-    monkeypatch.setattr(
-        cli, "read_csv_to_numpy", lambda p, delimiter=",": np.array([1, 2, 3])
-    )
-    # stub show/savefig
-    monkeypatch.setattr(cli.plt, "show", lambda: None)
-    monkeypatch.setattr(cli.plt, "savefig", lambda *a, **k: None)
-    return called
-
-
-def run_cli(argv):
-    # emulate command-line
-    old = sys.argv[:]
-    try:
-        sys.argv = ["kdiagram"] + argv
-        cli.main()
-    finally:
-        sys.argv = old
-
-
-def test_cli_version(capsys):
-    with pytest.raises(SystemExit):
-        run_cli(["--version"])
-    out = capsys.readouterr()
-    assert "kdiagram" in out.err or out.out  # depending on argparse version
-
-
-@pytest.mark.parametrize(
-    "argv,called_fn",
-    [
-        (
-            ["plot_coverage", "true.csv", "pred1.csv", "--q", "0.1", "0.9"],
-            "plot_coverage",
-        ),
-        (
-            [
-                "plot_model_drift",
-                "data.csv",
-                "--q10-cols",
-                "q10",
-                "--q90-cols",
-                "q90",
-                "--horizons",
-                "h1",
-            ],
-            "plot_model_drift",
-        ),
-        (["plot_velocity", "data.csv", "--q50-cols", "q50"], "plot_velocity"),
-        (
-            [
-                "plot_interval_consistency",
-                "data.csv",
-                "--qlow-cols",
-                "q10",
-                "--qup-cols",
-                "q90",
-            ],
-            "plot_interval_consistency",
-        ),
-        (
-            [
-                "plot_anomaly_magnitude",
-                "data.csv",
-                "--actual-col",
-                "obs",
-                "--q-cols",
-                "q10",
-                "q90",
-            ],
[…]
