### 2025-08-27 c685e6d
Sep taylor diagram and evaluation module
