+++ kdiagram/cli/plot_context_err.py
@@ -1,0 +1,293 @@
+# License: Apache-2.0
+# Author: LKouadio <etanoyau@gmail.com>
+
+from __future__ import annotations
+
+import argparse
+from typing import Any
+
+import matplotlib.pyplot as plt
+
+from ..plot.context import (
+    plot_error_distribution,
+    plot_error_pacf,
+)
+from ._utils import (
+    add_bool_flag,
+    ensure_columns,
+    ensure_numeric,
+    load_df,
+    parse_figsize,
+)
+
+__all__ = [
+    "add_plot_error_distribution",
+    "add_plot_error_pacf",
+    "add_context_err",
+]
+
+
+def _cmd_plot_error_distribution(ns: argparse.Namespace) -> None:
+    """
+    CLI runner for plot-error-distribution.
+    """
+    df = load_df(ns.input, format=ns.format)
+
+    need = [ns.actual_col, ns.pred_col]
+    ensure_columns(df, need)
+    df = ensure_numeric(df, need, copy=True, errors="raise")
+
+    hist_kwargs: dict[str, Any] = {}
+    if ns.figsize:
+        hist_kwargs["figsize"] = ns.figsize
+    if ns.bins is not None:
+        hist_kwargs["bins"] = ns.bins
+    hist_kwargs["kde"] = ns.kde
+    hist_kwargs["density"] = ns.density
+    if ns.hist_color:
+        hist_kwargs["hist_color"] = ns.hist_color
+    if ns.kde_color:
+        hist_kwargs["kde_color"] = ns.kde_color
+    if ns.alpha is not None:
+        hist_kwargs["alpha"] = ns.alpha
+
+    ax = plot_error_distribution(
+        df=df,
+        actual_col=ns.actual_col,
+        pred_col=ns.pred_col,
+        title=ns.title,
+        xlabel=ns.xlabel,
+        **hist_kwargs,
+    )
+
+    # the plotting function does not savefig; handle here
+    if ns.savefig and ax is not None:
+        fig = ax.figure
+        fig.savefig(str(ns.savefig), dpi=ns.dpi, bbox_inches="tight")
+        plt.close(fig)
+
+
+def _cmd_plot_error_pacf(ns: argparse.Namespace) -> None:
+    """
+    CLI runner for plot-error-pacf.
+    """
+    df = load_df(ns.input, format=ns.format)
+
+    need = [ns.actual_col, ns.pred_col]
+    ensure_columns(df, need)
+    df = ensure_numeric(df, need, copy=True, errors="raise")
+
+    pacf_kwargs: dict[str, Any] = {}
+    if ns.lags is not None:
+        pacf_kwargs["lags"] = ns.lags
+    if ns.alpha is not None:
+        pacf_kwargs["alpha"] = ns.alpha
+    if ns.method:
+        pacf_kwargs["method"] = ns.method
+
+    plot_error_pacf(
+        df=df,
+        actual_col=ns.actual_col,
+        pred_col=ns.pred_col,
+        title=ns.title,
+        xlabel=ns.xlabel,
+        ylabel=ns.ylabel,
+        figsize=ns.figsize,
+        show_grid=ns.show_grid,
+        grid_props=None,
+        savefig=str(ns.savefig) if ns.savefig else None,
+        dpi=ns.dpi,
+        **pacf_kwargs,
+    )
+
+
+# ------------------------------- parsers --------------------------------
+
+
+def add_plot_error_distribution(
+    sub: argparse._SubParsersAction,
+) -> None:
+    p = sub.add_parser(
+        "plot-error-distribution",
+        help=("Histogram + KDE of forecast errors " "(actual - predicted)."),
+        description=(
+            "Compute residuals and plot their distribution using "
[…]
