+++ kdiagram/cli/plot_eval_performance.py
@@ -1,0 +1,271 @@
+# License: Apache-2.0
+# Author: LKouadio <etanoyau@gmail.com>
+
+from __future__ import annotations
+
+import argparse
+
+from ..plot.evaluation import plot_regression_performance
+from ._utils import (
+    ColumnsListAction,
+    _collect_point_preds,
+    _flatten_cols,
+    _parse_metric_values,
+    _parse_name_bool_map,
+    _resolve_metric_labels,
+    add_bool_flag,
+    ensure_columns,
+    ensure_numeric,
+    load_df,
+    parse_figsize,
+)
+
+__all__ = [
+    "add_plot_regression_performance",
+]
+
+
+def _cmd_plot_regression_performance(
+    ns: argparse.Namespace,
+) -> None:
+    """
+    CLI runner for regression performance polar chart.
+    Two modes:
+      (A) values-mode via --metric-values
+      (B) data-mode via --y-true + preds
+    """
+
+    # values-mode
+    mv = _parse_metric_values(ns.metric_values)
+    if mv:
+        names = _flatten_cols(ns.names) if ns.names else None
+        n_models = len(next(iter(mv.values())))
+        if names and len(names) != n_models:
+            raise SystemExit(
+                f"--names has {len(names)} entries but "
+                f"--metric-values imply {n_models} models"
+            )
+
+        plot_regression_performance(
+            y_true=None,
+            metric_values=mv,
+            names=names,
+            metrics=None,
+            add_to_defaults=False,
+            metric_labels=_resolve_metric_labels(ns),
+            higher_is_better=_parse_name_bool_map(ns.higher_is_better),
+            title=ns.title,
+            figsize=ns.figsize,
+            cmap=ns.cmap,
+            show_grid=ns.show_grid,
+            grid_props=None,
+            mask_radius=ns.mask_radius,
+            savefig=str(ns.savefig) if ns.savefig else None,
+            dpi=ns.dpi,
+        )
+        return
+
+    # data-mode
+    if not ns.y_true:
+        raise SystemExit(
+            "Provide --y-true and predictions via "
+            "--pred/--pred-cols/--model, or use "
+            "--metric-values for values-mode."
+        )
+
+    df = load_df(ns.input, format=ns.format)
+
+    # y_true
+    ensure_columns(df, [ns.y_true])
+    df = ensure_numeric(df, [ns.y_true], copy=True, errors="raise")
+    y_true = df[ns.y_true].to_numpy()
+
+    # collect predictions (each group => exactly one column)
+    yps, names_cli = _collect_point_preds(df, ns)
+
+    # names
+    names = _flatten_cols(ns.names) if ns.names else None
+    if names is None and names_cli:
+        names = list(names_cli)
+
+    # metrics list
+    metrics = _flatten_cols(ns.metrics) if ns.metrics else None
+
+    plot_regression_performance(
+        y_true,
+        *yps,
+        names=names,
+        metrics=metrics,
+        metric_values=None,
+        add_to_defaults=ns.add_to_defaults,
+        metric_labels=_resolve_metric_labels(ns),
+        higher_is_better=_parse_name_bool_map(ns.higher_is_better),
+        title=ns.title,
+        figsize=ns.figsize,
+        cmap=ns.cmap,
+        show_grid=ns.show_grid,
+        grid_props=None,
+        mask_radius=ns.mask_radius,
+        savefig=str(ns.savefig) if ns.savefig else None,
+        dpi=ns.dpi,
+    )
+
+
+def add_plot_regression_performance(
+    sub: argparse._SubParsersAction,
+) -> None:
+    p = sub.add_parser(
+        "plot-regression-performance",
+        help=(
+            "Polar performance chart for regression models using "
+            "multiple metrics."
[…]
