+++ kdiagram/datasets/make.py
@@ -11,8 +11,10 @@ the `k-diagram` package, particularly those focused on uncertainty.
 """
 from __future__ import annotations
 
+import re
 import textwrap
 import warnings
+from typing import Any, Callable
 
 import numpy as np
 import pandas as pd
@@ -24,6 +26,8 @@ __all__ = [
     "make_taylor_data",
     "make_multi_model_quantile_data",
     "make_cyclical_data",
+    "make_regression_data",
+    "make_classification_data",
 ]
 
 
@@ -1395,6 +1399,9 @@ def make_multi_model_quantile_data(
     if not bias_range[0] <= bias_range[1]:
         raise ValueError("bias_range must be (min, max) with min <= max.")
 
+    bias_ranges = _expand_param(bias_range, n_models, "bias_range")
+    width_ranges = _expand_param(width_range, n_models, "width_range")
+
     # --- Setup ---
     # Ensure unique and sorted quantiles
     quantiles_sorted = sorted(list(set(quantiles)))
@@ -1442,9 +1449,14 @@ def make_multi_model_quantile_data(
 
     # --- Generate predictions for each model ---
     for _i, model_name in enumerate(model_names_list):
-        # Sample model-specific parameters
-        model_bias = rng.uniform(bias_range[0], bias_range[1])
-        model_width = rng.uniform(width_range[0], width_range[1])
+        # Sample model-specific parameters from the expanded lists
+        current_bias_range = bias_ranges[_i]
+        current_width_range = width_ranges[_i]
+
+        model_bias = rng.uniform(current_bias_range[0], current_bias_range[1])
+        model_width = rng.uniform(
+            current_width_range[0], current_width_range[1]
+        )
 
         # Store generated quantiles temporarily before sorting
         temp_model_quantiles = {}
@@ -1724,3 +1736,1010 @@ References
 ----------
 .. footbibliography::
 """
+
+
+def make_regression_data(
+    n_samples: int = 200,
+    n_features: int = 1,
+    feature_range: tuple[float, float] = (0.0, 10.0),
+    n_models: int = 3,
+    model_profiles: dict[str, dict[str, Any]] | None = None,
+    true_func: Callable[[np.ndarray], np.ndarray] | None = None,
+    true_kind: str = "linear",  # 'linear'|'quadratic'|'sine'
+    true_coeff_range: tuple[float, float] = (-5.0, 5.0),
+    intercept: float = 5.0,
+    noise_on_true: float | Callable[[np.ndarray], np.ndarray] = 1.0,
+    heteroskedastic: bool = False,
+    hetero_strength: float = 0.5,
+    prefix: str = "pred_",
+    seed: int | None = 0,
+    as_frame: bool = False,
+    clip_negative: bool = False,
+    shuffle: bool = True,
+    model_names: list[str] | None = None,
+    feature_names: list[str] | None = None,
+) -> Bunch | pd.DataFrame:
+
+    # ---------- RNG ----------
+    rng = np.random.default_rng(seed)
+
+    # ---------- features ----------
+    lo, hi = float(feature_range[0]), float(feature_range[1])
+    if hi <= lo:
+        raise ValueError("feature_range must satisfy hi > lo.")
+
+    X = rng.uniform(lo, hi, size=(n_samples, n_features))
+
+    if not feature_names:
+        feature_names = [f"feature_{i+1}" for i in range(n_features)]
+    elif len(feature_names) != n_features:
+        raise ValueError("len(feature_names) must equal n_features.")
+
+    # ---------- true signal ----------
+    # allow user supplied function(X) -> shape (n_samples,)
+    if true_func is not None:
+        y_signal = np.asarray(true_func(X))
+        if y_signal.shape != (n_samples,):
[…]
