+++ kdiagram/plot/evaluation.py
@@ -2,7 +2,7 @@
 #   Author: LKouadio <etanoyau@gmail.com>
 
 import warnings
-from typing import Any, Callable, Optional, Union
+from typing import Any, Callable, Literal, Optional, Union
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -191,6 +191,7 @@ Positive Rate (FPR) at various threshold settings.
    \text{FPR} = \frac{FP}{FP + TN}
 
 This function adapts the concept to a polar plot:
+    
 - The **angle (θ)** is mapped to the False Positive Rate,
   spanning from 0 at 0° to 1 at 90°.
 - The **radius (r)** is mapped to the True Positive Rate,
@@ -780,6 +781,7 @@ various threshold settings.
    \text{Recall} = \frac{TP}{TP + FN}
 
 This function adapts the concept to a polar plot:
+    
 - The **angle (θ)** is mapped to **Recall**, spanning from 0
   at 0° to 1 at 90°.
 - The **radius (r)** is mapped to **Precision**, spanning from 0
@@ -1280,6 +1282,10 @@ def plot_regression_performance(
     add_to_defaults: bool = False,
     metric_labels: Optional[Union[dict[str, str], bool, list]] = None,
     higher_is_better: Optional[dict[str, bool]] = None,
+    norm: Literal["per_metric", "global", "none"] = "per_metric",
+    global_bounds: Optional[dict[str, tuple[float, float]]] = None,
+    min_radius: float = 0.02,
+    clip_to_bounds: bool = True,
     title: str = "Regression Model Performance",
     figsize: tuple[float, float] = (8, 8),
     cmap: str = "viridis",
@@ -1289,18 +1295,25 @@ def plot_regression_performance(
     savefig: Optional[str] = None,
     dpi: int = 300,
 ):
-    # --- Input Validation and Data Preparation ---
+    # --- 1. Determine Mode and Calculate Scores ---
+    # The function operates in two modes:
+    # a) "Values Mode": Pre-computed scores are provided.
+    # b) "Data Mode": Scores are computed from y_true and y_preds.
+
     if metric_values is not None:
-        if y_true is not None or y_preds:
+        # --- a) Values Mode: Use pre-computed scores ---
+        if (y_true is not None) or y_preds:
             raise ValueError(
-                "If `metric_values` is provided, `y_true`"
-                " and `y_preds` must be None."
+                "If `metric_values` is provided, `y_true` and "
+                "`y_preds` must be None."
             )
         scores = metric_values
         metric_names = list(scores.keys())
-        # Infer number of models from the first metric's list of scores
-        n_models = len(list(scores.values())[0])
-    elif y_true is not None and y_preds:
+        # Infer number of models from the first metric's list
+        n_models = len(next(iter(scores.values())))
+
+    elif (y_true is not None) and y_preds:
+        # --- b) Data Mode: Compute scores from data ---
         n_models = len(y_preds)
         default_metrics = [
             "r2",
@@ -1312,59 +1325,123 @@ def plot_regression_performance(
             metrics_to_use = default_metrics
         else:
             user_metrics = columns_manager(metrics)
-            if add_to_defaults:
-                metrics_to_use = default_metrics + user_metrics
-            else:
-                metrics_to_use = user_metrics
-
+            metrics_to_use = (
+                default_metrics + user_metrics
+                if add_to_defaults
+                else user_metrics
+            )
+        # Helper function calculates scores, ensuring higher is better
         scores = _get_scores(
-            y_true, y_preds, metrics_to_use, higher_is_better
+            y_true,
+            list(y_preds),
+            metrics_to_use,
[…]
