### 2025-08-28 db9cc44
Revise the plot_regression_performance and quickstart
