+++ docs/source/_autosummary/context/kdiagram.plot.context.plot_error_autocorrelation.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.plot.context.plot\_error\_autocorrelation
-==================================================
-
-.. currentmodule:: kdiagram.plot.context
-
-.. autofunction:: plot_error_autocorrelation
