+++ docs/source/_autosummary/utils/kdiagram.utils.compute_crps.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.utils.compute\_crps
-============================
-
-.. currentmodule:: kdiagram.utils
-
-.. autofunction:: compute_crps
