+++ docs/source/_autosummary/taylor_diagram/kdiagram.plot.taylor_diagram.taylor_diagram.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.plot.taylor\_diagram.taylor\_diagram
-=============================================
-
-.. currentmodule:: kdiagram.plot.taylor_diagram
-
-.. autofunction:: taylor_diagram
