+++ docs/source/_autosummary/utils/kdiagram.utils.calculate_probabilistic_scores.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.utils.calculate\_probabilistic\_scores
-===============================================
-
-.. currentmodule:: kdiagram.utils
-
-.. autofunction:: calculate_probabilistic_scores
