+++ docs/source/_autosummary/evaluation/kdiagram.plot.evaluation.plot_polar_roc.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.plot.evaluation.plot\_polar\_roc
-=========================================
-
-.. currentmodule:: kdiagram.plot.evaluation
-
-.. autofunction:: plot_polar_roc
