+++ docs/source/_autosummary/evaluation/kdiagram.plot.evaluation.plot_pinball_loss.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.plot.evaluation.plot\_pinball\_loss
-============================================
-
-.. currentmodule:: kdiagram.plot.evaluation
-
-.. autofunction:: plot_pinball_loss
