+++ docs/source/_autosummary/evaluation/kdiagram.plot.evaluation.plot_polar_confusion_matrix_in.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.plot.evaluation.plot\_polar\_confusion\_matrix\_in
-===========================================================
-
-.. currentmodule:: kdiagram.plot.evaluation
-
-.. autofunction:: plot_polar_confusion_matrix_in
