+++ docs/source/_autosummary/utils/kdiagram.utils.pivot_forecasts_long.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.utils.pivot\_forecasts\_long
-=====================================
-
-.. currentmodule:: kdiagram.utils
-
-.. autofunction:: pivot_forecasts_long
