+++ docs/source/_autosummary/utils/kdiagram.utils.compute_coverage_score.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.utils.compute\_coverage\_score
-=======================================
-
-.. currentmodule:: kdiagram.utils
-
-.. autofunction:: compute_coverage_score
