+++ docs/source/_autosummary/utils/kdiagram.utils.bin_by_feature.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.utils.bin\_by\_feature
-===============================
-
-.. currentmodule:: kdiagram.utils
-
-.. autofunction:: bin_by_feature
