+++ docs/source/_autosummary/utils/kdiagram.utils.compute_winkler_score.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.utils.compute\_winkler\_score
-======================================
-
-.. currentmodule:: kdiagram.utils
-
-.. autofunction:: compute_winkler_score
