+++ docs/source/_autosummary/errors/kdiagram.plot.errors.plot_error_bands.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.plot.errors.plot\_error\_bands
-=======================================
-
-.. currentmodule:: kdiagram.plot.errors
-
-.. autofunction:: plot_error_bands
