+++ docs/source/index.rst
@@ -1,70 +1,136 @@
-.. k-diagram documentation master file, created by
-   sphinx-quickstart on Thu Apr 10 12:44:32 2025.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
 .. k-diagram documentation master file
 
-#############################################
+#########################################
 k-diagram: Polar Insights for Forecasting
-#############################################
+#########################################
+
+.. card::
+   :class-card: sd-border-0 sd-shadow-none sd-p-0 hero-image-card
+   :margin: 0
+
+   .. image:: /_static/hero_image.png
+      :alt: A collage of beautiful polar plots from k-diagram
+      :align: center
+      :width: 100%
 
 .. card:: **Navigate the complexities of forecast uncertainty and model behavior with specialized polar visualizations.**
-    :margin: 0 0 1 0
+   :class-card: sd-border-0 sd-shadow-none
+   :text-align: center
+   :margin: 4 0 1 0
 
-    Welcome to the official documentation for `k-diagram`. This package
-    provides a unique perspective on evaluating forecasting models,
-    especially when uncertainty quantification is crucial. Dive in to
-    discover how polar plots can reveal deeper insights into your
-    model's performance, stability, and potential weaknesses.
+   Welcome to `k-diagram`! This package provides a unique perspective
+   on evaluating forecasting models by leveraging the power of polar
+   coordinates. Move beyond standard metrics and discover how circular
+   plots can reveal deeper insights into your model's performance,
+   stability, and hidden weaknesses.
 
 .. container:: text-center
 
-    .. button-ref:: installation
-        :color: primary
-        :expand:
-        :outline:
+   .. button-ref:: installation
+      :color: primary
+      :expand:
+      :outline:
 
-        Install k-diagram
+      Install k-diagram
 
-    .. button-ref:: quickstart
-        :color: secondary
-        :expand:
+   .. button-ref:: quickstart
+      :color: secondary
+      :expand:
 
-        Quick Start Guide
+      Quick Start Guide
 
-    .. button-ref:: gallery/index     
-        :color: secondary
-        :expand:
-        
-        Plot Gallery
-        
-.. card:: Key Features
-    :class-header: text-center sd-font-weight-bold
-    :margin: auto  
+.. raw:: html
 
-    * **Intuitive Polar Perspective:** Visualize multi-dimensional
-      aspects like uncertainty spread, temporal drift, and spatial
-      patterns in a compact circular layout.
-    * **Advanced Error Analysis:** A dedicated suite of plots to
-      diagnose systemic bias vs. random error, compare error
-      distributions, and visualize 2D uncertainty.
-    * **Targeted Diagnostics:** Functions specifically designed to
-      assess interval coverage, consistency, anomaly magnitude, model
-      velocity, and drift.
-    * **Uncertainty-Aware Evaluation:** Move beyond point-forecast
-      accuracy and evaluate the reliability of your model's
-      uncertainty estimates.
-    * **Identify Model Weaknesses:** Pinpoint where and when your
-      forecasts are less reliable or exhibit significant anomalies.
-    * **Clear Communication:** Generate publication-ready plots to
-      effectively communicate model performance and uncertainty
-      characteristics.
+   <p align="center" style="margin-top: 2em; margin-bottom: 1.5em;">
+     <img alt="PyPI - Version" src="https://img.shields.io/pypi/v/k-diagram">
[…]

[…]
