### 2025-08-29 abd2fec
Remove auto-summary folder in docs
