+++ docs/source/_autosummary/relationship/kdiagram.plot.relationship.plot_conditional_quantiles.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.plot.relationship.plot\_conditional\_quantiles
-=======================================================
-
-.. currentmodule:: kdiagram.plot.relationship
-
-.. autofunction:: plot_conditional_quantiles
