+++ docs/source/_autosummary/probabilistic/kdiagram.plot.probabilistic.plot_polar_sharpness.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.plot.probabilistic.plot\_polar\_sharpness
-==================================================
-
-.. currentmodule:: kdiagram.plot.probabilistic
-
-.. autofunction:: plot_polar_sharpness
