+++ docs/source/_autosummary/evaluation/kdiagram.plot.evaluation.plot_regression_performance.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.plot.evaluation.plot\_regression\_performance
-======================================================
-
-.. currentmodule:: kdiagram.plot.evaluation
-
-.. autofunction:: plot_regression_performance
