+++ docs/source/_autosummary/datasets/kdiagram.datasets.make_regression_data.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.datasets.make\_regression\_data
-========================================
-
-.. currentmodule:: kdiagram.datasets
-
-.. autofunction:: make_regression_data
