+++ docs/source/_autosummary/utils/kdiagram.utils.compute_forecast_errors.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.utils.compute\_forecast\_errors
-========================================
-
-.. currentmodule:: kdiagram.utils
-
-.. autofunction:: compute_forecast_errors
