+++ docs/source/_autosummary/utils/kdiagram.utils.plot_hist_kde.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.utils.plot\_hist\_kde
-==============================
-
-.. currentmodule:: kdiagram.utils
-
-.. autofunction:: plot_hist_kde
