+++ docs/source/_autosummary/datasets/kdiagram.datasets.make_uncertainty_data.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.datasets.make\_uncertainty\_data
-=========================================
-
-.. currentmodule:: kdiagram.datasets
-
-.. autofunction:: make_uncertainty_data
