+++ docs/source/_autosummary/feature_based/kdiagram.plot.feature_based.plot_feature_interaction.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.plot.feature\_based.plot\_feature\_interaction
-=======================================================
-
-.. currentmodule:: kdiagram.plot.feature_based
-
-.. autofunction:: plot_feature_interaction
