+++ docs/source/_autosummary/datasets/kdiagram.datasets.make_multi_model_quantile_data.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.datasets.make\_multi\_model\_quantile\_data
-====================================================
-
-.. currentmodule:: kdiagram.datasets
-
-.. autofunction:: make_multi_model_quantile_data
