+++ docs/source/_autosummary/uncertainty/kdiagram.plot.uncertainty.plot_polar_quiver.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.plot.uncertainty.plot\_polar\_quiver
-=============================================
-
-.. currentmodule:: kdiagram.plot.uncertainty
-
-.. autofunction:: plot_polar_quiver
