+++ docs/source/_autosummary/relationship/kdiagram.plot.relationship.plot_residual_relationship.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.plot.relationship.plot\_residual\_relationship
-=======================================================
-
-.. currentmodule:: kdiagram.plot.relationship
-
-.. autofunction:: plot_residual_relationship
