+++ docs/source/_autosummary/uncertainty/kdiagram.plot.uncertainty.plot_actual_vs_predicted.rst
@@ -1,6 +1,0 @@
-﻿kdiagram.plot.uncertainty.plot\_actual\_vs\_predicted
-=====================================================
-
-.. currentmodule:: kdiagram.plot.uncertainty
-
-.. autofunction:: plot_actual_vs_predicted
