### 2025-08-29 46586f4
Fix RMSE compatibility in scikit-learn>=1.40
