### 2025-08-29 2cae0d9
Reformat bloc compatibility with black .
