+++ docs/source/user_guide/probabilistic.rst
@@ -49,12 +49,12 @@ Summary of Probabilistic Diagnostic Functions
    * - :func:`~kdiagram.plot.probabilistic.plot_crps_comparison`
      - Provides an overall performance score using the Continuous
        Ranked Probability Score (CRPS).
-   * - :func:`~kdiagram.plot.probabilistic.plot_credibility_bands`
-     - Visualizes how the forecast's median and credibility bands
-       change as a function of another feature.
    * - :func:`~kdiagram.plot.probabilistic.plot_calibration_sharpness`
      - Visualizes the direct trade-off between calibration and
        sharpness for multiple models.
+   * - :func:`~kdiagram.plot.probabilistic.plot_credibility_bands`
+     - Visualizes how the forecast's median and credibility bands
+       change as a function of another feature.
 
 .. raw:: html
 
@@ -113,9 +113,81 @@ In the polar plot, PIT bins map to the **angle**; frequencies map to the
 * Diagnose **overconfidence**, **underconfidence**, or **bias**.
 * Compare calibration across models before evaluating **sharpness**.
 
-**Example:**
+The first and most fundamental test of any probabilistic forecast is to
+assess its **calibration**. A forecast is considered calibrated if its
+predicted probability distributions are statistically consistent with the
+observed outcomes. In essence, it's a test of the model's statistical
+"honesty." The Probability Integral Transform (PIT) histogram is the
+primary diagnostic tool for this task. Let's go for a practical example 
+for a better understanding.
 
-See the gallery example and code: :ref:`gallery_plot_pit_histogram`.
+.. admonition:: Practical Example
+
+   Let say a energy company relies on a model to produce probabilistic
+   forecasts of the power output from a wind farm. Before using these
+   forecasts for operational planning, they must verify that the model
+   is well-calibrated. If the forecast intervals are consistently too
+   narrow (over-confident) or too wide (under-confident), any
+   decisions based on them could be costly.
+
+   The polar PIT histogram will give us an immediate visual diagnosis of
+   the forecast's calibration by comparing the distribution of PIT
+   values to a perfect uniform circle.
+
+   .. code-block:: pycon
+
+      >>> import numpy as np
+      >>> import kdiagram as kd
+      >>>
+      >>> # --- 1. Simulate a probabilistic wind power forecast ---
+      >>> np.random.seed(0)
+      >>> # True power output often follows a skewed distribution
+      >>> y_true = np.random.weibull(a=2., size=1000) * 50
+      >>> quantiles = np.linspace(0.05, 0.95, 19)
+      >>> # Simulate an OVER-CONFIDENT model (predictive intervals are too narrow)
+      >>> noise = np.random.normal(0, 5, (1000, 500)) # Underestimate the real noise
+      >>> y_preds_quantiles = np.quantile(
+      ...     y_true[:, np.newaxis] + noise, q=quantiles, axis=1
+      ... ).T
+      >>>
+      >>> # --- 2. Generate the plot ---
+      >>> ax = kd.plot_pit_histogram(
+      ...     y_true,
+      ...     y_preds_quantiles,
+      ...     quantiles=quantiles,
+      ...     title='Calibration Check for Wind Power Forecast'
+      ... )
+
+   .. figure:: ../images/userguide_plot_pit_histogram.png
+      :align: center
+      :width: 80%
+      :alt: A polar PIT histogram for a wind power forecast.
+
+      A polar PIT histogram used to diagnose the calibration of a
+      probabilistic wind power forecast. Deviations from the dashed
+      uniform circle indicate miscalibration.
+
+   This plot provides an instant diagnosis of the forecast's
[…]
