+++ docs/source/user_guide/feature_based.rst
@@ -136,9 +136,85 @@ is the number of features.
 * Normalization allows comparing relative patterns effectively, even if
   absolute importance scales differ significantly between layers.
 
-**Example:**
-(See the :ref:`Gallery <gallery_feature_based>` section for a runnable code example and plot)
 
+Understanding which features a model relies on is a cornerstone of
+interpretation and trust. While a simple bar chart can show feature
+importance for a single model, the real insights often come from
+comparing these patterns across different models or contexts. This
+"Feature Fingerprint" plot is designed for exactly that kind of
+comparative analysis.
+
+.. admonition:: Practical Example
+
+   A telecommunications company has two models competing to predict
+   customer churn: a classic ``Logistic Regression`` model and a more
+   complex ``Gradient Boosting`` model. To trust and deploy one of
+   them, the company needs to understand their decision-making
+   processes. Which features does each model consider most important?
+   Do they rely on the same information, or do they have fundamentally
+   different "views" of the problem?
+
+   This plot will create a unique "fingerprint" for each model,
+   visualizing their feature importance profiles on the same set of
+   axes for a direct comparison.
+
+   .. code-block:: pycon
+
+      >>> import numpy as np
+      >>> import kdiagram as kd
+      >>>
+      >>> # --- 1. Define feature names and model importance scores ---
+      >>> features = [
+      ...     'tenure', 'monthly_charges', 'total_charges',
+      ...     'data_usage', 'support_calls', 'contract_type'
+      ... ]
+      >>> labels = ['Logistic Regression', 'Gradient Boosting']
+      >>>
+      >>> # Logistic Regression relies heavily on a few key features
+      >>> logreg_importances = [0.8, 0.9, 0.7, 0.1, 0.2, 0.6]
+      >>> # Gradient Boosting uses a wider range of features
+      >>> boosting_importances = [0.5, 0.6, 0.6, 0.8, 0.7, 0.4]
+      >>>
+      >>> importances = np.array([logreg_importances, boosting_importances])
+      >>>
+      >>> # --- 2. Generate the plot ---
+      >>> ax = kd.plot_feature_fingerprint(
+      ...     importances,
+      ...     features=features,
+      ...     labels=labels,
+      ...     title='Churn Model Feature Importance Fingerprints'
+      ... )
+
+   .. figure:: ../images/userguide_plot_feature_fingerprint.png
+      :align: center
+      :width: 80%
+      :alt: A feature fingerprint radar chart comparing two models.
+
+      A polar radar chart comparing the feature importance profiles
+      ("fingerprints") of a Logistic Regression and a Gradient
+      Boosting model for customer churn prediction.
+
+   This plot allows for an immediate visual comparison of the models'
+   internal logic. By comparing the shapes of the colored polygons, we
+   can see which features dominate each model's decision-making.
+
+   **Quick Interpretation:**
+    The plot reveals the distinctly different "fingerprints" of the two
+    models. The ``Logistic Regression`` model (blue) has a spiky
+    profile, indicating it relies heavily on a few core features like
+    ``tenure``, ``monthly_charges``, and ``total_charges``, while paying
+    little attention to others. In contrast, the ``Gradient Boosting``
+    model (brown) displays a more well-rounded fingerprint, showing
+    that it has learned to incorporate a wider array of information,
+    assigning significant importance to features like ``data_usage`` and
+    ``support_calls`` as well.
+
+This ability to compare feature importance profiles is crucial for
+model selection, debugging, and ensuring alignment with domain
+knowledge. To see the full implementation, please explore the gallery.
+
+**Example:**
+See the gallery example and code: :ref:`gallery_plot_feature_fingerprint`.
 
[…]

[…]
