### 2025-09-01 c40d006
Rebuild gallery plots
