+++ kdiagram/utils/plot.py
@@ -2,13 +2,15 @@
 #   Author: LKouadio <etanoyau@gmail.com>
 
 import re
-from typing import Optional, Union
+from typing import Literal, Optional, Union
 
 import matplotlib.pyplot as plt
 import numpy as np
+from matplotlib.axes import Axes
 from matplotlib.colors import Colormap
 from scipy.stats import gaussian_kde
 
+from ..api.typing import Acov
 from ..compat.matplotlib import get_cmap
 from .generic_utils import get_valid_kwargs
 
@@ -692,3 +694,157 @@ def _sample_colors(
     else:
         xs = np.linspace(trim, 1.0 - trim, n)
     return [tuple(cmap_obj(float(x))) for x in xs]
+
+
+def _setup_axes_for_reliability(
+    ax: Optional[Axes],
+    counts_panel: str,
+    figsize: Optional[tuple[float, float]],
+) -> tuple:
+    """
+    Return (fig, ax, axb) given an optional `ax` and `counts_panel` setting.
+
+    - If `ax` is None:
+        * counts_panel == "bottom": new Figure + GridSpec (main + bottom)
+        * else:                      new Figure + single Axes
+    - If `ax` is provided:
+        * counts_panel == "bottom": append a bottom axis to the same figure
+                                     (using axes_grid1 divider) that shares x.
+        * else:                      reuse `ax` and no bottom axis.
+    """
+    if ax is None:
+        if counts_panel == "bottom":
+            fig = plt.figure(figsize=figsize)
+            gs = fig.add_gridspec(2, 1, height_ratios=(3.0, 1.0), hspace=0.12)
+            ax = fig.add_subplot(gs[0, 0])
+            axb = fig.add_subplot(gs[1, 0], sharex=ax)
+        else:
+            fig, ax = plt.subplots(figsize=figsize)
+            axb = None
+    else:
+        fig = ax.figure
+        if counts_panel == "bottom":
+            # Append a bottom panel below the provided `ax`
+            try:
+                from mpl_toolkits.axes_grid1 import make_axes_locatable
+            except ModuleNotFoundError as err:
+                raise RuntimeError(
+                    "counts_panel='bottom' with a provided `ax` requires "
+                    "mpl_toolkits.axes_grid1 (install matplotlib's toolkits)."
+                ) from err
+            divider = make_axes_locatable(ax)
+            # ~20% height for counts; tweak pad as you like
+            axb = divider.append_axes(
+                "bottom", size="20%", pad=0.25, sharex=ax
+            )
+        else:
+            axb = None
+    return fig, ax, axb
+
+
+def resolve_polar_span(acov: Acov = "default") -> float:
+    """Return angular span (radians) for the given coverage keyword."""
+    acov = (acov or "default").lower()
+    spans = {
+        "default": 2 * np.pi,  # 360°
+        "half_circle": 1 * np.pi,  # 180°
+        "quarter_circle": 0.5 * np.pi,  # 90°
+        "eighth_circle": 0.25 * np.pi,  # 45°
+    }
+    try:
+        return spans[acov]
+    except KeyError as e:
+        raise ValueError(
+            f"Invalid acov={acov!r}. Use one of "
+            "{'default','half_circle','quarter_circle','eighth_circle'}."
+        ) from e
+
+
+def setup_polar_axes(
+    ax: Optional[Axes],
+    *,
+    acov: Acov = "default",
+    figsize: Optional[tuple[float, float]] = None,
[…]
