### 2025-09-03 995747a
Terminate gallery revision
