+++ kdiagram/plot/feature_based.py
@@ -9,14 +9,28 @@ from typing import Any
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
+from matplotlib.axes import Axes
 
+from ..api.typing import Acov
 from ..compat.matplotlib import get_cmap
 from ..decorators import check_non_emptiness, isdf
 from ..utils.handlers import columns_manager
-from ..utils.plot import set_axis_grid
+from ..utils.plot import (
+    acov_to_span,
+    map_theta_to_span,
+    resolve_polar_axes,
+    set_axis_grid,
+    set_polar_angular_span,
+    setup_polar_axes,
+    warn_acov_preference,
+)
 from ..utils.validator import ensure_2d, exist_features
 
-__all__ = ["plot_feature_fingerprint", "plot_feature_interaction"]
+__all__ = [
+    "plot_feature_fingerprint",
+    "plot_feature_interaction",
+    "plot_fingerprint",
+]
 
 
 @check_non_emptiness(params=["df"])
@@ -31,6 +45,7 @@ def plot_feature_interaction(
     theta_period: float | None = None,
     theta_bins: int = 24,
     r_bins: int = 10,
+    acov: Acov = "default",
     title: str | None = None,
     figsize: tuple[float, float] = (8, 8),
     cmap: str = "viridis",
@@ -39,11 +54,23 @@ def plot_feature_interaction(
     mask_radius: bool = False,
     savefig: str | None = None,
     dpi: int = 300,
+    ax: Axes | None = None,
 ):
-    # --- Input Validation ---
-    required_cols = [theta_col, r_col, color_col]
-    exist_features(df, features=required_cols)
-    data = df[required_cols].dropna().copy()
+
+    warn_acov_preference(
+        acov,
+        preferred="default",
+        plot_name="feature_interaction",
+        advice=(
+            "A full 360° span usually produces the most readable comparison; "
+            "proceeding with the requested span."
+        ),
+    )
+    # ---- validate & subset
+    required = [theta_col, r_col, color_col]
+    exist_features(df, features=required)
+
+    data = df[required].dropna().copy()
     if data.empty:
         warnings.warn(
             "DataFrame is empty after dropping NaNs.",
@@ -52,67 +79,114 @@ def plot_feature_interaction(
         )
         return None
 
-    # --- Binning and Aggregation ---
-    if theta_period:
-        data["theta_rad"] = (
-            ((data[theta_col] % theta_period) / theta_period) * 2 * np.pi
+    # axes setup (acov-aware)
+    # - returns (fig, ax, span_radians)
+    fig, ax, span = setup_polar_axes(
+        ax,
+        acov=acov,
+        figsize=figsize,
+        zero_at="N",  # 0° at north by default
+        clockwise=True,  # CW increasing angles (common in polar charts)
+    )
+
+    # map theta data to [0, span]
+    th_raw = data[theta_col].to_numpy()
+    if theta_period is not None:
+        th_scaled = map_theta_to_span(
+            th_raw,
+            span=span,
+            theta_period=theta_period,
         )
     else:
-        min_theta, max_theta = data[theta_col].min(), data[theta_col].max()
-        if (max_theta - min_theta) > 1e-9:
-            data["theta_rad"] = (
-                ((data[theta_col] - min_theta) / (max_theta - min_theta))
-                * 2
-                * np.pi
-            )
-        else:
-            data["theta_rad"] = 0
+        # min-max scale to the selected span
+        tmin, tmax = float(data[theta_col].min()), float(
+            data[theta_col].max()
+        )
+        th_scaled = map_theta_to_span(
+            th_raw,
+            span=span,
+            data_min=tmin,
+            data_max=tmax,
+        )
[…]
