### 2025-09-04 f83fbde
Add technical report to support Joss-paper
