+++ docs/source/references.bib
@@ -219,12 +219,16 @@
   doi       = {10.22541/au.175390529.91420978/v1}
 }
 
-@article{kouadiob2025,
-  author = {Kouadio, Kouao Laurent and Liu, Rong and Loukou, Kouam{\'e} Gb{\`e}l{\`e} Hermann and Liu, Jianxin and Liu, Wenxiang},
-  title = {Analytics Framework for Interpreting Spatiotemporal Probabilistic Forecasts},
-  journal = {International Journal of Forecasting},
-  year = {2025},
-  note = {Manuscript submitted}
+@techreport{kouadiob2025,
+  author      = {Kouadio, Kouao Laurent},
+  title       = {{k-diagram}: Technical Report — Derivations and Details},
+  institution = {School of Geosciences and Info-physics, Central South University},
+  year        = {2025},
+  version     = {1.2.3},
+  month       = {Sep},
+  note        = {Software technical report accompanying the k-diagram package},
+  doi         = {10.5281/zenodo.17051183},
+  url         = {https://doi.org/10.5281/zenodo.17051183}
 }
 
 @book{Jolliffe2012,
