### 2025-09-04 8bda646
fix latex formatage error
