### 2025-09-04 earthai-tech/k-diagram#14 @earthai-tech
# Documentation

1. *As you mentioned, the warnings and errors ... in decorators.py, e.g*

I have now removed the redundant entry from our `conf.py` to clean up the configuration. See https://github.com/earthai-tech/k-diagram/blob/main/docs/source/conf.py . I have also corrected all the incorrect underline lengths in that ``decorators.py``,
and  run a check across the rest of the codebase to ensure consistency with numpydoc standards.

I have performed another pass on the project to fix any remaining minor formatting issues and ensure the documentation
builds as cleanly as possible  see : https://github.com/earthai-tech/k-diagram/actions/runs/17455503003/job/49568119789

# Packaging

2. *While you choose a valid and improved implementation approach, I still do ... but I was not able to find a corresponding explanation,  does this mean anything?*

I have removed the legacy `kdiagram/_version.py` file entirely.  See the new [pyproject.toml](https://github.com/earthai-tech/k-diagram/blob/main/pyproject.toml)

3. *Furthermore, in docs/source/conf.py:24-52 you are trying to import the ... try block is enough for everything to work.*

I simplified `docs/source/conf.py` by using a single  `try: import kdiagram` … `except:` fallback for the version
and dropped reading `pyproject.toml`. I also kept a minimal fallback (`release = "0+unknown"`) only when building docs without the package installed.  See https://github.com/earthai-tech/k-diagram/blob/main/docs/source/conf.py .

4. *While I complimented you on the API, you respond "Action taken…" What action was taken?*

On AI assistance & responsibility: I occasionally used an AI tool to prototype tests (e.g., edge-case scaffolding). It did not generate
library code or APIs. All implementation and final test code were written and reviewed by me, and I take full responsibility for the
codebase, including this response. The spatiotemporal probabilistic forecasting methods and their implementation reflect my own work and domain deep understanding.

5. *SyntaxWarnings… If you want to be sure your documentation is rendering correctly, please fix them. If you do not see them, do a fresh sphinx build. *
I have fixed all Sphinx issues  in read, on a **fresh build** with ``dirbuild``, and the build successfully passed, as you can see in the image screenshot below

  ```bash
rm -rf docs/_build && sphinx-build -b dirhtml docs/source docs/_build/html
  ```

# CI

3. *Test more current versions: ✔️ … test until 3.12. … exclude 3.13, or test for it.*

I also added **Python 3.13**  to the GitHub Actions test matrix, and trove classifiers in `pyproject.toml`.

4. *Lint: … ruff format results in 27 files reformatted … In your CI I see some commented out code `ruff --fix .` which is not supported by ruff.*

You are right, the correct  usage is  `ruff check --fix .`, but I only use auto-fix **locally**
via pre-commit, not in CI. CI now verifies style without modifying files and the clean the section:

```yaml
- name: Lint (ruff and black)
  run: |
    echo "Checking for linter issues with Ruff..."
    ruff check .  # Checks for linting issues

```
Please refer to Actions configuration to see the latest change : https://github.com/earthai-tech/k-diagram/blob/main/.github/workflows/python-package-conda.yml

- *Looking into the CI… "If you switch to Ruff:" … Copy pasted too much?*

I have removed that leftover comment. The workflow now reflects the current minimal Ruff setup to check linting issue only.

- *In the Lint step you use pip, `python -m pip install -e .[dev]`. I suppose  this is not what you want to do, as all dependencies should already be satisfied.*

I dropped `pip install -e .[dev]`  from the lint job. The conda env already  contains the dev toolchain; we only `pip install . --no-deps` once when needed for imports during docs/tests.

# Development / dependency harmonization

*README… environment.yml is aligned now. I still see  differences between dependencies in environment.yml and pyproject.toml…
“Misc / build helpers”… “Optional extras”… tqdm not included… “Pip-only packages” available on conda-forge… docs/requirements.txt still has pytest and flake8… three different NumPy pins… keep dependencies in sync…*

I have made sure **Aligned** dependency sets across  `pyproject.toml`, `environment.yml`, and `docs/requirements.txt`:

  * **NumPy pin:** unified to `numpy>=1.22` in all three places.
  * **Extras:** introduced `.[docs]` (Sphinx, Furo, extensions, `numpydoc`, `sphinxcontrib-bibtex`);  https://github.com/earthai-tech/k-diagram/blob/main/pyproject.toml
  `.[dev]` now **includes** `.[docs]` plus tests & linters.
  * **`docs/requirements.txt`:** limited strictly to docs build (Sphinx + extensions + `numpydoc`).  See https://github.com/earthai-tech/k-diagram/blob/main/docs/requirements.txt
  *  **Removed** `pytest` and `flake8`.
  * **`environment.yml`:** installs runtime + tests + linters + docs via `conda-forge` when available; kept a minimal `pip` section only for packages not on conda-forge. See https://github.com/earthai-tech/k-diagram/blob/main/environment.yml
  * **Build helpers:** removed redundant `setuptools`/`wheel`  from the conda env.
  * **`tqdm`:** moved to `.[dev]`.

I have ** Updated **README** and installation docs and added a “Build docs in isolation” section   (using `.[docs]` or `docs/requirements.txt`) with `sphinx-build` HTML and LaTeX/PDF steps in our documentation :https://k-diagram.readthedocs.io/en/latest/installation.html#building-documentation  which provided both `venv` and `conda` workflows, with `dirhtml/html`  build commands and clean rebuild instructions.

3. warnings issues

Thank you for your detailed follow-up and for pushing for more rigor  regarding the test warnings. I appreciate your diligence. You were absolutely  right to be concerned about the use of `filterwarnings` in `pyproject.toml`,  as globally silencing warnings can indeed hide underlying issues.

I have fixed the warnings from doc builds and packages tests after removing the `filterwarnings` from `pyproject.toml` entirely.
You can see that there are no warnings issued from  the tests with Python 3.9 (https://github.com/earthai-tech/k-diagram/actions/runs/17455503003/job/49568119768)   to  3.12 (https://github.com/earthai-tech/k-diagram/actions/runs/17455503003/job/49568119789). Only Python  3.13,  (https://github.com/earthai-tech/k-diagram/actions/runs/17455503003/job/49568119763) shows two warnings about the Python Resource Package.

I have gone through the entire test suite and  **removed all `@pytest.mark.skip` markers**.
The issues that previously required them (like version-dependent warnings or minor cache problems) have now been fully resolved.

4. *kdiagram/tests/test_uncertainty.py:241..*

In fact,  the test is `tests/test_uncertainty_plots.py` (rather than `test_uncertainty.py` ) was failing in newer Python
environments. This has been fixed by making the test conditional on the NumPy  version, ensuring it passes on all supported versions from Python 3.9 to 3.13. I can confirm that the entire test suite now passes without any skips. You
can see the successful build log on the following commit here, which shows  **0 skipped tests,  935 passed in 58.19s**:  https://github.com/earthai-tech/k-diagram/actions/runs/17455503003/job/49568119789

5. *What does the compat/numpy.py file do in the case of the package? Does it bring the numpy 2.x api when having installed 1.x? Is there functionality that is only in numpy 1.x and not in 2.x that k-diagram uses?*

The `kdiagram/compat/numpy.py` file is a **forward-compatibility shim**. Its purpose is to ensure that `k-diagram`
works smoothly with both older (1.x) and newer (2.x) versions of NumPy. My strategy is to ensure  `k-diagram` is ready for NumPy 2.0. To manage this transition, I created a compatibility module (`kdiagram/compat/numpy.py`) to handle version differences.
For example, `np.find_common_type` was deprecated in NumPy 1.25 and removed in 2.0. ``ndarray. ptp``  is now removed in numpy<2 and kept only ``np.ptp``.  Our compatibility file checks the NumPy version and provides a stable function that uses the old one on NumPy 1.x and the new, recommended one on NumPy 2.x.

# Reply to Small Notes:

"The if ``__name__ == "__main__":``  blocks were a remnant of my personal development workflow in the Spyder IDE, where it's a convenient way to run a single file. However, you are correct that  it is not a standard pattern for a pytest suite and is unnecessary. I have removed this boilerplate from all test files to adhere to best practices and improve clarity for all contributors."

https://github.com/earthai-tech/k-diagram/issues/14#issuecomment-3191703275
