### 2025-09-04 e7e271e
updated paper.bib
