+++ paper.bib
@@ -64,7 +64,8 @@
   journal   = {Journal of Machine Learning Research},
   volume    = {12},
   pages     = {2825-2830},
-  year      = {2011}
+  year      = {2011}, 
+  url       = {https://scikit-learn.org/stable/}
 }
 
 @article{Liu2024,
