### 2025-09-04 53f9409
Upgrade to release 1.2.3
