+++ docs/source/release_notes/index.rst
@@ -13,12 +13,12 @@ releases. We follow semantic versioning (MAJOR.MINOR.PATCH). For upgrade
 guidance, breaking changes are called out in each version page.
 
 .. note::
-   The latest stable release is **v1.2.2**.
+   The latest stable release is **v1.2.3**.
 
 Quick links
 -----------
 
-- :doc:`v1.2.2 <v1.2.2>` — New contextual plots and data processing utilities.
+- :doc:`v1.2.3 <v1.2.3>` — New contextual plots and data processing utilities.
 - :doc:`v1.2.0 <v1.2.0>` — New error analysis module and advanced uncertainty plots.
 - :doc:`v1.1.0 <v1.1.0>` — New features, improvements, and fixes.
 - :doc:`v1.0.3 <v1.0.3>` — Maintenance and bug fixes on the 1.0 line.
@@ -31,7 +31,7 @@ Contents
    :maxdepth: 1
    :caption: Versions
 
-   v1.2.2
+   v1.2.3
    v1.2.0
    v1.1.0
    v1.0.3
