### 2025-09-05 dd50393
Update Joss fig2 paper.md
