### 2025-09-07 5e49cfe
Update Joss paper: Functionality section
