### 2025-09-07 65aee4c
Update paper.md: Functionality section
