### 2025-09-19 #8661 @jromanowska
@cbueth, I've read the issues you were refering to. It seems like you did a great work with testing, reading the documentation and also the code of k-diagram. Good job! Thank you, @earthai-tech, for the effort to adapt the code according to the reviewers comments. It's especially important to be able to install the package easily, have well designed CI actions, and a code that can be easily maintained. I agree that the amount of warnings when running the tests was too much and that simply silencing them is not a good way to solve the problem. On the other hand, I think @earthai-tech is doing a good job in improving the software. Let's wait for the new round of changes :)
