### 2025-09-19 #8661 @earthai-tech
> Agreed, thanks @jromanowska and @earthai-tech !

I guess the thanks belong to you all because I'm still continue to learn! See you to the next round!
