### 2025-09-25 41b167c
Add anomaly plots
