+++ tests/test_datasets_make_new.py
@@ -1,0 +1,360 @@
+import numpy as np
+import pandas as pd
+import pytest
+
+from kdiagram.datasets.make import (
+    make_classification_data,
+    make_cyclical_data,
+    make_fingerprint_data,
+    make_multi_model_quantile_data,
+    make_regression_data,
+    make_taylor_data,
+    make_uncertainty_data,
+)
+
+# ------------------------------- make_cyclical_data ---------------------------
+
+
+def test_make_cyclical_basic_and_errors():
+    # scalar params should broadcast; as_frame=True order/shape
+    df = make_cyclical_data(
+        n_samples=24,
+        n_series=3,
+        pred_phase_shift=0.1,
+        pred_amplitude_factor=0.9,
+        pred_noise_factor=1.2,
+        pred_bias=0.0,
+        as_frame=True,
+        seed=123,
+    )
+    assert list(df.columns[:2]) == ["y_true", "time_step"]
+    assert df.shape == (24, 2 + 3)
+
+    # Bunch path, seed=None branch
+    b = make_cyclical_data(
+        n_samples=10, n_series=2, seed=None, as_frame=False
+    )
+    assert set(
+        [
+            "frame",
+            "data",
+            "feature_names",
+            "target_names",
+            "target",
+            "series_names",
+            "prediction_columns",
+            "DESCR",
+        ]
+    ).issubset(b.keys())
+    assert b.data.shape == (10, 1 + 2)  # time_step + two series
+
+    # wrong series_names length -> ValueError
+    with pytest.raises(ValueError):
+        make_cyclical_data(
+            n_series=2, series_names=["only_one"], as_frame=True
+        )
+
+    # bad param type -> TypeError
+    with pytest.raises(TypeError):
+        make_cyclical_data(n_series=2, pred_bias={"oops": 1}, as_frame=True)
+
+
+# ------------------------------ make_fingerprint_data ------------------------
+
+
+def test_make_fingerprint_variants_and_validation():
+    # invalid sparsity
+    with pytest.raises(ValueError):
+        make_fingerprint_data(sparsity=-0.1, as_frame=True)
+    # invalid value_range
+    with pytest.raises(ValueError):
+        make_fingerprint_data(value_range=(2.0, -1.0), as_frame=True)
+    # feature names length mismatch
+    with pytest.raises(ValueError):
+        make_fingerprint_data(
+            n_features=4, feature_names=["a", "b"], as_frame=True
+        )
+    # layer names length mismatch
+    with pytest.raises(ValueError):
+        make_fingerprint_data(n_layers=2, layer_names=["L1"], as_frame=True)
+
+    # as_frame=True with sparsity > 0 => expect some zeros
+    df = make_fingerprint_data(
+        n_layers=4, n_features=7, sparsity=0.25, seed=1, as_frame=True
+    )
+    assert df.shape == (4, 7)
+    assert (df.values == 0).sum() > 0
+
+    # Bunch path, add_structure toggled
+    b = make_fingerprint_data(
+        n_layers=3, n_features=5, add_structure=False, seed=2
+    )
+    assert b.importances.shape == (3, 5)
+    assert isinstance(b.frame, pd.DataFrame)
+
+
+# ------------------------------ make_uncertainty_data ------------------------
+
+
+def test_make_uncertainty_anomalies_and_layout():
+    ds = make_uncertainty_data(
+        n_samples=50, n_periods=3, anomaly_frac=0.2, seed=7
+    )  # Bunch path
+    assert (
+        len(ds.q10_cols) == 3
+        and len(ds.q90_cols) == 3
+        and len(ds.q50_cols) == 3
+    )
+
+    # anomaly count equals int(frac * n_samples)
[…]
