### 2025-09-25 6a13df3
Drop comma before the citations in doc
