### 2025-09-25 e13b255
Make pyproject the single source of truth
