+++ environment.yml
@@ -7,41 +7,39 @@ channels:
   - defaults
 
 dependencies:
-  # Core Python
   - python>=3.9
-
-  # Runtime dependencies (from pyproject [project].dependencies)
-  - numpydoc>=1.6
+  # Runtime via conda (matches pyproject lower bounds)
   - numpy>=1.22
   - pandas
   - scipy
   - matplotlib
   - seaborn
   - scikit-learn
+  - numpydoc>=1.6
 
-  # Testing / coverage dependencies
+  # Test/dev via conda
   - pytest>=8.0
   - pytest-xdist
   - pytest-cov
   - coverage
-
-  # Dev utilities (match pyproject dev extras)
   - ruff>=0.4
   - black>=24.0
-
-  # Optional extras (like tqdm)
   - tqdm
 
-  # Docs toolchain (from docs requirements + pyproject dev)
-  # Pip-only packages (especially Sphinx extensions)
   - pip:
-    - pytest-mock
+    # Project install without deps because conda provided runtime
+    - -e . --no-deps
+    # Docs toolchain (same as [project.optional-dependencies].docs)
     - sphinx>=7.0,<8.0
     - furo
     - sphinx-copybutton
     - myst-parser
     - sphinx-design
     - sphinx-gallery
+    - sphinxcontrib-jsmath
     - sphinx-mathjax-offline
     - sphinx-issues
     - sphinx-prompt
+    - sphinxcontrib-bibtex
+    - pytest-mock
+
