### 2025-09-25 f7cfa83
Update the installation page/README.md after split the optional dependencies in [docs] and [dev] in pyproject.toml
