### 2025-09-25 e97fa8b
Uniformize CI & docs requirements/environment.yml to rely only on pyproject.toml
