+++ docs/requirements.txt
@@ -1,24 +1,3 @@
 # Sphinx and Core Extensions
-sphinx>=7.0,<8.0
-sphinx-copybutton
-myst-parser
-sphinx-design
-furo
-sphinx-gallery
-sphinxcontrib-jsmath
-sphinx-mathjax-offline
-sphinx-issues
-sphinx-prompt
-sphinxcontrib-bibtex
-numpydoc>=1.6
-
-# k-diagram Core Dependencies
-numpy>=1.22
-pandas>=1.0
-scipy>=1.5
-matplotlib>=3.3
-seaborn>=0.11
-scikit-learn>=0.24
-
-# Testing and development
-pytest-mock
+# Install k-diagram with its docs extra from pyproject.toml
+-e ..[docs]
