+++ pyproject.toml
@@ -78,9 +78,6 @@ dev = [
   "tqdm",
 ]
 
-
-# Testing and development
-pytest-mock
 [tool.setuptools.packages.find]
 include = ["kdiagram*"]
 
