### 2025-09-25 619525a
Remove standard alone pytest-mock and fix deprecated resources.is_resource
