+++ tests/test_datasets__property.py
@@ -1,6 +1,6 @@
 import shutil
 from pathlib import Path
-
+from contextlib import contextmanager
 import pytest
 
 from kdiagram.datasets import _property as prop
@@ -56,20 +56,35 @@ def test_remove_data_deletes_dir(tmp_path):
 
 # ---------- download_file_if : package resource path ----------
 
-
 def test_download_from_package_copies_to_cache(tmp_path, monkeypatch):
     cache_dir = tmp_path / "cache"
-    filename = "sample.txt"
-    pkg_file = tmp_path / "pkg" / filename
-    _touch(pkg_file, b"pkgdata")
+    cache_dir.mkdir(parents=True, exist_ok=True)
 
-    # Mock importlib.resources APIs used by the module
-    monkeypatch.setattr(
-        prop.resources, "is_resource", lambda mod, name: name == filename
-    )
-    monkeypatch.setattr(
-        prop.resources, "path", lambda mod, name: _PathCtx(pkg_file)
-    )
+    filename = "sample.txt"
+    pkg_dir = tmp_path / "pkg"
+    pkg_dir.mkdir(parents=True, exist_ok=True)
+    pkg_file = pkg_dir / filename
+    pkg_file.write_bytes(b"pkgdata")
+
+    # --- Fake Traversable objects for the modern resources API ---
+    class _Candidate:
+        def __init__(self, path: Path): self._path = path
+        def is_file(self): return True  # signal that the resource exists
+
+    class _Root:
+        def __init__(self, base: Path): self._base = base
+        def joinpath(self, name: str):
+            assert name == filename
+            return _Candidate(self._base / name)
+
+    # files(<pkg>) -> our fake root
+    monkeypatch.setattr(prop.resources, "files", lambda pkg: _Root(pkg_dir))
+
+    # as_file(candidate) -> yield the real path to pkg_file
+    @contextmanager
+    def _as_file(traversable):
+        yield pkg_file
+    monkeypatch.setattr(prop.resources, "as_file", _as_file)
 
     # Ensure downloader is never called in this path
     monkeypatch.setattr(
@@ -86,8 +101,8 @@ def test_download_from_package_copies_to_cache(tmp_path, monkeypatch):
     )
     assert out is not None
     out_path = Path(out)
-    assert out_path.parent == cache_dir
-    assert out_path.read_bytes() == b"pkgdata"
+    assert out_path.parent == cache_dir              # copied into cache
+    assert out_path.read_bytes() == b"pkgdata"       # content preserved
 
 
 def test_download_uses_cache_if_already_present(tmp_path, monkeypatch):
@@ -220,7 +235,6 @@ def test_download_if_missing_disabled_returns_none(tmp_path, monkeypatch):
 
 # ---------- error handling & metadata validation ----------
 
-
 def test_invalid_error_value_raises(tmp_path):
     with pytest.raises(ValueError):
         prop.download_file_if("x.txt", error="nope", verbose=False)
@@ -283,27 +297,41 @@ def test_download_returns_none_when_downloader_silent_failure(
 
 
 def test_package_copy_failure_falls_back_to_package_path_or_cache(
-    tmp_path, monkeypatch
-):
+        tmp_path, monkeypatch
+    ):
     cache_dir = tmp_path / "cache"
+    cache_dir.mkdir(parents=True, exist_ok=True)
+
     filename = "copyfail.txt"
-    pkg_file = tmp_path / "pkg2" / filename
-    _touch(pkg_file, b"pkg2")
+    pkg_dir = tmp_path / "pkg2"
+    pkg_dir.mkdir(parents=True, exist_ok=True)
+    pkg_file = pkg_dir / filename
[…]
