+++ tests/test_plot_context_new.py
@@ -12,7 +12,12 @@ from kdiagram.plot import context as ctx
 
 def make_df(n=60, seed=0, with_nans=False):
     rng = np.random.default_rng(seed)
-    t = pd.date_range("2024-01-01", periods=n, freq="h")
+    t = pd.to_datetime(
+        np.arange(n),
+        unit="h",
+        origin=pd.Timestamp("2024-01-01"),
+    )
+    
     y = 50 + np.linspace(0, 10, n) + 5 * np.sin(np.arange(n) * 2 * np.pi / 15)
     pred1 = y + rng.normal(0, 1.5, n)
     pred2 = 0.9 * y + 5 + rng.normal(0, 2.0, n)
