+++ tests/test_datasets_.py
@@ -1,9 +1,9 @@
-import os
-import re
+
 
 import numpy as np
 import pandas as pd
 import pytest
+from importlib import resources
 
 from kdiagram.api.bunch import Bunch
 from kdiagram.datasets.load import (
@@ -12,34 +12,58 @@ from kdiagram.datasets.load import (
 )
 
 
+# @pytest.fixture
+# def mock_load_zhongshan_subsidence(mocker):
+#     # Mock get_data to return the correct directory
+#     mocker.patch(
+#         "kdiagram.datasets.load.get_data",
+#         return_value=os.path.join(
+#             os.path.dirname(__file__), "kdiagram", "datasets", "data"
+#         ),
+#     )
+
+#     # Mock the file download path
+#     mocker.patch(
+#         "kdiagram.datasets.load.download_file_if",
+#         return_value=os.path.join(
+#             os.path.dirname(__file__),
+#             "kdiagram",
+#             "datasets",
+#             "data",
+#             "min_zhongshan.csv",
+#         ),
+#     )
+
+#     # Mock the package resource path
+#     mocker.patch(
+#         "kdiagram.datasets.load.resources.files",
+#         return_value=os.path.join(
+#             os.path.dirname(__file__), "kdiagram", "datasets", "data"
+#         ),
+#     )
+
 @pytest.fixture
 def mock_load_zhongshan_subsidence(mocker):
-    # Mock get_data to return the correct directory
+    # Use the actual packaged data directory
+    pkg_data_dir = resources.files("kdiagram.datasets.data")
+
+    # Point the cache to the packaged data dir (or use tmp_path if you prefer)
     mocker.patch(
         "kdiagram.datasets.load.get_data",
-        return_value=os.path.join(
-            os.path.dirname(__file__), "kdiagram", "datasets", "data"
-        ),
+        return_value=str(pkg_data_dir),
     )
 
-    # Mock the file download path
+    # Mock the download helper to "return" the packaged file path
     mocker.patch(
         "kdiagram.datasets.load.download_file_if",
-        return_value=os.path.join(
-            os.path.dirname(__file__),
-            "kdiagram",
-            "datasets",
-            "data",
-            "min_zhongshan.csv",
-        ),
+        return_value=str(pkg_data_dir / "min_zhongshan.csv"),
     )
 
-    # Mock the package resource path
+    # IMPORTANT: do NOT mock resources.files to a raw string.
+    # If you *really* want to mock it, return a Traversable:
     mocker.patch(
         "kdiagram.datasets.load.resources.files",
-        return_value=os.path.join(
-            os.path.dirname(__file__), "kdiagram", "datasets", "data"
-        ),
+        return_value=pkg_data_dir,  # this is a Traversable
     )
 
 
@@ -132,36 +156,32 @@ def test_load_uncertainty_data_invalid_parameters():
         # Invalid anomaly fraction (> 1)
         load_uncertainty_data(anomaly_frac=1.5, as_frame=True)
 
-
 @pytest.mark.network
 def test_load_zhongshan_subsidence_force_download(
-    mock_load_zhongshan_subsidence,
-):
-    """
-    Force-download the Zhongshan subsidence dataset
-    and check for expected warnings.
-    """
-    try:
-        # Use the context manager to capture all warnings that occur
-        with pytest.warns(UserWarning) as record:
-            result = load_zhongshan_subsidence(
-                force_download=True, as_frame=True
-            )
+        mocker, mock_load_zhongshan_subsidence
+    ):
+    # Simulate a failed download attempt
+    mocker.patch(
+        "kdiagram.datasets.load.download_file_if",
+        return_value=None,  # forces load() to warn "Forced download failed ..."
+    )
 
-            # If the function succeeds, it must have found a fallback resource.
[…]
