### 2025-09-25 1b884cb
Add requests to dev in pyproject.toml
