+++ pyproject.toml
@@ -77,6 +77,7 @@ dev = [
   "coverage[toml]>=7.0",
   "tqdm",
   "statsmodels",
+  "requests"
 ]
 
 [tool.setuptools.packages.find]
