+++ docs/requirements.txt
@@ -1,3 +1,3 @@
 # Sphinx and Core Extensions
 # Install k-diagram with its docs extra from pyproject.toml
--e ..[docs]
+-e .[docs]
