### 2025-09-25 f7ce79d
Make kdiagram to rely only pyproject.Done
