+++ pyproject.toml
@@ -104,9 +104,8 @@ omit = [
     "kdiagram/__init__.py",             # Only re-exports the public API
     "kdiagram/__main__.py", 
     "kdiagram/datasets/_property.py",   # Internal dataset helpers
-    
-    # Difficult to unit test reliably 
-    "kdiagram/cli/_utils.py",          # Tested via integration, not unit coverage
+                                        # Difficult to unit test reliably 
+    "kdiagram/cli/_utils.py",           # Tested via integration, not unit coverage
     "kdiagram/cli/plot_context_corr.py",
     "kdiagram/cli/plot_context_err.py", 
     "kdiagram/core/_io_utils.py",
@@ -115,7 +114,7 @@ omit = [
     "kdiagram/compat/*",                # Handles external library version differences
     "kdiagram/utils/deps.py",           # Handles optional dependencies
     "kdiagram/utils/_deps.py",          # Handles optional dependencies
-    "examples/cas/scripts/*.py"         # skip CAS examples for now
+    "examples/cas/scripts/*.py"         # skip CAS examples for now: will add it in future release
 ]
 
 [tool.coverage.report]
@@ -125,6 +124,8 @@ precision = 1
 
 omit = [
   "kdiagram/cli/_utils.py",
+  "kdiagram/cli/plot_context_corr.py",
+  "kdiagram/cli/plot_context_err.py", 
   "kdiagram/core/_io_utils.py",
   "kdiagram/core/io.py",
   "kdiagram/utils/plot.py",
