### 2025-09-26 16c3bcc
Add CAS data and examples for paper reproducibility
