+++ examples/cas/scripts/preprocessing_cas_data.py
@@ -1,0 +1,728 @@
+# License: Apache 2.0
+# Author: LKouadio <etanoyau@gmail.com>
+
+"""
+Build supervised CAS datasets from cleaned domain tables.
+
+This module converts domain-cleaned CSVs (or in-memory DataFrames)
+into time-indexed, feature-engineered tables ready for modeling.
+It:
+  * enforces a target sampling frequency (optional),
+  * adds calendar/time features and lag/rolling statistics,
+  * creates multi-horizon supervised targets (long format),
+  * performs time-ordered splits per series (train/val/test),
+  * writes outputs (Parquet preferred, CSV.GZ fallback),
+  * and emits a manifest JSON describing the run.
+
+Inputs are cleaned per-domain tables with minimal schema:
+``domain``, ``series_id``, ``t`` (datetime-parseable), ``y`` (numeric).
+Typical CAS layout:
+
+    data/cas/
+      raw/                 # raw inputs (from data sources)
+      preprocessed/        # cleaned + supervised outputs (this module)
+      modeling_results_ok/ # model predictions + metrics
+      outputs/             # paper figures/tables
+
+Main entry points
+-----------------
+- preprocess_dataframe(df, ...): core transform on a DataFrame.
+- preprocess_cas_file(path, outdir, ...): CSV → outputs + manifest.
+- preprocess_{wind,hydro,subsidence}(...): thin wrappers.
+
+See the CLI at the bottom for reproducible runs from the shell.
+"""
+
+from __future__ import annotations
+
+import json
+from pathlib import Path
+from typing import Any
+
+import numpy as np
+import pandas as pd
+
+# ---------------------------- defaults ---------------------------------
+
+DEFAULT_FREQ = "D"
+DEFAULT_HORIZONS = [1, 3, 7, 14, 28]
+DEFAULT_LAGS = [1, 2, 3, 7, 14, 28]
+DEFAULT_ROLLS = [7, 14, 30]
+DEFAULT_VAL = 0.15
+DEFAULT_TEST = 0.15
+DEFAULT_SEED = 13
+
+
+# ---------------------------- utilities --------------------------------
+
+
+def _safe_to_datetime(
+    s: pd.Series,
+    dayfirst: bool | None = None,
+) -> pd.Series:
+    # strip whitespace + common invisible marks
+    s = s.astype(str)
+    s = s.str.replace("\u200f", "", regex=False)
+    s = s.str.strip()
+
+    tries: list[dict[str, Any]] = []
+    if dayfirst is not None:
+        tries = [dict(dayfirst=dayfirst, errors="coerce")]
+    else:
+        tries = [
+            dict(dayfirst=False, errors="coerce"),
+            dict(dayfirst=True, errors="coerce"),
+        ]
+    out = None
+    for kw in tries:
+        out = pd.to_datetime(s, **kw)  # type: ignore
+        if out.notna().mean() > 0.8:
+            break
+    if out is None:
+        out = pd.to_datetime(s, errors="coerce")  # type: ignore
+    return out
+
+
+def _enforce_freq(
+    df: pd.DataFrame,
+    freq: str | None,
+) -> pd.DataFrame:
+    # Resample per (domain, series_id) if freq is set
+    if not freq:
+        return df
+    parts: list[pd.DataFrame] = []
+    for (d, sid), g in df.groupby(["domain", "series_id"], sort=False):
+        g = g.set_index("t").sort_index()
+        full = g.resample(freq).asfreq()
+        for c in g.columns:
+            full[c] = full[c].astype(g[c].dtype) if c in full else np.nan
+        nums = g.select_dtypes(include=[np.number]).columns
+        covars = [c for c in nums if c != "y"]
+        full[covars] = full[covars].ffill()
+        full["domain"] = d
[…]
