+++ examples/cas/scripts/results_R3.py
@@ -1,0 +1,345 @@
+"""
+R3: Per-horizon analysis — Table 3 + Figure 4
+
+Reads (via results_config):
+  metrics_all_domains.csv
+    cols: domain, model, horizon, n, coverage, delta_cov,
+          winkler, crps, cas
+
+Writes (under data/cas/outputs/):
+  table3_per_horizon_winners.(csv|tex)
+  figure4a_cas_vs_horizon.(png|pdf)
+  figure4b_cas_vs_horizon_log.(png|pdf)
+"""
+
+from __future__ import annotations
+
+import matplotlib.pyplot as plt
+import matplotlib.ticker as mtick
+import numpy as np
+import pandas as pd
+from matplotlib.lines import Line2D
+from results_config import (
+    DOMAIN_COLOR,
+    DOMAIN_LABEL,
+    DOMAIN_ORDER,
+    METRICS_PATH,
+    MODEL_LABEL,
+    MODEL_MARK,
+    MODEL_ORDER,
+    OUTDIR,
+)
+
+# ----------------------------
+# Config & aesthetics
+# ----------------------------
+HORIZONS = [1, 3, 7, 14, 28]
+
+# neutral line styles per model
+MODEL_STYLE = {"qar": "-", "qgbm": "--", "xtft": "-."}
+
+plt.rcParams.update(
+    {
+        "figure.dpi": 120,
+        "savefig.dpi": 400,
+        "font.size": 12,
+        "axes.labelsize": 12,
+        "axes.titlesize": 13,
+        "legend.fontsize": 10,
+        "xtick.labelsize": 11,
+        "ytick.labelsize": 11,
+        "axes.linewidth": 0.8,
+        "grid.alpha": 0.25,
+        "grid.linestyle": "--",
+        "grid.linewidth": 0.6,
+        "lines.linewidth": 2.0,
+        "figure.constrained_layout.use": True,
+        "font.family": "serif",
+        "font.serif": [
+            "DejaVu Serif",
+            "Times New Roman",
+            "Times",
+            "Computer Modern Roman",
+        ],
+    }
+)
+
+OUTDIR.mkdir(parents=True, exist_ok=True)
+
+# ----------------------------
+# Load & prepare
+# ----------------------------
+met = pd.read_csv(METRICS_PATH)
+met["abs_cas"] = met["cas"].abs()
+
+# ----------------------------
+# Table 3: winners by min |CAS|
+# per (domain, horizon)
+# ----------------------------
+w_idx = met.groupby(["domain", "horizon"])["abs_cas"].idxmin()
+
+winners = (
+    met.loc[
+        w_idx,
+        [
+            "domain",
+            "horizon",
+            "model",
+            "abs_cas",
+            "cas",
+            "coverage",
+            "delta_cov",
+            "crps",
+            "winkler",
+        ],
+    ]
+    .sort_values(["domain", "horizon"])
+    .reset_index(drop=True)
+)
+
+csv_path = OUTDIR / "table3_per_horizon_winners.csv"
+winners.to_csv(csv_path, index=False)
+
+
+def winners_to_latex(df: pd.DataFrame) -> str:
+    df2 = df.copy()
+    df2["model"] = (
+        df2["model"].map(MODEL_LABEL).apply(lambda s: f"\\textbf{{{s}}}")
+    )
+    return df2.to_latex(
+        index=False,
+        float_format="%.4g",
+        columns=[
+            "domain",
+            "horizon",
+            "model",
+            "abs_cas",
+            "cas",
+            "coverage",
+            "delta_cov",
+            "crps",
[…]
