### 2025-09-26 #8661 @earthai-tech
Hello @cbueth ,

I'm following up to let you know that I have completed the revisions based on your feedback. The **paper revision** is detailed in the comment at https://github.com/earthai-tech/k-diagram/issues/22#issuecomment-3337023483, and the **updated documentation** is referenced here https://github.com/openjournals/joss-reviews/issues/8661#issuecomment-3336874389.
Similarly, the requested **software revisions** with the changes are outlined in this issue: https://github.com/earthai-tech/k-diagram/issues/14#issuecomment-3337742616.

For full transparency, the **JOSS disclosure** for related works was also posted earlier in this review thread.

Thank you again for your time and valuable guidance throughout this process. I believe all points have been addressed, and I look forward to your feedback.
