### 2025-10-01 6495467
expand kernel explanation in CAS metric
