### 2025-10-27 earthai-tech/k-diagram#14 @cbueth
In the `pyproject.toml` in section `[tool.coverage.run]` there are several files ignored in coverage with various reasons. This section should be used for files that should be omitted in the coverage calculation, e.g. files that are tested in other ways files that are tested but they are not counted by a technical reason, e.g. jit compiled functionality.
You should not add functionality that is hard to test, as the coverage number should resemble what is actually being covered. Furthermore, you might want to add tests for core functionality, as `kdiagram/core/io.py` is hinting by it's name.
Please reconsider what you keep on that list and with which reason.
