### 2025-10-30 8bafe26
Ruff formattage done
