+++ tests/test_property_handlers.py
@@ -1,6 +1,7 @@
-
 import pandas as pd
+
 from kdiagram.core.property import PandasDataHandlers
+
 
 def test_parsers_and_writers_presence():
     h = PandasDataHandlers()
