### 2025-10-30 71cf8d2
Update the paper.md and add kind param to experimental evaluation
