+++ kdiagram/plot/evaluation.py
@@ -27,8 +27,10 @@ from ..utils.handlers import columns_manager
 from ..utils.mathext import compute_pinball_loss
 from ..utils.plot import (
     canonical_acov,
+    maybe_delegate_cartesian,
     set_axis_grid,
     setup_polar_axes,
+    validate_kind,
     warn_acov_preference,
 )
 from ..utils.validator import validate_yy
@@ -62,6 +64,7 @@ def plot_polar_roc(
     show_auc: bool = True,
     savefig: str | None = None,
     dpi: int = 300,
+    kind: str = "polar",
     ax: Axes | None = None,
 ):
     if not y_preds:
@@ -80,28 +83,35 @@ def plot_polar_roc(
 
     y_true, _ = validate_yy(y_true, y_preds[0])  # Validate first pred
 
-    # Respect parameter but force quarter-circle
-    canon = canonical_acov(acov, raise_on_invalid=False)
-    if canon != "quarter_circle":
-        warnings.warn(
-            "plot_polar_roc currently renders best as a quarter circle. "
-            f"Received acov='{acov}'. For ROC, θ=0 at the right (East) and a "
-            "90° span give the clearest reading of FPR (angle) vs TPR (radius). "
-            "Proceeding with acov='quarter_circle'.",
-            UserWarning,
-            stacklevel=2,
-        )
-
+    # Branch on rendering kind; keep existing warnings/messages for polar path.
+    kind = validate_kind(kind)
     # --- Plotting Setup ---
-    if ax is None:
-        fig, ax = plt.subplots(
-            figsize=figsize, subplot_kw={"projection": "polar"}
-        )
-    else:
-        fig = ax.figure
+    if kind == "polar":
+        # Respect parameter but force quarter-circle (keep original warning text)
+        canon = canonical_acov(acov, raise_on_invalid=False)
+        if canon != "quarter_circle":
+            warnings.warn(
+                "plot_polar_roc currently renders best as a quarter circle. "
+                f"Received acov='{acov}'. For ROC, θ=0 at the right (East) and a "
+                "90° span give the clearest reading of FPR (angle) vs TPR (radius). "
+                "Proceeding with acov='quarter_circle'.",
+                UserWarning,
+                stacklevel=2,
+            )
 
-    # cmap_obj = get_cmap(cmap, default="viridis")
-    # colors = cmap_obj(np.linspace(0, 1, len(y_preds)))
+        if ax is None:
+            fig, ax = plt.subplots(
+                figsize=figsize, subplot_kw={"projection": "polar"}
+            )
+        else:
+            fig = ax.figure
+    else:  # cartesian
+        if ax is None:
+            fig, ax = plt.subplots(figsize=figsize)
+        else:
+            fig = ax.figure
+
+    # Colors (unchanged behavior / messages)
     colors = get_colors(
         len(y_preds),
         colors=colors,
@@ -110,55 +120,98 @@ def plot_polar_roc(
         failsafe="discrete",
     )
 
-    # --- Plot No-Skill Reference Spiral ---
-    # In polar, the y=x line becomes an Archimedean spiral
+    # --- No-skill reference ---
     if show_no_skill:
-        no_skill_theta = np.linspace(0, np.pi / 2, 100)
-        no_skill_radius = np.linspace(0, 1, 100)
-        ax.plot(
-            no_skill_theta,
-            no_skill_radius,
-            color="gray",
-            linestyle="--",
-            lw=1.5,
-            label="No-Skill (AUC = 0.5)",
-        )
+        if kind == "polar":
[…]
