### 2025-10-30 ef7fe8c
Update the API conventions and ExtensibiliTY
