### 2025-10-30 e0e9141
Build doc and lastest update of paper.md
