+++ paper.md
@@ -71,8 +71,11 @@ them effectively.
 
 `k-diagram` addresses this gap by providing a scalable and intuitive 
 toolkit designed specifically for the visual diagnosis of spatiotemporal 
-probabilistic forecasts [@Liu2024; @kouadiob2025]. The package's novelty lies 
-in its use of polar coordinates to map different dimensions of forecast 
+probabilistic forecasts [@Liu2024; @kouadiob2025]. Furthermore, in modern deployments, 
+a single run may yield thousands to millions of (location, horizon) forecasts; 
+`k-diagram` is designed to compress these large spatiotemporal distributions into stable, 
+comparable views that preserve geography and lead-time structure.
+The package's novelty lies in its use of polar coordinates to map different dimensions of forecast 
 performance, such as uncertainty magnitude, reliability, and temporal stability, onto 
 angle and radius. This approach provides compact overviews that reveal patterns 
 obscured in traditional Cartesian plots. By providing clear visual 
@@ -140,15 +143,14 @@ mapping them to angle and radius. All functionalities are supported by
 a set of helper utilities and a command-line interface for
 generating key plots directly from data files. 
 
-# Software Engineering
+# Engineering and Design
 
 `k-diagram` is built as a small stack of composable layers rather than a 
 monolithic script. User-facing families—uncertainty and error diagnostics, 
 evaluation and comparison, distribution/feature views, relationships, and 
 Taylor diagrams—live in focused modules that expose thin entry points 
 (`plot_*`, `taylor_diagram`). These functions validate inputs, assemble 
-plotting primitives, and return a Matplotlib `Axes` for downstream editing, 
-while the heavy lifting is handled by a utility layer that performs column 
+plotting primitives while the heavy lifting is handled by a utility layer that performs column 
 detection and normalization, quantile pairing and checks, angular coverage 
 mapping, and color handling. This separation keeps "what to draw" independent 
 from "how to draw", and concentrates cross-cutting concerns in one place: 
@@ -161,7 +163,7 @@ orientation and grid.
 From a performance and testing standpoint, data transforms are vectorized with 
 NumPy/Pandas and only compact arrays reach Matplotlib. There is no hidden global
 state; each function depends solely on its arguments and returns an `Axes` 
-object. That purity makes the code easy to test: transforms and validators 
+object for downstream editing. That purity makes the code easy to test: transforms and validators 
 are unit-tested directly, while rendering is exercised with headless smoke 
 tests that assert on ticks, labels, and returned objects rather than pixel 
 hashes. The CLI wraps the same public API, so command-line workflows and 
@@ -177,17 +179,20 @@ convenient; the function chooses the least surprising path and validates
 shapes early. Polar grammar is explicit: angular coverage is controlled by 
 `acov`, orientation by `zero_at` and `clockwise`, and periodic contexts can 
 label angles meaningfully via `theta_ticks`, `theta_ticklabels` 
-(e.g., weekdays instead of degrees), and color mapping to tailor the
-visualizations for their specific domain. The entire package is
-designed for ease of use, allowing users to control plot aesthetics. 
+(e.g., weekdays instead of degrees), In addition, all plots share a small 
+set of display controls: `show_grid` toggles the background grid and 
+`grid_props` customizes its style on both polar and Cartesian axes, 
+while `mask_angle` and `mask_radius` hide angular and radial tick labels 
+respectively on polar axes (these two are ignored in Cartesian mode). 
 Every call returns a Matplotlib `Axes`, enabling the usual compositional 
 workflow—annotate, combine subplots, or `savefig`—with no special wrappers.
 
 Where the community convention is Cartesian (ROC/PR curves, classification 
-summaries), the same functions accept `kind="polar"|"cartesian"`; the 
[…]
