+++ docs/source/user_guide/evaluation.rst
@@ -19,35 +19,19 @@ the performance of multiple models and diagnose their strengths
 and weaknesses.
 
 .. note::
-   **Polar vs. Cartesian rendering (``kind``)**
-
    Many evaluation plots accept ``kind={'polar','cartesian'}``
    (default is ``'polar'`` unless stated otherwise). When
    ``kind='cartesian'``, the function **delegates** to a Cartesian
    renderer while preserving common styling (``figsize``, ``colors``,
    ``show_grid``). Polar-only options (e.g., ``acov``, ``zero_at``,
    ``clockwise``) are ignored in Cartesian mode. The return value is
-   always the actual ``Axes`` used.
-
-   *Use Cartesian* when you want the conventional reading for ROC/PR and
-   classification plots (FPR/TPR on x/y, Precision/Recall on y/x,
-   grouped bars). *Use Polar* when you want compact overviews, periodic
-   angles, or comparative radial layouts. For ROC/PR in polar, a
-   quarter-circle is used for readability.
-
-   **Examples**
-
-   .. code-block:: python
-
-      # Delegates to Cartesian
-      ax = kd.plot_polar_pr_curve(y_true, y_pred1, y_pred2,
-                                  names=['A', 'B'],
-                                  kind='cartesian')
-
-      # Polar with angular coverage controls
-      ax = kd.plot_polar_confusion_matrix(y_true, y_pred,
-                                          kind='polar', acov='default')
-
+   always the actual ``Axes`` used. *Use Cartesian* when you want the 
+   conventional reading for ROC/PR and classification plots (FPR/TPR on x/y,
+   Precision/Recall on y/x, grouped bars). *Use Polar* when you want 
+   compact overviews, periodic angles, or comparative radial layouts. 
+   For ROC/PR in polar, a quarter-circle is used for readability. In the following
+   examples, we use the default behavior ( i.e ``kind="polar"``).
+   
 
 Summary of Evaluation Functions
 -------------------------------
