### 2025-10-30 1a91308
Fix _get_valid_kwargs for py3.13
