### 2025-10-30 earthai-tech/k-diagram#14 @earthai-tech
> The package makes heavy use of polar coordinate plots. For standard Taylor Diagrams (deviation/correlation) polar coordinates are integral, but for many of the plots I do not see any benefit of using polar coordinates, or even make it more hard to interpret, e.g. the [Polar ROC Curve](https://k-diagram.readthedocs.io/en/latest/user_guide/evaluation.html#polar-roc-curve-plot-polar-roc), [Polar Classification Report](https://k-diagram.readthedocs.io/en/latest/user_guide/evaluation.html#polar-classification-report-plot-polar-classification-report), or [Polar Sharpness Diagram](https://k-diagram.readthedocs.io/en/latest/user_guide/probabilistic.html#polar-sharpness-diagram-plot-polar-sharpness).
>
> For [Polar Credibility Bands](https://k-diagram.readthedocs.io/en/latest/user_guide/probabilistic.html#polar-credibility-bands-plot-credibility-bands), polar coordinats can show the periodicity. Is there a possibility to pass the angle ticks? The data is about the binned weekday, but then the ticks show degrees.
>
> I still need some time to go through all the plot types for the [openjournals/joss-reviews#8661](https://github.com/openjournals/joss-reviews/issues/8661).

I see, and I understand. Initially, the core of the project (and the paper) is uncertainty diagnostics, and that’s where the polar design is original and most useful. The “evaluation” pieces (ROC/PR/Classification Report/Sharpness) came later as an experimental add-on. They’re interesting as alternate views, but they shouldn’t replace the standard rectangular versions people already know and trust.
We are planning for a future release, as we have already started with evaluation. py (71cf8d2dbe3d0c92bf528b45f54e9f6b57d39cfb), to keep the polar views for ROC/PR/Classification Report as optional, but ship the standard Cartesian versions alongside them. In the docs, we’ll add a short note on when polar plots help (e.g., cyclic structure, angular comparisons across groups) and when rectangular plots are clearer  ( see also 71cf8d2dbe3d0c92bf528b45f54e9f6b57d39cfb in evaluation.rst).

Secondly, for the Polar Credibility Bands, you get a point and I've implemented what you asked for. For instance, you can now control the angular ticks and labels, and you can choose where “zero” sits and whether angles run clockwise. Concretely, the function henceforth exposes `theta_ticks` and `theta_ticklabels`, and it surfaces `zero_at` (`"N"|"E"|"S"|"W"`) and `clockwise`.  ( see a4c059600d557ae6b5cd6150359273aa3e16635e)
That means your weekday example can be labeled as weekdays instead of degrees ( see [doc] (https://k-diagram.readthedocs.io/en/latest/user_guide/probabilistic.html#polar-credibility-bands-plot-credibility-bands)  and/or commit a4c059600d557ae6b5cd6150359273aa3e16635e.
