### 2025-10-30 #8661 @earthai-tech
> > ## Delineation of Contributions
> > I have been careful to separate the scopes of these works to avoid overlap and ensure each makes a unique contribution.
> > ```
> > * **1. This JOSS Submission** focuses exclusively on the **software artifact**. Its contribution lies
> >   in the software's architecture, API design, the efficient implementation of the plotting
> >   algorithms, and the quality of the testing, documentation, and tutorials that support its use by
> >   the scientific community.
> >
> > * **2. The Scientific Manuscript**:
> >   > _Analytics Framework for Interpreting Spatiotemporal Probabilistic Forecasts_
> >   > (manuscript submitted; no DOI yet).
> >   > Authors: Kouao Laurent Kouadio, Rong Liu, Kouamé Gbèlè Hermann Loukou,
> >   > Wenxiang Liu, Zhanhui Qing, Zhuo Liu (corresponding).
> >
> >   presents the **novel scientific framework** that the software implements. It contains the theoretical
> >   development, validation through case studies, and the scientific interpretation of the results.
> >
> > * **3. The Technical Report** (_"k-diagram: Technical Report — Derivations and Details"_,
> >   (DOI: **https://doi.org/10.5281/zenodo.17051182**) serves as a stable, archival, and **citable companion**
> >   for the underlying mathematics. I created this report to provide a persistent reference for
> >   the derivations while the primary methods manuscript is under peer review, ensuring the
> >   software's methods are transparently documented.
> > ```
>
> I have been re-reading the [Co-publication of science, methods, and software section](https://joss.readthedocs.io/en/latest/submitting.html#co-publication-of-science-methods-and-software) of the JOSS documentation. Reading the current `paper.md`, I see the value of the documentation in 'creation of tutorials' and implementations of plot types. On the contrary, to me the paper does not focus enough on "the software's architecture, API design, the efficient implementation of the plotting algorithms, and the quality of the testing", as stated in the [@earthai-tech](https://github.com/earthai-tech)'s earlier comment. While the scholary effort comes more from the tutorials and plot types, I must concurr and say in my impression the manuscript does not establish the algorithmic efforts.
>
> The package definately has it's selling point of plotting data on polar coordinates, delivers many types of them and also describes and documents this thoroughly. What I am missing until now is the motivation of doing so (aside of Taylor Diagrams), but maybe the motivation is just doing so and scientific curiosity. I am happy to hear from you, [@earthai-tech](https://github.com/earthai-tech). Detailled discussion in [earthai-tech/k-diagram#14](https://github.com/earthai-tech/k-diagram/issues/14).

Thanks for the careful read—and for pushing us to sharpen the *software* contribution in the JOSS paper. You’re right that the current draft lingers on plot types and tutorials. I revised  the manuscript so the engineering story is front and center.  A compact **Engineering & Design** section  outline the package's architecture, data flow, and rendering layers, and an **API Conventions & Extensibility** section spells out stable interfaces, validation rules, and how contributors add new diagnostics ( [71cf8d2dbe3d0c92bf528b45f54e9f6b57d39cfb](https://github.com/earthai-tech/k-diagram/commit/71cf8d2dbe3d0c92bf528b45f54e9f6b57d39cfb#diff-e5bd582629f4e5f64783dd018bad2253d2e07f309bf0370e8d2ff95b0464a5f2) and [ef7fe8c06ef409b6a59fa1e5e29b8848b2440779](https://github.com/earthai-tech/k-diagram/commit/ef7fe8c06ef409b6a59fa1e5e29b8848b2440779). I have also make the testing picture explicit in [development page](https://k-diagram.readthedocs.io/en/latest/development.html)  (unit checks for transforms and validators; smoke tests for rendering; CI across versions) so readers can see the quality of the artifact, not just the gallery ([71cf8d2dbe3d0c92bf528b45f54e9f6b57d39cfb](https://github.com/earthai-tech/k-diagram/commit/71cf8d2dbe3d0c92bf528b45f54e9f6b57d39cfb#diff-e5bd582629f4e5f64783dd018bad2253d2e07f309bf0370e8d2ff95b0464a5f2) in documentation.rst).

On the motivation for polar forms beyond Taylor diagrams. Like I said earlier in https://github.com/earthai-tech/k-diagram/issues/14#issuecomment-3466732465, the core of the k-diagram is an uncertainty diagnostic. Our aim isn’t novelty for its own sake, but a practical grammar for uncertainty at scale. In spatio-temporal settings, polar layouts let you encode complementary behaviours as angles and radii, so clustered failure modes “pop” visually. That’s why patterns like [model drift](https://k-diagram.readthedocs.io/en/latest/gallery/uncertainty.html#model-drift ), [uncertainty drift](https://k-diagram.readthedocs.io/en/latest/gallery/uncertainty.html#uncertainty-drift), and [temporal uncertainty](https://k-diagram.readthedocs.io/en/latest/gallery/uncertainty.html#temporal-uncertainty)often read more clearly in polar space. The gallery already hints at this, *Model drift* shows systematic shifts sector-by-sector over time; *Uncertainty drift* makes widening (or tightening) spread visible at a glance; *Temporal uncertainty* for instance, surfaces seasonality and regime changes without drowning in small multiples.

To better understand, let me consider a concrete example. Imagine monitoring a 10 km x 10 km region with Interferometric Synthetic Aperture Radar (InSAR)  from 2015 to 2025. If we grid the area at 100 m resolution, that’s 100 × 100 = **10,000** locations. With monthly forecasts spanning **11 years**, we have **132** time points, about **1.32 million** point forecasts. In this regime, tables and single-number scores blur structure. Polar diagnostics, by contrast, let you see where credible intervals are steadily widening (uncertainty drift), where centers are sliding in one direction (model drift), and where both evolve together, exactly the cells you would prioritize for field checks, QC, or model retraining. In other words, the polar view earns its keep when you must synthesize dense spatio-temporal behavior into interpretable geometry. This is one of the visualisations we use, `k-diagram`, to do in our diagnosis paper, addressing a land subsidence problem mentioned in  https://github.com/openjournals/joss-reviews/issues/8661#issuecomment-3336891048.

I have woven these clarifications into the paper in the **Statement of Need** section ([e0e914199d148df6de29fd05c3014a6b0e5833a7](https://github.com/earthai-tech/k-diagram/commit/ef7fe8c06ef409b6a59fa1e5e29b8848b2440779) in paper.md). We hope we have made it clearer.
