+++ paper.md
@@ -152,18 +152,18 @@ Taylor diagrams—live in focused modules that expose thin entry points
 (`plot_*`, `taylor_diagram`). These functions validate inputs, assemble 
 plotting primitives while the heavy lifting is handled by a utility layer that performs column 
 detection and normalization, quantile pairing and checks, angular coverage 
-mapping, and color handling. This separation keeps "what to draw" independent 
-from "how to draw", and concentrates cross-cutting concerns in one place: 
-compatibility shims in `kdiagram.compat` stabilize public behavior across 
+mapping, and color handling. This separation keeps **"what to draw"** independent 
+from **"how to draw"**, and concentrates cross-cutting concerns in one place: 
+**compatibility shims** in `kdiagram.compat` stabilize public behavior across 
 Matplotlib and Pandas versions; lightweight validators and decorators catch 
 ambiguous inputs early; and shared helpers (`setup_polar_axes`, `set_axis_grid`,
  `_sample_colors`, `map_theta_to_span`) give every polar figure a consistent 
 orientation and grid.
 
-From a performance and testing standpoint, data transforms are vectorized with 
+From a **performance** and **testing** standpoint, data transforms are vectorized with 
 NumPy/Pandas and only compact arrays reach Matplotlib. There is no hidden global
 state; each function depends solely on its arguments and returns an `Axes` 
-object for downstream editing. That purity makes the code easy to test: transforms and validators 
+object for downstream editing. That purity makes the code easy to test: **transforms and validators** 
 are unit-tested directly, while rendering is exercised with headless smoke 
 tests that assert on ticks, labels, and returned objects rather than pixel 
 hashes. The CLI wraps the same public API, so command-line workflows and 
@@ -176,8 +176,8 @@ The API is data-first and predictable. Plots accept either tidy DataFrames
 with explicit column selectors (for example, `q_cols=('q10','q50','q90')`, 
 `y_true='actual'`, `y_pred='pred')` or plain arrays when that is more 
 convenient; the function chooses the least surprising path and validates 
-shapes early. Polar grammar is explicit: angular coverage is controlled by 
-`acov`, orientation by `zero_at` and `clockwise`, and periodic contexts can 
+shapes early. Polar grammar is explicit: **angular coverage** is controlled by 
+`acov`, orientation by `zero_at` and `clockwise`, and **periodic contexts** can 
 label angles meaningfully via `theta_ticks`, `theta_ticklabels` 
 (e.g., weekdays instead of degrees), In addition, all plots share a small 
 set of display controls: `show_grid` toggles the background grid and 
@@ -185,11 +185,11 @@ set of display controls: `show_grid` toggles the background grid and
 while `mask_angle` and `mask_radius` hide angular and radial tick labels 
 respectively on polar axes (these two are ignored in Cartesian mode). 
 Every call returns a Matplotlib `Axes`, enabling the usual compositional 
-workflow—annotate, combine subplots, or `savefig`—with no special wrappers.
+workflow, annotate, combine subplots, or `savefig`, with no special wrappers.
 
 Where the community convention is Cartesian (ROC/PR curves, classification 
 summaries), the same functions accept `kind="polar"|"cartesian"`; the default 
-is "polar" for compact overviews, with a "cartesian" variant available when 
+is **"polar"** for **compact overviews**, with a **"cartesian"** variant available when 
 user want standard axes and conventional readings. Keep the polar view for 
 cyclic or directional comparisons and switch to Cartesian when familiarity or 
 axis-aligned interpretation matters most.
@@ -199,10 +199,8 @@ into validated arrays or grouped summaries; lay out the coordinate system
 (polar or Cartesian) using the shared helpers; and render with standard 
 Matplotlib primitives. Because mapping, grid styling, and color handling 
[…]
