### 2025-10-30 #8661 @jromanowska
Wonderful! Thank you, @cbueth and @earthai-tech, for the dialog and collaborative effort to make this package better. I think this is a great example of how a good critique and peer-review should work.

@ChinniAbburi, please, finish up your review, or let us know when to expect your final decision.
