### 2025-11-01 f8b79f5
Fix polar labels, add mode annular , and cbueth to clearer interpretation
