+++ kdiagram/plot/errors.py
@@ -4,37 +4,38 @@
 from __future__ import annotations
 
 import warnings
-from typing import Any
+from typing import Any, Literal
 
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 from matplotlib.axes import Axes
 from matplotlib.colors import Normalize
-from scipy.stats import gaussian_kde
+from scipy.stats import gaussian_kde, skew
 
 from ..api.typing import Acov
-from ..compat.matplotlib import get_cmap
+from ..compat.matplotlib import get_cmap, get_colors
 from ..decorators import check_non_emptiness, isdf
 from ..utils.plot import (
     map_theta_to_span,
     set_axis_grid,
     setup_polar_axes,
 )
+from ..utils.fs import savefig as safe_savefig 
+from ..core._io_utils import _get_valid_kwargs
 from ..utils.validator import exist_features
 
 __all__ = ["plot_error_ellipses", "plot_error_bands", "plot_error_violins"]
 
 
-@check_non_emptiness
-@isdf
-def plot_error_violins(
+def _plot_error_violins(
     df: pd.DataFrame,
     *error_cols: str,
     names: list[str] | None = None,
     title: str | None = None,
     figsize: tuple[float, float] = (9.0, 9.0),
     cmap: str = "viridis",
+    colors: list[str] =None, 
     show_grid: bool = True,
     grid_props: dict[str, Any] | None = None,
     savefig: str | None = None,
@@ -88,13 +89,20 @@ def plot_error_violins(
     angles = np.linspace(0.0, float(span), k, endpoint=False)
 
     # colors
-    cmap_obj = get_cmap(cmap, default="viridis")
-    colors = cmap_obj(np.linspace(0.0, 1.0, k))
+    colors = get_colors(
+        k,
+        colors=colors,
+        cmap=cmap,
+        default="tab10",
+        failsafe="discrete",
+    )
 
     # width of each violin in radians, scaled by span
     width = (float(span) / max(1, k)) * 0.8
 
     # draw violins (two-lobed polygon around the angle line)
+    violin_kws = _get_valid_kwargs(ax.fill, violin_kws, error="ignore")
+    
     for i, (ang, dens) in enumerate(zip(angles, violin_data)):
         if dens is None:
             continue
@@ -107,7 +115,7 @@ def plot_error_violins(
 
         theta = x + float(ang)
         r = y
-
+        
         ax.fill(
             theta,
             r,
@@ -139,14 +147,254 @@ def plot_error_violins(
 
     set_axis_grid(ax, show_grid=show_grid, grid_props=grid_props)
 
-    fig.tight_layout()
-    if savefig:
-        fig.savefig(savefig, dpi=dpi, bbox_inches="tight")
-        plt.close(fig)
-    else:
-        plt.show()
+    # saving 
+    final =safe_savefig(
+        savefig,
+        fig, 
+        dpi=dpi,
+        bbox_inches="tight",
+    )
+    if final is not None: 
+        plt.close(fig) 
+    else: 
+        fig.tight_layout()
+        plt.show() 
 
     return ax
+
+
+@check_non_emptiness
+@isdf
+def plot_error_violins(
+    df: pd.DataFrame,
+    *error_cols: str,
+    names: list[str] | None = None,
+    title: str | None = None,
+    figsize: tuple[float, float] = (9.0, 9.0),
+    cmap: str = "viridis",
+    colors: list[str] =None, 
+    show_grid: bool = True,
+    grid_props: dict[str, Any] | None = None,
+    savefig: str | None = None,
+    dpi: int = 300,
+    acov: Acov = "default",
+    ax: Axes | None = None,
[…]
