+++ kdiagram/__init__.py
@@ -67,6 +67,7 @@ if _missing_dependencies:
 
 # Re-export config helpers
 from .config import configure_warnings, warnings_config  # noqa: F401, E402
+from .utils import savefig # Noqa: E402
 from .plot import (  # noqa: E402
     plot_actual_vs_predicted,
     plot_anomaly_glyphs,
@@ -173,4 +174,5 @@ __all__ = [
     "plot_cas_profile",
     "plot_glyphs",
     "plot_cas_layers",
+    "savefig"
 ]
