### 2025-11-01 53ace52
merge Engineering design and API convention in singe Much more section in paper.md
