+++ paper.md
@@ -140,70 +140,65 @@ Finally, the package provides tools for **Relationship Visualization**,
 where polar plots in `kdiagram.plot.relationship` offer a novel
 perspective on the correlation between true and predicted values by
 mapping them to angle and radius. All functionalities are supported by
-a set of helper utilities and a command-line interface for
+a set of helper utilities and a command-line interface (CLI) for
 generating key plots directly from data files. 
 
-# Engineering and Design
 
-`k-diagram` is built as a small stack of composable layers rather than a 
-monolithic script. User-facing families—uncertainty and error diagnostics, 
-evaluation and comparison, distribution/feature views, relationships, and 
-Taylor diagrams—live in focused modules that expose thin entry points 
-(`plot_*`, `taylor_diagram`). These functions validate inputs, assemble 
-plotting primitives while the heavy lifting is handled by a utility layer that performs column 
-detection and normalization, quantile pairing and checks, angular coverage 
-mapping, and color handling. This separation keeps "what to draw" independent 
-from "how to draw", and concentrates cross-cutting concerns in one place: 
-compatibility shims in `kdiagram.compat` stabilize public behavior across 
-Matplotlib and Pandas versions; lightweight validators and decorators catch 
-ambiguous inputs early; and shared helpers (`setup_polar_axes`, `set_axis_grid`,
- `_sample_colors`, `map_theta_to_span`) give every polar figure a consistent 
-orientation and grid.
+# Engineering, API, and Extensibility
 
-From a performance and testing standpoint, data transforms are vectorized with 
-NumPy/Pandas and only compact arrays reach Matplotlib. There is no hidden global
-state; each function depends solely on its arguments and returns an `Axes` 
-object for downstream editing. That purity makes the code easy to test: transforms and validators 
-are unit-tested directly, while rendering is exercised with headless smoke 
-tests that assert on ticks, labels, and returned objects rather than pixel 
-hashes. The CLI wraps the same public API, so command-line workflows and 
-Python sessions produce identical figures. The next section outlines the 
-API conventions and extension points that guide usage and contributions.
+`k-diagram` is designed as a **small stack of composable layers** rather than a
+monolithic script. This architecture is directly reflected in its public API,
+which is **data-first**, **predictable**, and geared toward **integration** with the
+scientific Python ecosystem.
 
-# API Conventions and Extensibility
+The primary API entry points are the `plot_*` functions. These functions are
+designed to be thin, predictable wrappers that accept data in common
+formats—either a tidy `pandas` DataFrame with explicit column selectors
+(e.g., `q_cols=('q10','q50','q90')`, `y_true='actual'`) or plain `numpy` arrays.
+In either case, inputs are **validated early** to provide clear error messages.
+A core feature of the API is its **explicit "polar grammar"**. Users have direct
+control over the geometry of their plots through parameters like `acov`
+(angular coverage), `zero_at` (orientation), and `clockwise` (direction).
+For periodic data, this grammar extends to human-readable labels via
+`theta_ticks` and `theta_ticklabels`, allowing axes to show "Monday" instead
+of "0.0 radians." These controls are paired with shared display options
+like `show_grid` to ensure a consistent look and feel.
 
-The API is data-first and predictable. Plots accept either tidy DataFrames 
-with explicit column selectors (for example, `q_cols=('q10','q50','q90')`, 
-`y_true='actual'`, `y_pred='pred')` or plain arrays when that is more 
-convenient; the function chooses the least surprising path and validates 
-shapes early. Polar grammar is explicit: angular coverage is controlled by 
-`acov`, orientation by `zero_at` and `clockwise`, and periodic contexts can 
-label angles meaningfully via `theta_ticks`, `theta_ticklabels` 
[…]
