+++ docs/source/development.rst
@@ -21,20 +21,69 @@ without breaking existing users.
 
 
 Architecture at a Glance
-------------------------
+-------------------------
 
 The package is a small stack of composable layers:
 
-- ``kdiagram.plot`` – user functions grouped by task
-  (``uncertainty``, ``errors``, ``probabilistic``, ``evaluation``,
-  ``comparison``, ``feature_based``, ``relationship``, ``taylor_diagram``).
-- ``kdiagram.utils`` – shared helpers for polar setup, colors, KDE, grids,
-  validation, and tiny cross-cutting utilities.
-- ``kdiagram.core`` – I/O and post-processing (column coercion, quantiles,
-  binning, indexing, error policies).
-- ``kdiagram.compat`` – shims for Matplotlib/Pandas version differences.
-- ``kdiagram.cli`` – command-line front-ends that mirror the public API.
-- ``kdiagram.datasets`` – small helpers to stage example assets for docs/tests.
+- ``kdiagram.plot``
+
+  This is the main user-facing API, containing all plotting functions grouped
+  by task. Each file represents a family of plots,
+  such as ``probabilistic.py`` for PIT histograms and sharpness diagrams, 
+  ``errors.py`` for violin and band plots, or ``anomaly.py`` for visualizing 
+  forecast failures. Other modules like ``evaluation.py`` and ``context.py`` 
+  provide evaluation and contextual plots and  ``taylor_diagram`` 
+  for model evaluations.
+
+- ``kdiagram.utils``
+
+  This package provides the core machinery shared by all plots.
+  It includes essential plot utilities like ``setup_polar_axes`` for creating
+  the polar canvas and ``set_axis_grid`` for consistent styling. It also  
+  contains robust data-handling logic, such as the quantile helpers in
+  ``diagnose_q.py`` (e.g., ``validate_qcols``, ``detect_quantiles_in``)
+  and metric calculation functions in ``mathext.py`` (e.g., ``compute_crps``).
+  Filesystem operations are centralized in ``fs.py``, which provides the
+  ``safe_savefig`` helper.
+
+- ``kdiagram.core``
+
+  This layer handles data input/output and internal post-processing.
+  ``io.py`` provides the main ``read_data`` and ``write_data`` functions. 
+  This is supported by ``_io_utils.py``, which contains
+  internal helpers like ``_post_process`` (for ``fillna``, ``dropna``)
+  and ``_get_valid_kwargs``. This ``_get_valid_kwargs``
+  filter is used throughout the plotting modules (e.g., in ``errors.py``) 
+  to safely pass user customizations to Matplotlib.
+
+- ``kdiagram.compat``
+
+  This is a crucial layer for ensuring stable behavior across different
+  versions of Matplotlib, Pandas, and Scikit-learn.
+  It provides compatibility shims, such as ``get_cmap``
+  and ``get_colors`` for consistent color handling,
+  and backports metrics like ``root_mean_squared_error``
+  if they aren't available in the user's environment.
+
+- ``kdiagram.cli``
+
+  This package exposes the library's functionality to the command line. 
+  The main ``__init__.py`` file builds the primary
+  parser (``build_parser``) and registers all available plots as subcommands. 
+  Each subcommand, like ``add_plot_anomalies`` or ``add_plot_taylord``, 
+  links a CLI command (e.g., ``k-diagram plot-anomalies``) directly to the 
+  corresponding Python plotting function.
+
+``kdiagram.datasets``
+
+  This provides utilities for documentation and testing.
+  ``load.py`` contains functions to load real-world sample data, such as
+  ``load_zhongshan_subsidence``. ``make.py`` contains
+  generators for creating synthetic, reproducible datasets
+  (e.g., ``make_regression_data``, ``make_cyclical_data``,
[…]
