### 2025-11-01 4f4d541
Add test violins plot mode cbueth
