+++ kdiagram/plot/errors.py
@@ -15,14 +15,14 @@ from scipy.stats import gaussian_kde, skew
 
 from ..api.typing import Acov
 from ..compat.matplotlib import get_cmap, get_colors
+from ..core._io_utils import _get_valid_kwargs
 from ..decorators import check_non_emptiness, isdf
+from ..utils.fs import savefig as safe_savefig
 from ..utils.plot import (
     map_theta_to_span,
     set_axis_grid,
     setup_polar_axes,
 )
-from ..utils.fs import savefig as safe_savefig 
-from ..core._io_utils import _get_valid_kwargs
 from ..utils.validator import exist_features
 
 __all__ = ["plot_error_ellipses", "plot_error_bands", "plot_error_violins"]
@@ -35,7 +35,7 @@ def _plot_error_violins(
     title: str | None = None,
     figsize: tuple[float, float] = (9.0, 9.0),
     cmap: str = "viridis",
-    colors: list[str] =None, 
+    colors: list[str] = None,
     show_grid: bool = True,
     grid_props: dict[str, Any] | None = None,
     savefig: str | None = None,
@@ -102,7 +102,7 @@ def _plot_error_violins(
 
     # draw violins (two-lobed polygon around the angle line)
     violin_kws = _get_valid_kwargs(ax.fill, violin_kws, error="ignore")
-    
+
     for i, (ang, dens) in enumerate(zip(angles, violin_data)):
         if dens is None:
             continue
@@ -115,7 +115,7 @@ def _plot_error_violins(
 
         theta = x + float(ang)
         r = y
-        
+
         ax.fill(
             theta,
             r,
@@ -147,18 +147,18 @@ def _plot_error_violins(
 
     set_axis_grid(ax, show_grid=show_grid, grid_props=grid_props)
 
-    # saving 
-    final =safe_savefig(
+    # saving
+    final = safe_savefig(
         savefig,
-        fig, 
+        fig,
         dpi=dpi,
         bbox_inches="tight",
     )
-    if final is not None: 
-        plt.close(fig) 
-    else: 
+    if final is not None:
+        plt.close(fig)
+    else:
         fig.tight_layout()
-        plt.show() 
+        plt.show()
 
     return ax
 
@@ -172,40 +172,37 @@ def plot_error_violins(
     title: str | None = None,
     figsize: tuple[float, float] = (9.0, 9.0),
     cmap: str = "viridis",
-    colors: list[str] =None, 
+    colors: list[str] = None,
     show_grid: bool = True,
     grid_props: dict[str, Any] | None = None,
     savefig: str | None = None,
     dpi: int = 300,
     acov: Acov = "default",
     ax: Axes | None = None,
-    mode: Literal["basic", "cbueth"] ="cbueth", # to honor the reviewer , I kept 
-    # his name as a mode "cbueth", 
-    # --- NEW options (backward compatible) ---
-    bw_method: float | str | None = None,   
-    overlay: bool = False,                  
-    overlay_angle: float | None = None,     
-    show_stats: bool = False,               
+    mode: Literal["basic", "cbueth"] = "cbueth",
+    bw_method: float | str | None = None,
+    overlay: bool = False,
+    overlay_angle: float | None = None,
+    show_stats: bool = False,
     **violin_kws,
 ):
     mode = str(mode).lower()
-    if mode =="basic": 
+    if mode == "basic":
         return _plot_error_violins(
             df,
             *error_cols,
-            names=names, 
-            title=title, 
-            figsize=figsize, 
-            cmap=cmap, 
-            show_grid=show_grid, 
-            grid_props=grid_props, 
-            savefig=savefig, 
-            dpi=dpi, 
-            acov=acov, 
-            ax=ax, 
[…]
