### 2025-11-01 8150cc4
REV DOC plot fingerprint
