+++ kdiagram/plot/feature_based.py
@@ -1494,3 +1494,225 @@ def _draw_angular_labels(
             va="center",
             fontsize=fs,
         )
+
+
+plot_fingerprint.__doc__ = r"""
+Create a flexible polar 'fingerprint' (radar) chart.
+
+This function is a versatile wrapper for creating feature importance
+fingerprints. It operates in two modes:
+
+1.  **Precomputed Mode** (``precomputed=True``):
+    It plots an existing importance matrix (as a ``pd.DataFrame``
+    or ``np.ndarray``), similar to
+    :func:`~kdiagram.plot.feature_based.plot_feature_fingerprint`.
+
+2.  **Computation Mode** (``precomputed=False``):
+    It computes the importance scores "on-the-fly" from a raw
+    ``pd.DataFrame``. This mode can calculate feature correlations
+    against a target (``method='abs_corr'``) or compute feature
+    variability (``method='std'``, ``'var'``, ``'mad'``). It can
+    also generate separate fingerprints for different groups using
+    the ``group_col`` parameter.
+
+Parameters
+----------
+importances : pd.DataFrame or np.ndarray
+    The input data.
+    
+    - If ``precomputed=True``, this is the (n_layers, n_features)
+      matrix of importance scores.
+    - If ``precomputed=False``, this is the raw ``pd.DataFrame``
+      containing features, target, and (optionally) group columns.
+
+precomputed : bool, default=True
+    Controls the function's behavior.
+    
+    - If ``True``, plots ``importances`` as a precomputed score matrix.
+    - If ``False``, computes scores from the ``importances`` DataFrame
+      using the specified ``method``.
+
+y_col : str, optional
+    The name of the target column (e.g., 'y_true') within the
+    ``importances`` DataFrame. Required when ``precomputed=False``
+    and ``method='abs_corr'``.
+
+group_col : str, optional
+    The name of a categorical column in the ``importances`` DataFrame
+    (e.g., 'model_name', 'year'). If provided (and
+    ``precomputed=False``), one 'fingerprint' polygon will be
+    generated for each unique group.
+
+method : {'abs_corr', 'std', 'var', 'mad'}, default='abs_corr'
+    The method used to compute importance scores when
+    ``precomputed=False``.
+    
+    - ``'abs_corr'``: Absolute Pearson correlation of each feature
+      with ``y_col``.
+    - ``'std'``: Standard deviation of each feature.
+    - ``'var'``: Variance of each feature.
+    - ``'mad'``: Median Absolute Deviation of each feature.
+
+features : list of str, optional
+    Names of the features.
+    
+    - If ``precomputed=False``, this is the list of columns to
+      compute importance for. If ``None``, all numeric columns
+      (excluding ``y_col`` and ``group_col``) are used.
+    - If ``precomputed=True`` and ``importances`` is an array,
+      these are used as the angular (feature) labels.
+
+labels : list of str, optional
+    Display names for the layers (polygons), used in the legend.
+    
+    - If ``precomputed=True``, this labels the rows of the matrix.
+    - If ``precomputed=False`` and ``group_col`` is provided, this
+      is ignored, and the unique group names are used as labels.
+
+normalize : bool, default=True
+    If ``True``, normalizes each layer (row) of the importance
+    matrix so that its maximum value is 1.0. This is useful for
+    comparing the *shape* of fingerprints, regardless of their
+    absolute scales.
+
+fill : bool, default=True
[…]
