### 2025-11-01 d896658
update development page
fix reST indentation in ``kdiagram.datasets`` section
