### 2025-11-01 #8661 @earthai-tech
> [@jromanowska](https://github.com/jromanowska) I was able to finish my review and check all boxes of my checklist [#8661 (comment)](https://github.com/openjournals/joss-reviews/issues/8661#issuecomment-3164146675). The last two rounds [@earthai-tech](https://github.com/earthai-tech) addressed the issues more direct, facilitating a more swift review. I was able to reproduce examples from the gallery in the documentation and was able to understand the premise of the package better.
>
> `k-diagram` provides a vast variety of plot types with the intention to improve grasping uncertainties in forecasting models. The focus on polar coordinates is a sometimes unconventional, but refreshing approach. The _Gallery_ and _User Guide_ in the documentation demonstrate all types of visualisations, well documented. With this I would like to suggest the paper for acceptance! 🎉
>
> I left last minor comment in [earthai-tech/k-diagram#14 (comment)](https://github.com/earthai-tech/k-diagram/issues/14#issuecomment-3467539516) which should not block the acceptance. Furthermore, best luck with the review of the scientific paper and case studies.

@cbueth Thank you so much for this incredibly positive and supportive review!
I truly appreciate all the time and effort you've dedicated to reviewing `k-diagram`. Your detailed feedback has been invaluable.
I also wanted to let you know that I have just addressed your final minor comment regarding the label overlaps. I've pushed the fixes for that in the issue you mentioned ([earthai-tech/k-diagram#14](https://github.com/earthai-tech/k-diagram/issues/14#issuecomment-3475774300)). Thank you for catching those last details!

Thank you again for all your constructive feedback and for your good wishes on the scientific paper. It has been a pleasure to collaborate with you during this review.
