### 2025-11-04 earthai-tech/k-diagram#14 @earthai-tech
> Hello [@earthai-tech](https://github.com/earthai-tech), you're welcome! Nice you already looked into the minor notes. About the new plot mode of the violins, no need to name it after me, rather use a more descriptive name for it, so a user can understand it better. Take care and feel free to close this issue.

ok. well noted! thanks
