### 2025-11-18 #8661 @ChinniAbburi
@jromanowska Yes, I can go through the checklist one more time and confirm whether I’m satisfied with the software. I’ll be able to complete this by the end of this week.
