### 2025-12-08 96221eb
change cbueth mode to optimized
