+++ chatgpt_estimators_guide.txt
@@ -1,0 +1,725 @@
+verbose : int, default=0
+    Controls the verbosity when fitting.
+
+ if self.verbose:
+    progress_bar = tqdm(range(self.n_estimators), ascii=True, ncols= 100,
+        desc=f'Fitting {self.__class__.__name__}', 
+    )
+
+    if self.verbose: 
+        progress_bar.update (1)
+                
+if self.verbose: 
+    progress_bar.close() 
+
+#since BaseAdalineStochastic has abstractMethod __init__, im not sure , it can be instanciated, 
+# therefore in the example using either AdalineStochasticClassifier or AdalineStochasticRegressor who inherit from BaseAdalineStochastic 
+
+# find the most intuitive name to replace <EncodingTransformer>
+
+
+#----------------------------------------------------------------------------------------------------------------------------------------------
+Read the code, Reorganize it and optimize it  the code so to implement robustly either *fit* or *run* method ( not both ) since we want to design 
+a strong API. Read below the fit/run implementation section so select the method that is suitable for the class. 
+
+
+`<subpackage>= mlops
+`<module_name>= automation`
+`<object>= `SimpleAutomation`
+
+**Instructions for documenting the `<object>`**:
+- Provide a detailed explanation of each parameter, ensuring that methods that do not start with `_` are properly documented for readability and 
+understanding.
+- Use the **Numpy docstring format** and **reST** markup for parameters, which is standard for scientific and technical packages. 
+- When mentioning methods (excluding those starting with `_`), ensure that parameters are properly formatted with single backticks (` `) for inline 
+parameters and double backticks (`` ``) for arguments, similar to the **scikit-learn API**. Also, remember to end the inner attributes with an 
+underscore (`_`) ( thoses who you think importants and not listed in the <object>.__init__, similar like scikit-learn API).
+- skip the documentation of __init__ method as parameters should be documented in class docstring 
+
+**Mathematical Formulations**:
+- If the `<object>` performs mathematical calculations, use `.. math::` for mathematical equations and `:math:` for inline mathematical expressions.
+
+**Examples and Notes**:
+- Include example usages in your documentation. For example:
+  
+  ```python
+  >>> from gofast.<subpackage>.<module_name> import <object>
+  ```
+  
+  - Add **Notes** that highlight important details or behaviors, and make sure the examples are clear for users.
+
+**See Also and References**:
+- Add a **See Also** section to reference similar classes or methods.
+- Use reST reference styles for citations, i.e., `[1]_` for inline citations, and `.. [1] <authors, etc.>` in the References section.
+
+**Line Length and Formatting**:
+- For better readability, come to a new line after 70 characters in both the class and method docstrings.
+- Vertically align parameters in the docstring, ensuring clarity and professionalism.
+
+**Code Improvements**:
+- **Error Fixing**: Ensure that the code is free from errors and check attributes naming, and that comments provide clarity to other developers.
+- If there are placeholders or simulate in the code or comments indicating that implementation is needed, make sure to fully implement these sections
+because we are designing a complete software package that need dealing with real-world scenarios.
+- Make the code clearer, well-documented, and syntactically correct to reflect a high level of Python expertise.
+- comment the code to help developper where you think it is very intuitive
+
+**Parameter Validation**:
+- Use `validate_params` from `sklearn.utils._param_validation` to validate the parameters. If necessary, use specific validators like `Interval`, 
+`StrOptions`, `HasMethods`, or `Hidden`, etc.
+  
+**External Dependencies**:
+- If the `<object>` methods require external dependencies that do not come with Python by default, use the `@ensure_pkg` decorator 
+from `gofast.tools.funcutils`:
[…]
