+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "k-diagram"
-version = "1.3.2"
+version = "1.3.3"
 description = "Rethinking Forecasting Uncertainty via Polar-Based Visualization"
 readme = { file = "README.md", content-type = "text/markdown" }
 authors = [{ name = "Laurent Kouadio", email = "etanoyau@gmail.com" }]
