### 2025-12-09 df466ae
Update paper.bib
