+++ paper.bib
@@ -266,7 +266,9 @@
   title     = {Density Estimation for Statistics and Data Analysis},
   year      = {1986},
   publisher = {Chapman and Hall/CRC},
-  isbn      = {9780412246203}
+  address   = {London},
+  isbn      = {9780412246203},
+  doi       = {10.1201/9781315140919}
 }
 
 @article{Biswas2025,
