### 2025-12-09 #8661 @earthai-tech
> [@earthai-tech](https://github.com/earthai-tech), the paper is very well written! There is, however, one thing I'd like you to fix: the section "Engineering - API and Extensibility" has some bold words that seem to be weirdly placed. In the other sections, you use bold font to mark various objects and terms, so perhaps you want to stick to this and simply remove the bold in this "Engineering" section?

Thank you for the careful read and the helpful comment about the boldface in the Engineering – API and Extensibility section.
I have revised that section to remove the extra bold formatting and now only use bold to highlight named objects and terms, consistent with the rest of the paper.
The updated paper.md has been pushed to the repository.
