### 2025-12-09 ad56db0
Update paper.bib
