+++ paper.bib
@@ -89,21 +89,25 @@
   publisher = {SciPy}, 
   doi       = {10.25080/majora-92bf1922-00a}, 
 }
-
 @article{2020SciPy-NMeth,
-  title = {{SciPy 1.0: fundamental algorithms for scientific computing in Python}},
-  author = {{Virtanen}, P. and {Gommers}, R. and {Oliphant}, T.~E. and {Haberland}, M. and {Reddy}, T. and {Cournapeau}, D. and {Burovski}, E. and {Peterson}, P. and {Wehrer}, W. and {Huttunen}, J. and {H{\aa}cklin}, T. and {Straub}, S. and {Pike}, J. and {Petroff}, A. and {Conti}, K. and {Lieberman}, R. and {Pollard}, J. and {Nichols}, B. and {Raimi}, J. and {Hoydu}, E. and {Kerkwijk}, K.~B. and {Brett}, M. and {Haldane}, A. and {del R{\'{i}}o}, J. and {Manolopoulos}, G. and {Ivanov}, A. and {Millman}, K.~J. and {Wehner}, M.~C. and {Briggs}, S.~J. and {Yurkewycz}, N. and {Knothe}, R. and {Butterworth}, J. and {Goodman}, N. and {de la Vega}, S.~J. and {Routchenko}, O. and {Perron}, A. and {Rothberg}, Z. and {Ryaboy}, E. and {Williams}, N. and {Turrell}, P. and {Behnel}, S. and {Sieber}, L. and {Blake}, R. and {Kern}, B.~R. and {Larson}, R. and {Edwards}, C.~A. and {Fazio}, E. and {Juric}, I. and {Pitkanen}, T. and {Cijesek}, S. and {Taylor}, B.~E. and {Miller}, R.~T. and {Jessen}, E. and {Derksen}, D. and {Carey}, C.~J. and {Wejchert}, S. and {Olen{\'i}n}, S. and {Petrov}, A. and {Jachalsky}, A. and {Slavov}, V. and {Bokhove}, O. and {Jung}, J. and {Aivazis}, M. and {Hendriksen}, M.~D. and {Halderman}, K.~A. and {de Miranda Cardoso}, A. and {Levy}, D. and {V{\'a}zquez-Baeza}, Y.},
+  title   = {SciPy 1.0: fundamental algorithms for scientific computing in Python},
+  author  = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and
+             Haberland, Matt and Reddy, Tyler and Cournapeau, David and
+             Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and
+             Bright, Jonathan and others},
   journal = {Nature Methods},
-  volume = {17},
-  number = {3},
-  pages = {261-272},
-  year = {2020},
-  doi = {10.1038/s41592-019-0686-2}
+  volume  = {17},
+  number  = {3},
+  pages   = {261--272},
+  year    = {2020},
+  publisher = {Nature Publishing Group},
+  doi     = {10.1038/s41592-019-0686-2},
+  url     = {https://doi.org/10.1038/s41592-019-0686-2}
 }
 
 @article{Taylor2001,
   author    = {Taylor, Karl E.},
-  title     = {Summarizing quantitative measures of model performance in a single diagram},
+  title     = {Summarizing multiple aspects of model performance in a single diagram},
   journal   = {Journal of Geophysical Research: Atmospheres},
   volume    = {106},
   number    = {D7},
@@ -124,16 +128,25 @@
   doi = {10.1038/s41586-020-2649-2}
 }
 
-@article{reback2020pandas,
-  title={{pandas: a foundational Python library for data analysis and manipulation}},
[…]
