### 2025-12-09 4bf861b
Revise pandas citation in paper.md
Updated citation for pandas in the documentation.
