### 2025-12-09 445d293
Update paper.bib
