+++ paper.bib
@@ -139,7 +139,7 @@
 }
 
 @software{pandasdev,
-  author    = {{The pandas development team}},
+  author    = {{Pandas development team}},
   title     = {pandas-dev/pandas: Pandas},
   year      = {2020},
   publisher = {Zenodo},
