### 2025-12-09 fcf5fe9
Update paper.bib
