+++ paper.bib
@@ -128,7 +128,7 @@
   doi = {10.1038/s41586-020-2649-2}
 }
 
-@InProceedings{reback2020,
+@inproceedings{lu2020securing,
   author    = { {H}aw-minn {L}u and {A}drian {K}wong and {J}os\'e {U}npingco },
   title     = { {S}ecuring {Y}our {C}ollaborative {J}upyter {N}otebooks in the {C}loud using {C}ontainer and {L}oad {B}alancing {S}ervices },
   booktitle = { {P}roceedings of the 19th {P}ython in {S}cience {C}onference },
