### 2025-12-09 cb5f979
Update scipy citation in paper.md
Added a citation to the scipy library in the documentation.
