### 2025-12-09 #8661 @earthai-tech
> Moreover, I can see that have some errors in references:
>
> * Reback, J., McKinney, ... pandas: a foundational Python library for data analysis and manipulation
>   is linked to [this article](https://proceedings.scipy.org/articles/Majora-342d178e-001)
> * Taylor, K. E. (2001). Summarizing quantitative measures of model performance in a single diagram
>   the title is not correct - [original](https://agupubs.onlinelibrary.wiley.com/doi/10.1029/2000JD900719)
> * SciPy 1.0 - there are wrong author names - [see original](https://www.nature.com/articles/s41592-019-0686-2)

@jromanowska
You are right, I have now corrected the bibliography as requested:
 – The SciPy 1.0 reference has been updated to match the official Nature Methods citation (full author list and title).
– The pandas citation now points to the official Zenodo software record (doi:10.5281/zenodo.3509134), and the previously mis-linked SciPy 2020 proceedings DOI has been removed from the pandas entry.
 – The Taylor (2001) reference title has been corrected to “Summarizing Multiple Aspects of Model Performance in a Single Diagram.”

The updated `paper.bib` and manuscript are pushed to the repository; all references should now be consistent with the originals.
