### 2025-12-20 e0c5bb2
Update paper.md
