### 2025-12-20 cfe7338
Update paper.bib
