+++ paper.bib
@@ -90,7 +90,7 @@
   doi       = {10.25080/majora-92bf1922-00a}, 
 }
 @article{2020SciPy-NMeth,
-  title   = {SciPy 1.0: fundamental algorithms for scientific computing in Python},
+  title   = {SciPy 1.0: fundamental algorithms for scientific computing in {P}ython},
   author  = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and
              Haberland, Matt and Reddy, Tyler and Cournapeau, David and
              Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and
@@ -293,7 +293,7 @@
     number = {111},
     pages = {8336},
     author = {Biswas, Anubhab and Nespoli, Lorenzo and Azzimonti, Dario and Zambon, Lorenzo and Rubattu, Nicolò and Corani, Giorgio},
-    title = {BayesReconPy: A Python package for forecast reconciliation},
+    title = {BayesReconPy: A {P}ython package for forecast reconciliation},
     journal = {Journal of Open Source Software}
 }
 
