+++ CITATION.cff
@@ -1,0 +1,28 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Kouadio
+  given-names: Kouao Laurent
+  orcid: "https://orcid.org/0000-0001-7259-7254"
+doi: 10.5281/zenodo.17855722
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Kouadio
+    given-names: Kouao Laurent
+    orcid: "https://orcid.org/0000-0001-7259-7254"
+  date-published: 2025-12-20
+  doi: 10.21105/joss.08661
+  issn: 2475-9066
+  issue: 116
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 8661
+  title: "k-diagram: Rethinking Forecasting Uncertainty via Polar-based
+    Visualization"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.08661"
+  volume: 10
+title: "k-diagram: Rethinking Forecasting Uncertainty via Polar-based
+  Visualization"
