### 2025-12-21 bbd396f
Update the plot-error-violins: numpy2xx-trapezoid
