+++ kdiagram/plot/errors.py
@@ -15,6 +15,7 @@ from scipy.stats import gaussian_kde, skew
 
 from ..api.typing import Acov
 from ..compat.matplotlib import get_cmap, get_colors
+from ..compat.numpy import trapz 
 from ..core._io_utils import _get_valid_kwargs
 from ..decorators import check_non_emptiness, isdf
 from ..utils.fs import savefig as safe_savefig
@@ -307,12 +308,12 @@ def plot_error_violins(
     # order small → large so big shapes don’t hide small ones
     areas = []
     for dpos, dneg in violins:
-        a = (np.trapz(dpos, r_grid) if dpos is not None else 0.0) + (
-            np.trapz(dneg, r_grid) if dneg is not None else 0.0
+        a = (trapz(dpos, x=r_grid) if dpos is not None else 0.0) + (
+            trapz(dneg, x=r_grid) if dneg is not None else 0.0
         )
-        areas.append(a)
+        areas.append(float(a))
     order = np.argsort(areas)
-
+    
     for rank, i in enumerate(order):
         (dpos, dneg), ang, col = violins[i], float(angles[i]), colors[i]
 
