### 2025-12-22 2d53f45
Update the paper citation
